% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_class.R
\name{adam}
\alias{adam}
\title{Public constructor for adam objects}
\usage{
adam(adam_data, source = "")
}
\arguments{
\item{adam_data}{The ADaM datasets as list of data frames.}

\item{source}{Source information as character.}
}
\value{
An adam object.
}
\description{
Public constructor for adam objects
}
