% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ae.R
\name{add_ae_observation}
\alias{add_ae_observation}
\title{Add AE observation}
\usage{
add_ae_observation(
  nif,
  sdtm,
  ae_term,
  ae_field = "AEDECOD",
  parent = NULL,
  cmt = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  keep = "",
  debug = FALSE,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{debug}{Include debug fields, as logical.}

\item{silent}{Suppress messages, defaults to nif_option setting, if NULL.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function adds AE events as observations with the time of the AE onset
as the observation time point and the severity grade as the dependent
variable.
}
