% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_bl_lab}
\alias{add_bl_lab}
\title{Add baseline covariates to a NIF data set}
\usage{
add_bl_lab(nif, lb, lbspec, ...)
}
\arguments{
\item{nif}{NIF dataset.}

\item{lb}{SDTM LB domain as data frame.}

\item{lbspec}{The specimen, usually "BLOOD" or "URINE".}

\item{...}{Lab parameters as encoded by LBTESTCD, as strings.}
}
\value{
A NIF dataset
}
\description{
Lab parameters not found in LB will be reported in a warning message.
}
