% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_bl_lbm}
\alias{add_bl_lbm}
\title{Add baseline lean body mass (LBM)}
\usage{
add_bl_lbm(obj, method = lbm_boer)
}
\arguments{
\item{obj}{A nif object.}

\item{method}{The function to calculate LBM, i.e., \code{\link[=lbm_boer]{lbm_boer()}},
\code{\link[=lbm_hume]{lbm_hume()}} or \code{\link[=lbm_peters]{lbm_peters()}}.}
}
\value{
A nif object.
}
\description{
Add baseline lean body mass (LBM)
}
\seealso{
\code{\link[=lbm_hume]{lbm_hume()}}

\code{\link[=lbm_boer]{lbm_boer()}}

\code{\link[=lbm_peters]{lbm_peters()}}
}
