% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_bl_renal}
\alias{add_bl_renal}
\title{Add baseline renal function class}
\usage{
add_bl_renal(obj, method = egfr_cg)
}
\arguments{
\item{obj}{A NIF data set.}

\item{method}{The function to calculate eGFR (CrCL) from serum creatinine.
Currently either: egfr_mdrd, egfr_cg or egfr_raynaud}
}
\value{
A NIF data set.
}
\description{
Add baseline renal function class
}
