% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_covariates.R
\name{add_covariate}
\alias{add_covariate}
\title{Add time-varying covariate}
\usage{
add_covariate(
  nif,
  sdtm,
  domain,
  testcd,
  covariate = NULL,
  dtc_field = NULL,
  dv_field = NULL,
  testcd_field = NULL,
  observation_filter = "TRUE",
  cat = NULL,
  scat = NULL,
  duplicate_function = mean,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{The corresponding sdtm object.}

\item{domain}{The domain as character.}

\item{testcd}{The xxTESTCD with xx the domain name, as character.}

\item{covariate}{The name of the covariate, defaults to the testcd if NULL.}

\item{dtc_field}{The field to use as the date-time code for the observation.
Defaults to the two-character domain name followed by 'DTC', if NULL.}

\item{dv_field}{The name of the DV field as character.}

\item{testcd_field}{The name of the TESTCD field. defaults to xxTESTCD (with
xx the domain code), as character.}

\item{observation_filter}{A filter term for the \code{domain}, as character.}

\item{cat}{xxCAT filter to apply, as character.}

\item{scat}{xxSCAT filter to apply, as character.}

\item{duplicate_function}{The function to apply if multiple covariate values
are found by day.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A nif object with a new column added that contains the time-varying
covariate values. The name of this column is determined by the \code{covariate}
parameter (or defaults to the value of \code{testcd} if not specified). The
covariate values are matched to the nif object by USUBJID and date. For
each subject, missing covariate values are filled using the last observed
value (carrying forward).
}
\description{
Add a column to a \link{nif} object representing a time-varying covariate. In
contrast to observations (see \code{add_observation()}), covariates are not
captured as rows of observation events but are attached as a separate column
to the observations in a nif object. The values reflect the status of the
covariate at the time of the existing observations and are carried forward as
needed.
}
\details{
Covariate data may come from any domain, and like for observations, their
source is defined by the \code{domain} and \code{testcd} arguments. Covariate
observations can be further specified with the \code{cat} and \code{scat} arguments
that refer to the 'xxCAT' and 'xxSCAT' fields of the source domain, and the
\code{observation_filter} argument. This may be necessary, when observations
defined by the \code{testcd} alone are ambiguous.

In general, the covariate value and the respective observation time stamp are
taken from the 'xxSTRESN' and 'xxDTC' fields of the source (where xx refers
to the domain code). Other fields can be specified by the \code{dv_field} and
\code{dtc_field} arguments.

The name of the covariate columns can be specified by the \code{covariate}
argument. By default, it is set to the 'testcd' (without any prefix).
}
\examples{
add_covariate(examplinib_poc_nif, examplinib_poc, "vs", "WEIGHT",
  covariate = "wt"
) |>
head()
}
\seealso{
\code{\link[=add_baseline]{add_baseline()}}
}
