% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ddt.R
\name{add_dd}
\alias{add_dd}
\title{Add Field definition to Data Definition Table}
\usage{
add_dd(
  obj,
  name,
  definition,
  type,
  description,
  unit = NA_character_,
  source = ""
)
}
\arguments{
\item{obj}{Data definition Table as data frame.}

\item{name}{Field name as character.}

\item{definition}{Field definition as character.}

\item{type}{Field type as character.}

\item{description}{Field description as character.}

\item{unit}{Field unit as character.}

\item{source}{Field source as character.}
}
\value{
A data frame.
}
\description{
Add Field definition to Data Definition Table
}
