% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_domain}
\alias{add_domain}
\title{Add domain data}
\usage{
add_domain(obj, dom_name, dom_data)
}
\arguments{
\item{obj}{A sdtm object.}

\item{dom_name}{The domain name as character.}

\item{dom_data}{The domain data as data frame.}
}
\value{
A sdtm object.
}
\description{
Add domain data
}
