% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif_deprecated.R
\name{add_generic_observation}
\alias{add_generic_observation}
\title{Add an observation from a generic source}
\usage{
add_generic_observation(
  obj,
  source,
  DTC_field,
  selector,
  DV_field,
  analyte_name,
  cmt = NULL,
  parent = NA,
  silent = FALSE
)
}
\arguments{
\item{obj}{A NIF object.}

\item{source}{The data source as data frame.}

\item{DTC_field}{The date-time information field as character.}

\item{selector}{R code to conduct filtering of the data set.}

\item{DV_field}{The DV field as data character.}

\item{analyte_name}{The name of the analyte in the output, as character.}

\item{cmt}{The comopartment for the added observation.}

\item{parent}{The parent as character. If provided, the parent field defines
which parent compound administration is taken as the reference time when
time after dose (TAD) is calculated.}

\item{silent}{Suppress messages.}
}
\value{
A NIF object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
add_generic_observation(examplinib_poc_nif, domain(examplinib_poc, "vs"),
  "VSDTC", {VSTESTCD == "WEIGHT"}, "VSSTRESN", "wt")
  # ->
add_observation(examplinib_poc_nif, examplinib_poc, "vs", "WEIGHT",
  analyte = "wt")
}
\keyword{internal}
