% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_lab_covariate}
\alias{add_lab_covariate}
\title{#' Add baseline creatinine clearance field.
#'
#' @param obj A NIF data set.
#' @param method The function to calculate eGFR (CrCL) from serum creatinine.
#' Currently either: egfr_mdrd, egfr_cg or egfr_raynaud
#'
#' @return A NIF data set.
#' @seealso \code{\link[=egfr_mdrd]{egfr_mdrd()}}
#' @seealso \code{\link[=egfr_cg]{egfr_cg()}}
#' @seealso \code{\link[=egfr_raynaud]{egfr_raynaud()}}
#' @export
add_bl_crcl <- function(obj, method=egfr_mdrd) {
if("BL_CREAT" \%in\% colnames(obj)) {
obj <- obj \%>\%
as.data.frame() \%>\%
mutate(BL_CRCL=method(BL_CREAT, AGE, SEX, RACE, molar=T)) \%>\%
new_nif()
}
return(obj)
}
Add lab covariate}
\usage{
add_lab_covariate(obj, lb, lbspec = "SERUM", lbtestcd, silent = F)
}
\arguments{
\item{obj}{The NIF data set.}

\item{lb}{The LB SDTM domain}

\item{lbspec}{The specimen, e.g., SERUM.}

\item{lbtestcd}{Lab parameters to be included as character scalar or vector.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
A NIF data set
}
\description{
This functions adds columns for the lab parameters specified in \code{lbtestcd}, in
a time-varying way, i.e., the actual lab value at the time of the observation
or administration. This is in contrast to  \code{\link[=add_bl_lab]{add_bl_lab()}}.
In rows of missing lab data, the last value is carried forward.
}
\details{
Note that for some lab parameters, e.g., leukocytes, bili, etc., there may
be observations both in serum and in urine. It is therefore necessary to
specify the specimen tested. This corresponds to the \code{LBSPEC} field used in
the LB SDTM domain.
}
\seealso{
\code{\link[=add_bl_lab]{add_bl_lab()}}

\code{\link[=add_lab_observation]{add_lab_observation()}}
}
