% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_lab_observation}
\alias{add_lab_observation}
\title{Add Lab value as observation}
\usage{
add_lab_observation(obj, lb, lbtestcd, cmt, lbspec = "", silent = F)
}
\arguments{
\item{obj}{The NIF data set.}

\item{lb}{The LB SDTM domain.}

\item{lbtestcd}{The LBTESTCD.}

\item{cmt}{A numerical value to specify the compartment this observation is
assigned to.}

\item{lbspec}{The LBSPECT.}

\item{silent}{Don't issue warnings.}
}
\value{
The resulting NIF data set.
}
\description{
Add Lab value as observation
}
