% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_mapping}
\alias{add_mapping}
\title{Add a treatment-analyte mapping to an sdtm object}
\usage{
add_mapping(obj, EXTRT = "", PCTESTCD = "")
}
\arguments{
\item{obj}{A sdtm object}

\item{EXTRT}{The treatment from EX}

\item{PCTESTCD}{The analyte from PC}
}
\description{
Add a treatment-analyte mapping to an sdtm object
}
