% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_mapping.sdtm}
\alias{add_mapping.sdtm}
\title{Add a treatment-analyte mapping to an sdtm object}
\usage{
\method{add_mapping}{sdtm}(obj, extrt = "", pctestcd = "")
}
\arguments{
\item{obj}{A sdtm object}

\item{extrt}{The treatment from EX}

\item{pctestcd}{The analyte from PC}
}
\description{
Add a treatment-analyte mapping to an sdtm object
}
