% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_metabolite_mapping}
\alias{add_metabolite_mapping}
\title{Attach a parent-metabolite mapping to a SDTM object.}
\usage{
add_metabolite_mapping(obj, pctestcd_parent = "", pctestcd_metabolite = "")
}
\arguments{
\item{obj}{The SDTM object.}

\item{pctestcd_parent}{The PCTESTCD of the parent compound.}

\item{pctestcd_metabolite}{The PCTESTCD of the metabolite.}
}
\value{
The SDTM object.
}
\description{
Attach a parent-metabolite mapping to a SDTM object.
}
