% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{add_RFENDTC}
\alias{add_RFENDTC}
\title{Title}
\usage{
add_RFENDTC(dm, ex)
}
\arguments{
\item{dm}{The DM domain as data frame.}

\item{ex}{The EX domain as data frame.}
}
\value{
The updated DM domain as data frame.
}
\description{
Title
}
\keyword{internal}
