% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_rtb}
\alias{add_rtb}
\title{Add baseline and relative-to-baseline fields}
\usage{
add_rtb(obj, baseline_filter = "TIME <= 0", summary_function = median)
}
\arguments{
\item{obj}{A NIF object.}

\item{baseline_filter}{A filter term to identify the baseline condition.}

\item{summary_function}{The function to derive the baseline. This function is
applied over the DV values identified by the 'baseline_filter' term. The
default function is \code{median}. Alternatively, \code{mean}, \code{min} or \code{max} can be
considered.}
}
\value{
A NIF object
}
\description{
Add baseline and relative-to-baseline fields
}
\details{
Output fields:
\itemize{
\item \code{DVBL} Baseline value for the dependent variable DV.
\item \code{DVRTB} Relative-to-baseline value for the dependent variable DV.
}

The Baseline is calculated as the median (or the summary function
output) of the DV field for all time points identified by the
baseline_filter' term.
}
\examples{
head(add_rtb(examplinib_poc_nif))
head(add_rtb(examplinib_poc_min_nif))
}
