% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_tafd}
\alias{add_tafd}
\title{Add time after first dose column}
\usage{
add_tafd(nif)
}
\arguments{
\item{nif}{A NIF object.}
}
\value{
A NIF object.
}
\description{
Add time after first dose column
}
\examples{
add_tafd(examplinib_poc_nif)
}
