% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_time}
\alias{add_time}
\title{Add TIME field to table}
\usage{
add_time(x)
}
\arguments{
\item{x}{The table as data frame.}
}
\value{
A data frame with FIRSTDTC and TIME added.
}
\description{
TIME is created as the difference between the DTC field and the first DTC
field on the USUBJID level. TIME is in hours, rounded by 3 digits.
}
