% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pivot.R
\name{add_time_deviation}
\alias{add_time_deviation}
\title{Calculate time deviations for observations}
\usage{
add_time_deviation(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object with the 'TIME_DEV' field added
}
\description{
Calculate the time difference in hours between the actual time after dose
(TAD) and the scheduled time (NTIME) of observations.
}
\details{
If NTIME == 0, the data point is considered pre-dose, and instead of TAD, the
(negative) time difference to the next administration is calculated!
}
\examples{
library(dplyr)

examplinib_poc_nif \%>\%
  add_time_deviation() \%>\%
  head()

}
