% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_mappings.R
\name{add_time_mapping}
\alias{add_time_mapping}
\title{Attach a time mapping to an sdtm object}
\usage{
add_time_mapping(obj, ...)
}
\arguments{
\item{obj}{The SDTM object.}

\item{...}{Mappings in the form '"\if{html}{\out{<PCTPT>}}"=\if{html}{\out{<NTIME>}}' with multiple mappings
separated by commas. \if{html}{\out{<PCTPT>}} corresponds to the value in the PCTPT fields,
and NTIME corresponds to the nominal time in hours.}
}
\value{
The SDTM object
}
\description{
The nominal time of observations (e.g., \code{PCTPT}) is not required
to follow a strict format and is in most cases provided as a composite
string. This function can be used to explicitly define the nominal
observation times (in hours) for the values of, e.g., \code{PCTPT}.
}
\seealso{
\code{\link[=suggest]{suggest()}}
}
\keyword{internal}
