% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_time_mapping.sdtm}
\alias{add_time_mapping.sdtm}
\title{Attach a time mapping to an sdtm object}
\usage{
\method{add_time_mapping}{sdtm}(obj, ...)
}
\arguments{
\item{obj}{The sdtm object.}

\item{...}{Mappings in the form '"\if{html}{\out{<PCTPT>}}"=\if{html}{\out{<NTIME>}}' with multiple mappings
separated by commas. \if{html}{\out{<PCTPT>}} corresponds to the value in the PCTPT fiels,
and NTIME corresponds to the nominal time in hours.}
}
\value{
The modified sdtm object
}
\description{
Attach a time mapping to an sdtm object
}
\seealso{
\code{\link[=suggest]{suggest()}}
}
