% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_window.R
\name{add_time_window_flag}
\alias{add_time_window_flag}
\title{Flag time window violations}
\usage{
add_time_window_flag(
  obj,
  window,
  analyte = NULL,
  use_minutes = TRUE,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{window}{Time window definition as data frame.}

\item{analyte}{The analyte to apply the time window to.}

\item{use_minutes}{time window is given in minutes rather than hours,
defaults to TRUE.}

\item{silent}{Suppress messages, as logical.}
}
\value{
A nif object with TIME_DEV, EXCL, EXCL_REASON fields added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

.BEFORE and .AFTER are inclusive!
}
