% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_tv_lab}
\alias{add_tv_lab}
\title{Add time-variant lab covariate}
\usage{
add_tv_lab(obj, lb, lbtestcd, lbspec = NULL, silent = FALSE)
}
\arguments{
\item{obj}{A NIF dataset.}

\item{lb}{SDTM LB domain as data frame.}

\item{lbtestcd}{Lab parameter(s) as encoded by LBTESTCD.}

\item{lbspec}{The specimen, e.g., "BLOOD", "SERUM" or "URINE".}

\item{silent}{Disable messages.}
}
\value{
A NIF object.
}
\description{
Add column named 'TV_xxx' with the time-varying values for the respective
lab covariate. Last values are carried forward.
}
\details{
Lab parameters not found in LB will be reported in a warning message.
}
