% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_tv_vs}
\alias{add_tv_vs}
\title{Add time-variant vital sign covariate}
\usage{
add_tv_vs(obj, vs, vstestcd, duplicate_function = mean, silent = FALSE)
}
\arguments{
\item{obj}{A NIF dataset.}

\item{vs}{SDTM VS domain as data frame.}

\item{vstestcd}{Vital sign parameter(s) as encoded by VSTESTCD.}

\item{duplicate_function}{A function to resolve duplicate entries into a
single entry, e.g., \code{mean}, \code{min}, or \code{median}. Defaults to \code{mean}.}

\item{silent}{Disable messages.}
}
\value{
A NIF object.
}
\description{
Add column named 'TV_xxx' with the time-varying values for the respective
vital sign covariate. Last values are carried forward.
}
