% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{administration_summary}
\alias{administration_summary}
\title{Administration summary}
\usage{
administration_summary(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A data frame.
}
\description{
Summary statistics on the number of administrations per subject, by parent.
This may useful to characterized the population exposure to a treatment.
}
\examples{
administration_summary(examplinib_poc_nif)
}
