% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{analyte_overview}
\alias{analyte_overview}
\title{Overview on analytes and parents}
\usage{
analyte_overview(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A data frame.
}
\description{
Overview on analytes and parents
}
\examples{
analyte_overview(examplinib_poc_nif)
}
