% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{analytes.data.frame}
\alias{analytes.data.frame}
\title{Analytes within a NIF object}
\usage{
\method{analytes}{data.frame}(obj)
}
\arguments{
\item{obj}{A NIF object}
}
\value{
Character.
}
\description{
All analytes found in the NIF object. If the field 'ANALYTE' is not present,
The analyte title is derived from the compartment.
}
\examples{
analytes(examplinib_fe_nif)
analytes(examplinib_poc_nif)
analytes(examplinib_poc_min_nif)
}
