% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{analytes}
\alias{analytes}
\title{Analytes within a NIF or SDTM object}
\usage{
analytes(nif)
}
\arguments{
\item{obj}{A nif or sdtm object.}
}
\value{
Character.
}
\description{
Analytes within a NIF or SDTM object
}
\examples{
analytes(examplinib_fe_nif)
analytes(examplinib_poc_nif)
analytes(examplinib_poc_min_nif)
}
