% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{analytes.sdtm}
\alias{analytes.sdtm}
\title{Analytes in a sdtm object}
\usage{
\method{analytes}{sdtm}(obj)
}
\arguments{
\item{obj}{The sdtm object.}
}
\value{
Character.
}
\description{
Analytes in a sdtm object
}
\examples{
analytes(examplinib_sad)
}
\keyword{internal}
