% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{analytes_cmts}
\alias{analytes_cmts}
\title{Analytes and compartments within a NIF object}
\usage{
analytes_cmts(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A data frame.
}
\description{
Analytes and compartments within a NIF object
}
\examples{
analytes_cmts(examplinib_poc_nif)
analytes_cmts(examplinib_poc_min_nif)
}
