% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_statement}
\alias{append_statement}
\title{Add string to string, separated by a comma}
\usage{
append_statement(object, statement)
}
\arguments{
\item{object}{The base string as character.}

\item{statement}{The string to be added as character.}
}
\value{
A string
}
\description{
Add string to string, separated by a comma
}
\examples{
append_statement("A", "B")
append_statement("", "B")
append_statement(c("A1", "A2", "A3"), "B")
append_statement(c("A1", "", "A3", ""), "B")
}
\keyword{internal}
