% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_cat_filter}
\alias{apply_cat_filter}
\title{Apply CAT and SCAT filter}
\usage{
apply_cat_filter(obj, param, param_field)
}
\arguments{
\item{obj}{A data frame.}

\item{param}{The parameter (e.g., the cat or scat value).}

\item{param_field}{The field in the data frame (e.g., PCCAT, PCSCAT).}
}
\value{
A data frame.
}
\description{
Apply CAT and SCAT filter
}
