% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_auto.R
\name{auto_mapping}
\alias{auto_mapping}
\title{Create analyte mapping table from mapping formulae}
\usage{
auto_mapping(sdtm, ...)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{...}{Analyte mappings as formulae.}
}
\value{
A data frame with unique combinations of EXTRT and PCTESTCD.
}
\description{
Creates a mapping between treatments (EXTRT) and analytes (PCTESTCD). Each
combination of EXTRT and PCTESTCD must be unique. If duplicate mappings are
found, an error will be raised.
}
