% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif_deprecated.R
\name{baseline_covariates}
\alias{baseline_covariates}
\title{Extract baseline vital sign covariates from VS}
\usage{
baseline_covariates(vs)
}
\arguments{
\item{vs}{The VS domain as data frame.}
}
\value{
Baseline VS data as wide data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
