% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{bmi_by_age}
\alias{bmi_by_age}
\title{BMI by age scatterplot}
\usage{
bmi_by_age(obj, alpha = 0.7)
}
\arguments{
\item{obj}{The NIF object.}

\item{alpha}{The alpha for the point rendering.}
}
\value{
A plot object.
}
\description{
BMI by age scatterplot
}
\examples{
bmi_by_age(examplinib_poc_nif)
}
\keyword{internal}
