% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{bsa_dubois}
\alias{bsa_dubois}
\title{Body surface area (Du Bois formula)}
\usage{
bsa_dubois(weight, height)
}
\arguments{
\item{weight}{Body weight in kg, as numeric.}

\item{height}{Body height in cm, as numeric.}
}
\value{
BSA in m^2
}
\description{
Source: Du Bois D, Du Bois EF. A formula to estimate the approximate surface
area if height and weight be known. 1916. Nutrition.
1989 Sep-Oct;5(5):303-11; discussion 312-3.
\doi{10.1001/archinte.1916.00080130010002}
}
