% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{bsa_mosteller}
\alias{bsa_mosteller}
\title{Body surface area (Mosteller formula)}
\usage{
bsa_mosteller(weight, height)
}
\arguments{
\item{weight}{Body weight in kg, as numeric.}

\item{height}{Body height in cm, as numeric.}
}
\value{
BSA in m^2
}
\description{
Source: Mosteller RD. Simplified calculation of body-surface area.
N Engl J Med. 1987 Oct 22;317(17):1098.
\doi{10.1056/NEJM198710223171717}
}
