% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{carry_forward_dose}
\alias{carry_forward_dose}
\title{Complete DOSE field}
\usage{
carry_forward_dose(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
Complete DOSE field
}
\keyword{internal}
