% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{carry_forward_epoch}
\alias{carry_forward_epoch}
\title{Complete EPOCH field}
\usage{
carry_forward_epoch(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
Complete EPOCH field
}
\keyword{internal}
