% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{cat_boxplot}
\alias{cat_boxplot}
\title{boxplot over categories}
\usage{
cat_boxplot(obj, cat_field, val_field, title = NULL)
}
\arguments{
\item{obj}{The nif object.}

\item{cat_field}{The category field.}

\item{val_field}{The value field.}

\item{title}{The figure title.}
}
\value{
A ggplot2 object.
}
\description{
boxplot over categories
}
