% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdiscpilot01_adam}
\alias{cdiscpilot01_adam}
\title{Pharmaverse sample ADaM data}
\format{
An object of class \code{adam} (inherits from \code{list}) of length 4.
}
\source{
\url{https://github.com/pharmaverse/pharmaverseadam}
}
\usage{
cdiscpilot01_adam
}
\description{
An adam object based on the CDISCPILOT01 data set exported from the
'pharmaverseadam' package:
}
\keyword{datasets}
