% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdiscpilot01_sdtm}
\alias{cdiscpilot01_sdtm}
\title{Pharmaverse sample SDTM data}
\format{
sdtm object with 9 domains:

\preformatted{
-------- SDTM data summary --------
Study CDISCPILOT01

Safety and Efficacy of the Xanomeline Transdermal Therapeutic System (TTS) in
Patients with Mild to Moderate Alzheimers Disease.

Data disposition
  DOMAIN   SUBJECTS   OBSERVATIONS
  dm       306        306
  ex       254        591
  vs       254        29643
  pc       254        4572
  pp       168        2688
  lb       254        59580
  ts       0          0
  eg       254        26717
  ae       225        1191

Arms (DM):
  ACTARMCD   ACTARM
  Pbo        Placebo
  Scrnfail   Screen Failure
  Xan_Hi     Xanomeline High Dose
  Xan_Lo     Xanomeline Low Dose

Treatments (EX):
  PLACEBO, XANOMELINE

PK sample specimens (PC):
  PLASMA, URINE

PK analytes (PC):
  PCTEST       PCTESTCD
  XANOMELINE   XAN
}
}
\source{
\url{https://github.com/pharmaverse/pharmaversesdtm}
}
\usage{
cdiscpilot01_sdtm
}
\description{
An sdtm object based on the CDISCPILOT01 data set exported from the
'pharmaversesdtm' package:
}
\details{
Patil L, Bundfuss S, Dahnert K, Gautier F, Mancini E, Namai T, Nguyen V,
Shuliar V (2025).
\emph{pharmaversesdtm: SDTM Test Data for the 'Pharmaverse' Family of Packages}.
\url{https://github.com/pharmaverse/pharmaversesdtm}.
}
\keyword{datasets}
