% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_check.R
\name{check.nif}
\alias{check.nif}
\title{Check nif object for potential issues}
\usage{
\method{check}{nif}(
  obj,
  analyte = NULL,
  ntime_threshold = 0.2,
  ref_time = "TAD",
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{ntime_threshold}{The allowed fractional difference between TAD and NTIME.}

\item{ref_time}{Time field}

\item{silent}{Suppress messages.}
}
\value{
The nif object with the CHECK field added.
}
\description{
Check nif object for potential issues
}
