% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{check.sdtm}
\alias{check.sdtm}
\title{Check SDTM domains DM, EX and PC for missing date and time information.}
\usage{
\method{check}{sdtm}(obj, verbose = TRUE, ...)
}
\arguments{
\item{obj}{The SDTM as SDTM object.}

\item{verbose}{Boolean to indicate whether to include details.}

\item{...}{Further arguments.}
}
\value{
Nothing.
}
\description{
Check SDTM domains DM, EX and PC for missing date and time information.
}
\examples{
check(examplinib_poc)
}
