% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{check_date_time_format}
\alias{check_date_time_format}
\title{Check for incomplete date-time format in columns ending with "DTC"}
\usage{
check_date_time_format(obj, verbose = TRUE)
}
\arguments{
\item{obj}{The SDTM domain as data frame.}

\item{verbose}{Boolean to indicate whether to include details.}
}
\value{
The unchanged SDTM domain.
}
\description{
Check for incomplete date-time format in columns ending with "DTC"
}
