% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{check_last_exendtc}
\alias{check_last_exendtc}
\title{Check EXENDTC for the last administration.}
\usage{
check_last_exendtc(ex, verbose = TRUE)
}
\arguments{
\item{ex}{The SDTM EX domain as data frame.}

\item{verbose}{Boolean to indicate whether to include details.}
}
\value{
The unchanged EX domain.
}
\description{
Absent EXENDTC in the last administration by subject and treatment may
indicate that at the time of the data cut-off, this subject was still on
treatment.
}
