% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{check_missing_time}
\alias{check_missing_time}
\title{Check for missing time in columns ending with "DTC"}
\usage{
check_missing_time(obj, verbose = TRUE)
}
\arguments{
\item{obj}{The SDTM domain as data frame.}

\item{verbose}{Boolean to indicate whether to include details.}
}
\value{
The unchanged SDTM domain.
}
\description{
Check for missing time in columns ending with "DTC"
}
