% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{check_sdtm}
\alias{check_sdtm}
\title{Check SDTM domains DM, EX and PC for missing date and time information.}
\usage{
check_sdtm(sdtm, verbose = TRUE)
}
\arguments{
\item{sdtm}{The SDTM as SDTM object.}

\item{verbose}{Boolean to indicate whether to include details.}
}
\value{
Nothing.
}
\description{
Check SDTM domains DM, EX and PC for missing date and time information.
}
\examples{
check_sdtm(examplinib_poc)

}
