% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{cmt_mapping}
\alias{cmt_mapping}
\title{Analyte - compartment mapping}
\usage{
cmt_mapping(obj)
}
\arguments{
\item{obj}{A NIF object}
}
\value{
A data frame
}
\description{
Analyte - compartment mapping
}
\examples{
cmt_mapping(examplinib_poc_nif)
cmt_mapping(examplinib_poc_min_nif)
}
