% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalescing join}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  keep = c("left", "right"),
  suffix = c(".x", ".y"),
  join = c("full_join", "left_join", "right_join", "inner_join")
)
}
\arguments{
\item{x}{Left, as data frame.}

\item{y}{Right, as data frame.}

\item{by}{The 'by' argument of the join functions.}

\item{keep}{'left' means keep value from left table if values exist in both
tables.}

\item{suffix}{Same as the suffix argument in dplyr joins.}

\item{join}{Choose a join type from the list. The default is full_join.}
}
\value{
A data frame.
}
\description{
Source: https://www.r-bloggers.com/2023/05/replace-missing-value-from-other-columns-using-coalesce-join-in-dplyr/
}
