% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{compartments}
\alias{compartments}
\title{Compartments used in a nif object}
\usage{
compartments(obj)
}
\arguments{
\item{obj}{A NIF object}
}
\value{
A data frame
}
\description{
Compartments used in a nif object
}
\examples{
compartments(examplinib_poc_nif)
compartments(examplinib_poc_min_nif)
}
