% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compose_dtc}
\alias{compose_dtc}
\title{Compose DTC from date and time components}
\usage{
compose_dtc(date, time)
}
\arguments{
\item{date}{A date in POSIX or character format.}

\item{time}{A time in character format.}
}
\value{
A POSICct object.
}
\description{
Convert the date and time provided separately as character to a POSIXct
date-time object.
}
\examples{
compose_dtc(date = "2022-09-29", time = "09:30")
}
\keyword{internal}
