% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{compress.nif}
\alias{compress.nif}
\title{Compress a NIF data set}
\usage{
\method{compress}{nif}(nif, fields = NULL)
}
\arguments{
\item{nif}{A NIF data set object.}

\item{fields}{Fields to be included as character. If 'fields' is NULL
(default), the fields defined in \code{standard_nif_fields} plus all fields with
a name that starts with 'BL_' (baseline covariates) are included.}
}
\value{
A NIF data set object.
}
\description{
Compress a NIF data set
}
