% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{compress}
\alias{compress}
\title{Generic function for compress}
\usage{
compress(nif, fields = standard_nif_fields)
}
\arguments{
\item{nif}{A NIF data set object.}

\item{fields}{Fields to be kept as character.}
}
\value{
A NIF data set object.
}
\description{
Generic function for compress
}
