% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{compress_nif}
\alias{compress_nif}
\title{This function cleans up a NIF data set}
\usage{
compress_nif(nif, ...)
}
\arguments{
\item{nif}{NIF dataset as created by make_raw_nif() and potentially modified for
additional covariates.}

\item{...}{Further optional parameters are fields to be included in the
output. If none are provided, the standard set will be used.}
}
\value{
A NIF dataset with only the specified fields
}
\description{
This function cleans up a NIF data set
}
