% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conditional_message}
\alias{conditional_message}
\title{Issue message based on silent flag}
\usage{
conditional_message(...)
}
\arguments{
\item{...}{Further message components.}

\item{msg}{The message as character.}

\item{silent}{A Boolean.}
}
\value{
Nothing.
}
\description{
Issue message based on silent flag
}
\keyword{internal}
