% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{convert_char_numeric}
\alias{convert_char_numeric}
\title{Title}
\usage{
convert_char_numeric(
  x,
  min_prob = 0.9,
  exclude = c("USUBJID", "STUDYID"),
  silent = NULL
)
}
\arguments{
\item{silent}{}
}
\description{
Title
}
