% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pivot.R
\name{correlate_obs}
\alias{correlate_obs}
\title{Correlate observations}
\usage{
correlate_obs(
  obj,
  indep_analyte,
  dep_analyte,
  window = 10/60,
  time_field = "TIME",
  duplicate_function = mean
)
}
\arguments{
\item{obj}{A nif object.}

\item{indep_analyte}{The independent analyte as character.}

\item{dep_analyte}{The dependent analyte(s) as character.}

\item{window}{The allowed time window between the independent and dependent
analyte observations in hours.}

\item{time_field}{Time variable to use for the correlation of observations,
can be 'DTC' or 'TIME' (default).}

\item{duplicate_function}{A function to resolve duplicate values for
observations of the dependent analyte. Defaults to \code{mean}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
