% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{country_abbr}
\alias{country_abbr}
\title{Convert ISO 3166 country code to three-letter country name.}
\usage{
country_abbr(country_code)
}
\arguments{
\item{country_code}{The numerical country code according to ISO 3166.}
}
\value{
The three-letter country abbreviation.
}
\description{
Convert ISO 3166 country code to three-letter country name.
}
