% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{country_code}
\alias{country_code}
\title{Convert three-letter country name to ISO 3166 codes}
\usage{
country_code(country_abbr)
}
\arguments{
\item{country_abbr}{The Country name in three-letter format according to
ISO 3166.}
}
\value{
The numerical country code according to ISO 3166.
}
\description{
Convert three-letter country name to ISO 3166 codes
}
