% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{covariate_barplot}
\alias{covariate_barplot}
\title{Generic covariate barplot}
\usage{
covariate_barplot(obj, cov, group = NULL, title = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{cov}{The (categorical) baseline covariate as character.}

\item{group}{A grouping variable, as character.}

\item{title}{The plot title as character. Defaults to cov, if NULL.}
}
\value{
A ggplot object.
}
\description{
Generic covariate barplot
}
\examples{
covariate_barplot(examplinib_poc_nif, "SEX")
}
