% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{covariate_hist}
\alias{covariate_hist}
\title{Generic covariate distribution histogram}
\usage{
covariate_hist(obj, cov_params, nbins = 11)
}
\arguments{
\item{obj}{The NIF file object.}

\item{cov_params}{The covariate parameter as provided by the respective row
from 'covariate_plot_parameters'.}

\item{nbins}{The number of bins to be used if no binwidth is specified.
Defaults to 11.}
}
\value{
A plot object.
}
\description{
Generic covariate distribution histogram
}
