% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{crea_mdrd}
\alias{crea_mdrd}
\title{Serum creatinine estimation from eGFR (MDRD)}
\usage{
crea_mdrd(egfr, age, sex, race = "")
}
\arguments{
\item{egfr}{EGFR in ml/min/1.73 m^2.}

\item{age}{Age in years.}

\item{sex}{Sex encoded as number (female is 1) or character (female is "F").}

\item{race}{Race as per CDISC nomenclature. Black race is identified as the
occurrence of 'black' in the value.}
}
\value{
Serum creatinine in mg/dl.
}
\description{
Inverse of the function published in
\doi{10.7326/0003-4819-130-6-199903160-00002}.
}
\details{
To convert crea from mg/dl to umol/l, multiply by 88.4.
}
