% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{crea_raynaud}
\alias{crea_raynaud}
\title{Serum creatinine estimation from eGFR (Raynaud method)}
\usage{
crea_raynaud(egfr, age, sex, race = "")
}
\arguments{
\item{egfr}{EGFR in ml/min/1.73 m^2.}

\item{age}{Age in years.}

\item{sex}{Sex encoded as number (male is 0) or character (male is "M").}

\item{race}{Race. Dummy variable for compatibility, race is not used by this
method.}
}
\value{
Serum creatinine in mg/dl.
}
\description{
Inverse of the function published in
\url{https://doi.org/10.1136/bmj-2022-073654}.
}
\details{
To convert crea from mg/dl to umol/l, multiply by 88.4.
}
\seealso{
\code{\link[=crea_mdrd]{crea_mdrd()}}
}
