% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_administrations.R
\name{date_list}
\alias{date_list}
\title{Expand dates}
\usage{
date_list(stdtc, endtc, stdy = NA, endy = NA)
}
\arguments{
\item{stdtc}{STDTC as character.}

\item{endtc}{ENDTC as character.}

\item{stdy}{STDY as numeric.}

\item{endy}{ENDY as numeric.}
}
\value{
Vector of lists: Dates and Days.
}
\description{
Expand dates
}
