% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ddt.R
\name{ddt}
\alias{ddt}
\title{Data definition table for nif object}
\usage{
ddt(obj, silent = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A data table.
}
\description{
Generate a data definition table from a nif object. Known nif columns are
annotated automatically, for unknown fields, NA values are produced. Further
specifications can be added manually using \code{add_dd()}.
}
\examples{
ddt(examplinib_sad_nif)
}
