% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decompose_dtc}
\alias{decompose_dtc}
\title{Decompose DTC field into date and time components}
\usage{
decompose_dtc(obj, dtc_field)
}
\arguments{
\item{obj}{A data frame.}

\item{dtc_field}{The field to decompose as character.}
}
\value{
A data frame.
}
\description{
Decompose DTC field into date and time components
}
\keyword{internal}
