% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Extract the individual baseline value for an analyte}
\usage{
derive_baseline(
  obj,
  analyte = NULL,
  baseline_filter = "TAFD <= 0",
  summary_function = median,
  default_baseline = NA_real_,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{The analyte to derive the baseline for, as character. Defaults
to all analytes if NULL.}

\item{baseline_filter}{The baseline condition as character, defaults to
\code{TAFD <= 0}.}

\item{summary_function}{A function to reduce multiple baseline values,
defaults to \code{median}.}

\item{default_baseline}{The default value if the baseline filter computes to
NA.}

\item{silent}{Suppress messages, as logical.}
}
\value{
A nif object with the DVBL field added for the specified analyte.
}
\description{
Extract the individual baseline value for an analyte
}
\examples{
 head(derive_baseline(examplinib_sad_nif, "RS2023"))

}
