% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{derive_cfb_analyte}
\alias{derive_cfb_analyte}
\title{Derive a new analyte with change from baseline from an existing analyte}
\usage{
derive_cfb_analyte(
  obj,
  source_analyte,
  analyte = NULL,
  baseline_filter = "TIME <= 0",
  summary_function = median,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{source_analyte}{The original analyte.}

\item{analyte}{The name of the derived analyte. Defaults to "CFB_xx" with xx
the original analyte name.}

\item{baseline_filter}{A filter term to identify the baseline condition.}

\item{summary_function}{The function to derive the baseline. This function is
applied over the DV values identified by the 'baseline_filter' term. The
default function is \code{median}. Alternatively, \code{mean}, \code{min} or \code{max} can be
considered.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
