% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{derive_rtb}
\alias{derive_rtb}
\title{Calculate ratio to baseline}
\usage{
derive_rtb(
  obj,
  analyte = NULL,
  baseline_filter = "TAFD <= 0",
  summary_function = median,
  default_baseline = NA_real_,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{The analyte to derive the baseline for, as character. Defaults
to all analytes if NULL.}

\item{baseline_filter}{The baseline condition as character, defaults to
\code{TAFD <= 0}.}

\item{summary_function}{A function to reduce multiple baseline values,
defaults to \code{median}.}

\item{default_baseline}{The default value if the baseline filter computes to
NA.}

\item{silent}{Suppress messages, as logical.}
}
\value{
A nif object with the DVBL and DVRTB fields added for the specified
analyte.
}
\description{
Derive the individual baseline value (DVBL field) and the ratio of the
current value to baseline (DVRTB field) for an analyte. For analytes with a
baseline of zero, NA values are returned.
}
\examples{
head(derive_rtb(examplinib_sad_nif))

}
