% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{df.to.string}
\alias{df.to.string}
\title{Render data frame object to string}
\usage{
df.to.string(df, indent = "")
}
\arguments{
\item{df}{The data.frame to be rendered}

\item{indent}{A string that defines the left indentation of the rendered
output.}
}
\value{
The output as string
}
\description{
This function renders a data.frame into a string similar to its representation
when printed without line numbers
}
