% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_preprint}
\alias{df_preprint}
\title{Render data frame object to string}
\usage{
df_preprint(
  df,
  indent = 0,
  n = NULL,
  header = TRUE,
  header_sep = FALSE,
  color = FALSE,
  show_none = FALSE,
  na_string = "NA",
  abbr_lines = NULL,
  abbr_threshold = NULL
)
}
\arguments{
\item{df}{The data frame to be rendered.}

\item{indent}{Indentation level, as numeric.}

\item{n}{The number of lines to be included, or all if NULL.}

\item{header}{Boolean to indicate whether the header row is to be included.}

\item{header_sep}{Show separation line after header, as logical.}

\item{color}{Print headers in grey as logical.}

\item{show_none}{Show empty data frame as 'none', as logical.}

\item{na_string}{String to use for NA values. Defaults to "NA".}

\item{abbr_lines}{The row number to which long data frames are abbreviated
in the ouput if the threshold is exceeded. Defaults to nif_option settings if
NULL.}

\item{abbr_threshold}{The row number threshold beyond which long data frames
are abbreviated. Defaults to nif_option settings if NULL.}
}
\value{
Character vector.
}
\description{
This function renders a data.frame into a character vector, similar to its
representation when printed without line numbers
}
