% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{disposition_summary}
\alias{disposition_summary}
\title{Subject disposition overview}
\usage{
disposition_summary(sdtm_data)
}
\arguments{
\item{sdtm_data}{SDTM data object.}
}
\value{
A data frame
}
\description{
Subject disposition overview
}
\examples{
disposition_summary(examplinib_sad)
}
