% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{dose_red_sbs}
\alias{dose_red_sbs}
\title{Subjects with dose reduction}
\usage{
dose_red_sbs(obj, analyte = NULL)
}
\arguments{
\item{obj}{A NIF object object.}

\item{analyte}{The treatment of interest as character. Automatically selects
the treatment if NULL but fails if there are multiple treatments.}
}
\value{
A data frame with the ID and, if available, the USUBJID of subjects
with dose reductions.
}
\description{
Subjects with dose reduction
}
\examples{
dose_red_sbs(examplinib_poc_nif)
dose_red_sbs(examplinib_poc_nif, "RS2023")

}
