% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{doses.nif}
\alias{doses.nif}
\title{Doses within a nif object}
\usage{
\method{doses}{nif}(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A number vector of all doses (AMT) in the data set.
}
\description{
Doses within a nif object
}
