% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{doses}
\alias{doses}
\title{#' @export
studies <- function(obj) {
UseMethod("studies")
}
#' Doses within a nif object
#'
#' @param obj A nif object
#' @import dplyr
#' @return A number vector of all doses (AMT) in the data set.
#' @export
doses <- function(obj) {
UseMethod("doses")
}
Doses within a nif object}
\usage{
doses(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A number vector of all doses (AMT) in the data set.
}
\description{
#' @export
studies <- function(obj) {
UseMethod("studies")
}
#' Doses within a nif object
#'
#' @param obj A nif object
#' @import dplyr
#' @return A number vector of all doses (AMT) in the data set.
#' @export
doses <- function(obj) {
UseMethod("doses")
}
Doses within a nif object
}
