% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{doses.sdtm}
\alias{doses.sdtm}
\title{Doses in a sdtm object}
\usage{
\method{doses}{sdtm}(sdtm)
}
\arguments{
\item{sdtm}{A sdtm object.}
}
\value{
Numeric.
}
\description{
Doses in a sdtm object
}
\examples{
doses(examplinib_poc)
}
