% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dv_na_to_zero}
\alias{dv_na_to_zero}
\title{Set all NA values in DV to zero}
\usage{
dv_na_to_zero(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
Set all NA values in DV to zero
}
\examples{
dv_na_to_zero(nif::examplinib_sad_min_nif)
}
