% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{edish_plot}
\alias{edish_plot}
\title{Evaluation of drug-induced serious hepatotoxicity (eDISH) plot}
\usage{
edish_plot(
  nif,
  lb,
  enzyme = "ALT",
  show_labels = FALSE,
  nominal_time = TRUE,
  time = NULL,
  title = "eDISH plot: All time points",
  size = 3,
  alpha = 0.5,
  ...
)
}
\arguments{
\item{nif}{A nif object.}

\item{lb}{The LB SDTM domain as data frame.}

\item{enzyme}{The transaminase to be plotted on the x axis as character.}

\item{show_labels}{Show ID labels per point.}

\item{nominal_time}{Use NTIME as logical.}

\item{time}{time/nominal time filter as numeric.}

\item{title}{The plot title as character.}

\item{size}{The point size as numeric.}

\item{alpha}{The alpha value as numeric.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
DOI: 10.2165/11586600-000000000-00000
}
