% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_analysis.R
\name{egfr_mdrd}
\alias{egfr_mdrd}
\title{Glomerular filtration rate estimation from serum creatinine (MDRD)}
\usage{
egfr_mdrd(crea, age, sex, race = "", molar = F)
}
\arguments{
\item{crea}{Serum creatinine in mg/dl.}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (female is 1) or character (female is "F").}

\item{race}{Race as per CDISC nomenclature. Black race is identified as the
occurrence of 'black' in the value.}

\item{molar}{Switch to select whether the creatinine value is in mg/dl
(default) or umol/l units.
#'}
}
\value{
Estimated GFR in ml/min/1.73 m^2.
}
\description{
Reference:
\href{https://www.kidney.org/content/mdrd-study-equation}{National Kidney Foundation}
}
