% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_analysis.R
\name{egfr_raynaud}
\alias{egfr_raynaud}
\title{Glomerular filtration rate estimation from serum creatinine (Raynaud method)}
\usage{
egfr_raynaud(crea, age, sex, race = "", weight = NA, molar = F)
}
\arguments{
\item{crea}{Serum creatinine in mg/dl or umol/l (if molar=TRUE).}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (male is 0) or character (male is "M").}

\item{race}{Race. Dummy variable for compatibility, race is not used by this
method.}

\item{weight}{Body weight. Not used in this formula but included for
compatibility.}

\item{molar}{Switch to select whether the creatinine value is in mg/dl
(default) or umol/l units.}
}
\value{
Estimated GFR in ml/min/1.73 m^2.
}
\description{
Reference:
\href{https://doi.org/10.1136/bmj-2022-073654}{Raynaud, BMJ 2023;381:e073654}
}
\details{
Internally, the function uses mg/dl units
}
