% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_analyte}
\alias{ensure_analyte}
\title{Ensure that the ANALYTE field is present}
\usage{
ensure_analyte(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
If 'ANALYTE' is not in the column names, the field is created based on the
compartment (CMT).
}
\keyword{internal}
