% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ensure.R
\name{ensure_dose}
\alias{ensure_dose}
\title{Ensure that the DOSE field is present}
\usage{
ensure_dose(obj)
}
\arguments{
\item{obj}{A NIF object}
}
\value{
A NIF object with the DOSE field
}
\description{
This function ensures that the DOSE field exists in the NIF object. If it doesn't exist,
it creates it by:
\enumerate{
\item Using AMT values for dosing events (EVID == 1)
\item Filling missing values using forward and backward fill within each subject
}
}
\keyword{internal}
