% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_metabolite}
\alias{ensure_metabolite}
\title{Ensure that the METABOLITE field is present in a NIF file.}
\usage{
ensure_metabolite(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
Ensure that the METABOLITE field is present in a NIF file.
}
\keyword{internal}
