% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_ntime}
\alias{ensure_ntime}
\title{Ensure that NTIME is present in the NIF object}
\usage{
ensure_ntime(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
Experimental, do not use in production! Uses assumptions about the sampling
time points.
}
\keyword{internal}
