% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_parent}
\alias{ensure_parent}
\title{Ensure that the PARENT field is present in a NIF file.}
\usage{
ensure_parent(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
Ensure that the PARENT field is present in a NIF file.
}
\examples{
head(ensure_parent(examplinib_poc_nif))
head(ensure_parent(examplinib_poc_min_nif))
}
\keyword{internal}
