% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ensure.R
\name{ensure_tad}
\alias{ensure_tad}
\title{Ensure that TAD is present in the NIF object}
\usage{
ensure_tad(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object with the TAD field.
}
\description{
This function ensures that the time-after-dose (TAD) field exists in the NIF object.
If it doesn't exist, it creates it by calling add_tad(). TAD represents the time
elapsed since the most recent administration of the parent compound.
}
\keyword{internal}
