% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ensure.R
\name{ensure_tafd}
\alias{ensure_tafd}
\title{Ensure that TAFD is present in the NIF object}
\usage{
ensure_tafd(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object with the TAFD field.
}
\description{
This function ensures that the time-after-first-dose (TAFD) field exists in the NIF object.
If it doesn't exist, it creates it by calling add_tafd(). TAFD represents the time
elapsed since the first administration of the parent compound.
}
\keyword{internal}
