% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_time}
\alias{ensure_time}
\title{Ensure that all time fields are in the nif object}
\usage{
ensure_time(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
Ensure that all time fields are in the nif object
}
\keyword{internal}
