% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_issues.R
\name{ex_issues}
\alias{ex_issues}
\title{Identifies data inconsistencies and potential issues in the EX domain}
\usage{
ex_issues(ex, extrt, silent = TRUE)
}
\arguments{
\item{ex}{The EX domain as domain.}

\item{extrt}{The treatment, as character.}

\item{silent}{Suppress reporting, as logical.}
}
\value{
A list.
}
\description{
Identifies data inconsistencies and potential issues in the EX domain
}
\examples{
dummy <- ex_issues(domain(examplinib_poc_messy, "ex"), "EXAMPLINIB", silent = FALSE)

}
