% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_fe}
\alias{examplinib_fe}
\title{Synthetic SDTM data from a fictional clinical food effect study for
'examplinib'.}
\format{
A sdtm object
}
\usage{
examplinib_fe
}
\description{
Synthetic SDTM data from a fictional clinical food effect study for
'examplinib'.
}
\keyword{datasets}
