% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_fe_nif}
\alias{examplinib_fe_nif}
\title{Synthetic NIF data from a fictional clinical food effect study for
'examplinib'.}
\format{
A NIF object
}
\usage{
examplinib_fe_nif
}
\description{
Synthetic NIF data from a fictional clinical food effect study for
'examplinib'.
}
\keyword{datasets}
