% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_nif}
\alias{examplinib_nif}
\title{Sample nif data}
\format{
A nif object
}
\usage{
examplinib_nif
}
\description{
Synthetic nif data from the fictional 'examplinib' sdtm data set.
}
\keyword{datasets}
