% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_poc}
\alias{examplinib_poc}
\title{Synthetic SDTM data from a fictional clinical single-arm study with rich and
sparse pk sampling.}
\format{
A sdtm object
}
\usage{
examplinib_poc
}
\description{
Synthetic SDTM data from a fictional clinical single-arm study with rich and
sparse pk sampling.
}
\keyword{datasets}
