% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_poc_messy}
\alias{examplinib_poc_messy}
\title{Synthetic SDTM data from a fictional clinical single-arm study with rich and
sparse pk sampling. The ex domain contains some typical data inconsistencies
that need imputations.}
\format{
A sdtm object
}
\usage{
examplinib_poc_messy
}
\description{
Synthetic SDTM data from a fictional clinical single-arm study with rich and
sparse pk sampling. The ex domain contains some typical data inconsistencies
that need imputations.
}
\keyword{datasets}
