% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_poc_min_nif}
\alias{examplinib_poc_min_nif}
\title{Synthetic NIF data from a fictional clinical single-arm study with rich and
sparse pk sampling. Reduced to the minimally required fields.}
\format{
A NIF object
}
\usage{
examplinib_poc_min_nif
}
\description{
Synthetic NIF data from a fictional clinical single-arm study with rich and
sparse pk sampling. Reduced to the minimally required fields.
}
\keyword{internal}
