% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_sad}
\alias{examplinib_sad}
\title{Synthetic SDTM data from a fictional clinical SAD FIH study with rich PK
sampling}
\format{
A sdtm object
}
\usage{
examplinib_sad
}
\description{
Synthetic SDTM data from a fictional clinical SAD FIH study with rich PK
sampling
}
\keyword{datasets}
