% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_sad_nif}
\alias{examplinib_sad_nif}
\title{Synthetic NIF data from a fictional clinical SAD FIH study with rich PK
sampling}
\format{
A NIF object
}
\usage{
examplinib_sad_nif
}
\description{
Synthetic NIF data from a fictional clinical SAD FIH study with rich PK
sampling
}
\keyword{datasets}
