% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_administrations.R
\name{expand_ex}
\alias{expand_ex}
\title{Expand administration episodes}
\usage{
expand_ex(ex)
}
\arguments{
\item{ex}{The EX domain as data frame.}
}
\value{
A data frame.
}
\description{
Expand administration episodes
}
