% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{extract_date}
\alias{extract_date}
\title{Extract date component of a POSICct object}
\usage{
extract_date(dtc)
}
\arguments{
\item{dtc}{The POSIX-formatted datetime.}
}
\value{
The date as character.
}
\description{
Extract date component of a POSICct object
}
