% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{extract_time}
\alias{extract_time}
\title{Extract time component of a POSICct object}
\usage{
extract_time(dtc)
}
\arguments{
\item{dtc}{The POSIX-formatted datetime.}
}
\value{
The time as character.
}
\description{
Extract time component of a POSICct object
}
\keyword{internal}
