% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif_deprecated.R
\name{fill_bl_tv}
\alias{fill_bl_tv}
\title{Fill all baseline and time-varying covariates}
\usage{
fill_bl_tv(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
