% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{filter_correct_date_format}
\alias{filter_correct_date_format}
\title{Filter out any rows containing DTC rows with incomplete data format as per
ISO 8601.}
\usage{
filter_correct_date_format(obj, verbose = TRUE, silent = NULL)
}
\arguments{
\item{obj}{The SDTM domain as data frame.}

\item{verbose}{Boolean to indicate whether to include details.}

\item{silent}{Suppress messages, defaults to nif_option settin, if NULL.}
}
\value{
The filtered SDTM domain as data frame.
}
\description{
NA values and empty strings are preserved.
}
\keyword{internal}
