% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{filter_EXENDTC_after_EXSTDTC}
\alias{filter_EXENDTC_after_EXSTDTC}
\title{Remove administrations with EXSTDTC after EXENDTC}
\usage{
filter_EXENDTC_after_EXSTDTC(ex, dm, extrt, silent = NULL)
}
\arguments{
\item{ex}{The ex domain as data frame.}

\item{dm}{The dm domain as data frame.}

\item{extrt}{The treatment as character.}

\item{silent}{Suppress messages, defaults to nif_option setting, if NULL.}
}
\value{
A data frame.
}
\description{
Remove administrations with EXSTDTC after EXENDTC
}
