% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{filter_subject.nif}
\alias{filter_subject.nif}
\title{Filter NIF object for specific USUBJID}
\usage{
\method{filter_subject}{nif}(obj, usubjid)
}
\arguments{
\item{obj}{The NIF object.}

\item{usubjid}{The USUBJID as character}
}
\value{
A NIF object.
}
\description{
Filter NIF object for specific USUBJID
}
\examples{
filter_subject(examplinib_poc_nif, subjects(examplinib_poc_nif)[1, "USUBJID"])
}
