% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{filter_subject}
\alias{filter_subject}
\title{Filter for specific USUBJID (generic)}
\usage{
filter_subject(obj, usubjid)
}
\arguments{
\item{obj}{The object to filter.}

\item{usubjid}{The USUBJID as character.}
}
\value{
The filtered object.
}
\description{
Filter for specific USUBJID (generic)
}
\seealso{
\code{\link[=filter_subject.sdtm]{filter_subject.sdtm()}}

\code{\link[=filter_subject.nif]{filter_subject.nif()}}
}
