% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{filter_subject.sdtm}
\alias{filter_subject.sdtm}
\title{Filter sdtm object for specific USUBJID}
\usage{
\method{filter_subject}{sdtm}(obj, usubjid)
}
\arguments{
\item{obj}{The sdtm object.}

\item{usubjid}{The USUBJID as character}
}
\value{
A sdtm object.
}
\description{
Filter sdtm object for specific USUBJID
}
\examples{
filter_subject(examplinib_poc, subjects(examplinib_poc)[1, "USUBJID"])
}
