% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{filter_subjects}
\alias{filter_subjects}
\title{Filter nif object, eliminating subjects without observations or
administrations}
\usage{
filter_subjects(obj, keep_no_obs = FALSE, keep_no_admin = FALSE)
}
\arguments{
\item{obj}{A nif object.}

\item{keep_no_obs}{Retain subjects without observations, as logical.}

\item{keep_no_admin}{Retain subjects without administeration, as logical.}
}
\value{
A nif object.
}
\description{
Filter nif object, eliminating subjects without observations or
administrations
}
