% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{first_admin_dtc}
\alias{first_admin_dtc}
\title{DTC of first administration by subject}
\usage{
first_admin_dtc(x)
}
\arguments{
\item{x}{A nif object.}
}
\value{
A data frame.
}
\description{
DTC of first administration by subject
}
