% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_auto.R
\name{formula_to_mapping}
\alias{formula_to_mapping}
\title{Make analyte mapping table from formula object}
\usage{
formula_to_mapping(sdtm, f)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{f}{A formula.}
}
\value{
A data frame.
}
\description{
The formula object must be given as PCTESTCD ~ EXTRT or, if there are multiple
analytes from the same treatment: PCTESTCD1 + PCTESTCD2 ~ EXTRT, etc.
}
\details{
The return object is a data frame with one line for each analyte. The ANALYTE
field is the PCTESTCD for the respective analyte.
}
