% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{function_name}
\alias{function_name}
\title{Get the name of a function}
\usage{
function_name(fun)
}
\arguments{
\item{fun}{A function.}
}
\value{
The name of the function as a character string.
}
\description{
Returns the name of a function that is supplied as an argument.
}
\examples{
function_name(mean)
function_name(sum)
}
\keyword{internal}
