% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{get_admin_time_from_ntime}
\alias{get_admin_time_from_ntime}
\title{Title}
\usage{
get_admin_time_from_ntime(ex, sdtm, extrt, pctestcd = NULL, silent = NULL)
}
\arguments{
\item{silent}{}
}
\description{
Title
}
