% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{get_cov_plot_params}
\alias{get_cov_plot_params}
\title{Get covariate plot parameters}
\usage{
get_cov_plot_params(field)
}
\arguments{
\item{field}{The field of the NIF object}
}
\value{
A data frame.
}
\description{
Get covariate plot parameters
}
