% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pinboard.R
\name{get_pinboard}
\alias{get_pinboard}
\title{Get pinboard object}
\usage{
get_pinboard(board = NULL)
}
\arguments{
\item{board}{The path to the board object. Defaults to the 'pinboard' value
of nif_option, or the value of the 'NIF_PINBOARD' key in the .Renviron
setting if the 'nif_option' is not set.}
}
\value{
A pinboard object.
}
\description{
Get pinboard object
}
