% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{guess_analyte}
\alias{guess_analyte}
\title{Guess the most likely meant analyte}
\usage{
guess_analyte(obj)
}
\arguments{
\item{obj}{A NIF data set}
}
\value{
The ANALYTE as character
}
\description{
Guess the most likely meant analyte
}
