% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{guess_lbspec}
\alias{guess_lbspec}
\title{Guess the most likely LBSPEC}
\usage{
guess_lbspec(lb)
}
\arguments{
\item{lb}{The LB SDTM domain as data frame.}
}
\value{
the imputed LBSPEC as character.
}
\description{
Guess the most likely LBSPEC
}
\keyword{internal}
