% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{guess_ntime}
\alias{guess_ntime}
\title{Guess NTIME from PCTPT}
\usage{
guess_ntime(sdtm)
}
\arguments{
\item{sdtm}{A sdtm object.}
}
\value{
A data frame.
}
\description{
Guess NTIME from PCTPT
}
\examples{
guess_ntime(examplinib_poc)
}
