% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{guess_parent}
\alias{guess_parent}
\title{Guess the most likely parent based on its prevalence in the NIF object}
\usage{
guess_parent(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
The parent as character
}
\description{
Guess the most likely parent based on its prevalence in the NIF object
}
\keyword{internal}
