% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{guess_spec}
\alias{guess_spec}
\title{Guess a PCSPEC}
\usage{
guess_spec(pc, silent = T)
}
\arguments{
\item{pc}{A data frame.}

\item{silent}{Boolean.}
}
\value{
The imputed spec as character.
}
\description{
Guess a PCSPEC
}
\examples{
guess_spec(examplinib_poc$pc)
}
