% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{has_domain}
\alias{has_domain}
\title{Check whether a domain is present in an SDTM object}
\usage{
has_domain(obj, name)
}
\arguments{
\item{obj}{The sdtm object.}

\item{name}{The domain name to check as a single character string.}
}
\value{
Logical indicating whether the domain exists in the SDTM object.
}
\description{
Check whether a domain is present in an SDTM object
}
\examples{
# Check if DM domain exists
has_domain(examplinib_fe, "dm")

# Check if a non-existent domain exists
has_domain(examplinib_fe, "xyz")
}
