% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{hash.domain}
\alias{hash.domain}
\title{Generate the XXH128 hash of a SDTM domain object}
\usage{
\method{hash}{domain}(obj)
}
\arguments{
\item{obj}{A SDTM domain object.}
}
\value{
The XXH128 hash of the nif object as character.
}
\description{
Generate the XXH128 hash of a SDTM domain object
}
