% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{hash.nif}
\alias{hash.nif}
\title{Generate the XXH128 hash of a nif object}
\usage{
\method{hash}{nif}(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
The XXH128 hash of the nif object as character.
}
\description{
Generate the XXH128 hash of a nif object
}
\examples{
hash(examplinib_sad_nif)
}
