% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_dataset.R, R/nif_class.R, R/sdtm_class.R,
%   R/sdtm_domain.R, R/utilities.R
\name{hash.adam_dataset}
\alias{hash.adam_dataset}
\alias{hash.nif}
\alias{hash.sdtm}
\alias{hash.sdtm_domain}
\alias{hash}
\title{XXH128 hash}
\usage{
\method{hash}{adam_dataset}(x)

\method{hash}{nif}(x)

\method{hash}{sdtm}(x)

\method{hash}{sdtm_domain}(x)

hash(x)
}
\arguments{
\item{x}{An object.}
}
\value{
XXH128 hash as string.
}
\description{
XXH128 hash
}
