% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{hash.sdtm}
\alias{hash.sdtm}
\title{Generate the XXH128 hash of a sdtm object}
\usage{
\method{hash}{sdtm}(obj)
}
\arguments{
\item{obj}{A sdtm object.}
}
\value{
The XXH128 hash of the sdtm object as character.
}
\description{
Generate the XXH128 hash of a sdtm object
}
\examples{
hash(examplinib_sad)
}
\keyword{internal}
