% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{head.nif}
\alias{head.nif}
\title{Return the first lines of a nif object}
\usage{
\method{head}{nif}(x, ...)
}
\arguments{
\item{x}{A nif object.}

\item{...}{Further arguments.}
}
\value{
A data frame.
}
\description{
Return the first lines of a nif object
}
