% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{head}
\alias{head}
\title{Head (generic)}
\usage{
head(obj, ...)
}
\arguments{
\item{obj}{The object as data frame.}

\item{...}{Further parameters.}
}
\value{
A data frame.
}
\description{
Head (generic)
}
