% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{ht_by_wt}
\alias{ht_by_wt}
\title{Weight by height scatterplot}
\usage{
ht_by_wt(obj, alpha = 0.7)
}
\arguments{
\item{obj}{The NIF object.}

\item{alpha}{The alpha for the point rendering.}
}
\value{
A plot object.
}
\description{
Weight by height scatterplot
}
\examples{
ht_by_wt(examplinib_poc_nif)
}
