% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{import_from_connection}
\alias{import_from_connection}
\title{Import nif object from connection}
\usage{
import_from_connection(connection, format = c("csv", "NONMEM"))
}
\arguments{
\item{connection}{The connectino to read from.}

\item{format}{The input data format.}
}
\value{
A nif object.
}
\description{
Import nif object from connection
}
