% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{import_observation}
\alias{import_observation}
\title{Import observation from non-SDTM data frame}
\usage{
import_observation(
  nif,
  raw,
  analyte,
  parent = NULL,
  cmt = NULL,
  observation_filter = "TRUE",
  USUBJID_field = "USUBJID",
  DTC_field = NULL,
  NTIME_field = NA,
  DV_field = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{raw}{The raw observation data frame.}

\item{analyte}{The analyte name as character.}

\item{parent}{The parent as character.}

\item{cmt}{The compartment for the analte as numeric.}

\item{observation_filter}{Filter term, as character.}

\item{USUBJID_field}{The field specifying the USUBJID, as character.}

\item{DTC_field}{The field specifying the DTC, as character.}

\item{NTIME_field}{The field specifying the NTIME, as character.}

\item{DV_field}{The field specifying the dependent variable, as character.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
