% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\docType{data}
\name{imputation_1}
\alias{imputation_1}
\title{Alternative imputation rule set}
\format{
An object of class \code{list} of length 4.
}
\usage{
imputation_1
}
\description{
\itemize{
\item No administration time imputations
}
}
\keyword{datasets}
