% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputation_rules.R
\docType{data}
\name{imputation_rules_none}
\alias{imputation_rules_none}
\title{Void imputation rule set}
\format{
An object of class \code{list} of length 0.
}
\usage{
imputation_rules_none
}
\description{
\itemize{
\item No administration time imputations
\item No imputations on observations
}
}
\keyword{datasets}
