% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\docType{data}
\name{imputation_standard}
\alias{imputation_standard}
\title{standard imputation function}
\format{
An object of class \code{list} of length 2.
}
\usage{
imputation_standard
}
\arguments{
\item{step}{Imputation step}
}
\value{
A function
}
\description{
standard imputation function
}
\keyword{datasets}
