% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{imputation_summary}
\alias{imputation_summary}
\title{Summary of imputations in nif object}
\usage{
imputation_summary(obj, analyte = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{The analyte as character. Defaults to all analytes if NULL.}
}
\value{
A data frame.
}
\description{
Summary of imputations in nif object
}
