% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute.administration.time}
\alias{impute.administration.time}
\title{Impute missing administration times}
\usage{
impute.administration.time(admin, obs)
}
\arguments{
\item{admin}{An admin data set as created by 'make_admin()'.}

\item{obs}{An observation data set as created by 'make_obs()'.}
}
\value{
An admin data set.
}
\description{
This function fills in administration times from the PCREFDTC field when
available, and carries forward last administration dates
}
\details{
PCRFTDTC is the time for nominal_time=0
(ref: https://www.lexjansen.com/phuse/2016/pp/PP20.pdf)
}
