% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif_deprecated.R
\name{impute_admin_dtc_to_pcrftdtc}
\alias{impute_admin_dtc_to_pcrftdtc}
\title{Replace administration time with time found in PCRFTDTC}
\usage{
impute_admin_dtc_to_pcrftdtc(admin, obs, silent = FALSE)
}
\arguments{
\item{admin}{The administration data frame.}

\item{obs}{The observation data frame.}

\item{silent}{Switch to disable message output.}
}
\value{
The updated administration data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
