% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{impute_missing_baseline}
\alias{impute_missing_baseline}
\title{Impute missing baseline values}
\usage{
impute_missing_baseline(
  nif,
  baseline_fields = NULL,
  summary_function = median,
  silent = NULL
)
}
\arguments{
\item{nif}{The input nif object.}

\item{baseline_fields}{Baseline fields to impute, as character. Will be set
to default values (WEIGHT, HEIGHT, BMI, other fields starting with BL_), if
NULL.}

\item{summary_function}{A function to determine the population center value
of each baseline. Defaults to median.}

\item{silent}{Suppress messages.}
}
\value{
The nif object with missing baseline values imputed to the
respective population center value.
}
\description{
Fill individual missing baseline values, and replace missing baseline values
on subject level with the population center.
}
