% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_missing_end_time}
\alias{impute_missing_end_time}
\title{Title}
\usage{
impute_missing_end_time(ex, silent = F)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{silent}{Boolean.}
}
\value{
A data frame.
}
\description{
Title
}
