% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_missing_EXENDTC}
\alias{impute_missing_EXENDTC}
\title{Impute EXENDTC when not in the EX data set}
\usage{
impute_missing_EXENDTC(ex, cut.off.date = NA, silent = F)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{cut.off.date}{The cut-off date.}

\item{silent}{Boolean.}
}
\value{
The EX domain as data frame.
}
\description{
Impute EXENDTC when not in the EX data set
}
