% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_missing_EXENDTC_time}
\alias{impute_missing_EXENDTC_time}
\title{Impute time of EXENDTC to time of EXSTDTC, when missing.}
\usage{
impute_missing_EXENDTC_time(ex, silent = F)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{silent}{A boolean.}
}
\value{
The updated EX domain as data frame.
}
\description{
Impute time of EXENDTC to time of EXSTDTC, when missing.
}
