% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{indent_string}
\alias{indent_string}
\title{Convert indent level to padding string of spaces}
\usage{
indent_string(indent = 0)
}
\arguments{
\item{indent}{The indent level as numeric.}
}
\value{
A character string
}
\description{
Convert indent level to padding string of spaces
}
\keyword{internal}
