% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{index_nif}
\alias{index_nif}
\title{This function finalizes a NIF data set by indexing the rows}
\usage{
index_nif(nif)
}
\arguments{
\item{nif}{NIF dataset as created by make_raw_nif() and potentially modified for additional covariates}
}
\value{
A final NIF dataset including ID and REF fields
}
\description{
This function finalizes a NIF data set by indexing the rows
}
