% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{is_char_date}
\alias{is_char_date}
\title{Test vector whether consistent with Date format}
\usage{
is_char_date(x)
}
\arguments{
\item{x}{The input, as character.}
}
\value{
Logical.
}
\description{
Test vector whether consistent with Date format
}
