% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{is_char_datetime}
\alias{is_char_datetime}
\title{Test vector whether consistent with Datetime as per ISO 8601}
\usage{
is_char_datetime(x)
}
\arguments{
\item{x}{The input, as character.}
}
\value{
Logical.
}
\description{
Test vector whether consistent with Datetime as per ISO 8601
}
