% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{is_char_nonmem_numeric}
\alias{is_char_nonmem_numeric}
\title{Test vector whether consistent with NONMEM numeric format}
\usage{
is_char_nonmem_numeric(x)
}
\arguments{
\item{x}{The input, as character.}
}
\value{
Logical.
}
\description{
Test vector whether consistent with NONMEM numeric format
}
