% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_iso8601_pt}
\alias{is_iso8601_pt}
\title{Title}
\usage{
is_iso8601_pt(x)
}
\arguments{
\item{x}{A character string or vector of strings to check.}
}
\value{
A logical vector.
}
\description{
Title
}
\examples{
is_iso8601_pt("PT10M")
}
