% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_iso_date}
\alias{is_iso_date}
\title{Test whether string represents ISO 8601-formatted date}
\usage{
is_iso_date(x)
}
\arguments{
\item{x}{The input as character.}
}
\value{
Boolean.
}
\description{
The expected format is "dddd-dd-dd" with "d" a digit. This function tests
whether the above is part of the input, i.e., ISO 8601-formatted date-time
objects like "dddd-dd-ddTdd:dd" are also recognized.
}
\examples{
is_iso_date("2023-09-27")
is_iso_date(c("2023-03-21T11:55", "2023-07-18"))
}
\keyword{internal}
