% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{is_likely_numeric}
\alias{is_likely_numeric}
\title{Test whether character vector likely holds nueric values}
\usage{
is_likely_numeric(x, min_prob = 0.9)
}
\arguments{
\item{x}{The input as character.}

\item{min_prob}{The minimal fraction of values to correctly parse as numeric.}
}
\value{
Logical.
}
\description{
Test whether character vector likely holds nueric values
}
