% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{is_male}
\alias{is_male}
\title{Test whether sex is male}
\usage{
is_male(sex)
}
\arguments{
\item{sex}{Vector of numbers or strings.}
}
\value{
logical.
}
\description{
Test whether sex is male
}
