% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_valid_filter}
\alias{is_valid_filter}
\title{Test wheter a filter term is valid}
\usage{
is_valid_filter(data, filter_string)
}
\arguments{
\item{data}{A data frame.}

\item{filter_string}{A filter term as character.}
}
\value{
Logical.
}
\description{
Test wheter a filter term is valid
}
