% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{isofy_date_format}
\alias{isofy_date_format}
\title{Convert date fields to ISO 8601 format}
\usage{
isofy_date_format(obj, fields = NULL)
}
\arguments{
\item{obj}{A data frame.}

\item{fields}{Date variable names as character.}
}
\value{
A data frame.
}
\description{
Convert date fields to ISO 8601 format
}
