% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{isofy_dates}
\alias{isofy_dates}
\title{Convert all DTC fields into ISO 8601 string format}
\usage{
isofy_dates(obj)
}
\arguments{
\item{obj}{A data frame.}
}
\value{
A data frame.
}
\description{
Change all columns in the input data frame that end with 'DTC' from POSIct to
character using the ISO 8601 format.
}
\seealso{
\code{\link[=lubrify_dates]{lubrify_dates()}}
}
\keyword{internal}
