% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{last_dtc.data.frame}
\alias{last_dtc.data.frame}
\title{Last date in data.frame}
\usage{
\method{last_dtc}{data.frame}(obj)
}
\arguments{
\item{obj}{A data frame.}
}
\value{
A POSIXct scalar.
}
\description{
Last date in data.frame
}
