% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{last_dtc.domain}
\alias{last_dtc.domain}
\title{Last date in SDTM domain object}
\usage{
\method{last_dtc}{domain}(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A POSIXct scalar.
}
\description{
Last date in SDTM domain object
}
\examples{
last_dtc(examplinib_sad_nif)
}
\keyword{internal}
