% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{last_dtc.nif}
\alias{last_dtc.nif}
\title{Last date in nif object}
\usage{
\method{last_dtc}{nif}(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A POSIXct scalar.
}
\description{
Last date in nif object
}
\examples{
last_dtc(examplinib_sad_nif)
}
\keyword{internal}
