% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{last_dtc}
\alias{last_dtc}
\title{#' Generate the XXH128 hash of a nif object
#'
#' @param obj A nif object.
#'
#' @returns The XXH128 hash of the nif object as character.
#' @export
#' @importFrom rlang hash
hash.nif <- function(obj) {
validate_nif(obj)
NextMethod(obj)
}
Last recorded date-time}
\usage{
last_dtc(obj)
}
\arguments{
\item{obj}{A nif, sdtm or domain domain object.}
}
\value{
A POSIXct scalar.
}
\description{
#' Generate the XXH128 hash of a nif object
#'
#' @param obj A nif object.
#'
#' @returns The XXH128 hash of the nif object as character.
#' @export
#' @importFrom rlang hash
hash.nif <- function(obj) {
validate_nif(obj)
NextMethod(obj)
}
Last recorded date-time
}
