% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{last_dtc.sdtm}
\alias{last_dtc.sdtm}
\title{Last date in SDTM domain object}
\usage{
\method{last_dtc}{sdtm}(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A POSIXct scalar.
}
\description{
Last date in SDTM domain object
}
\examples{
last_dtc(examplinib_sad)
}
\keyword{internal}
