% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{last_ex_dtc}
\alias{last_ex_dtc}
\title{Last administration DTC}
\usage{
last_ex_dtc(ex)
}
\arguments{
\item{ex}{The EX domain as data table.}
}
\value{
The last administration in DTC format.
}
\description{
Last administration DTC
}
\keyword{internal}
