% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{last_obs_dtc}
\alias{last_obs_dtc}
\title{Extract last observation time from observation tibble}
\usage{
last_obs_dtc(obs)
}
\arguments{
\item{obs}{tibble as created with make_obs}
}
\value{
A datetime object representing the last recorded observation time
}
\description{
Extract last observation time from observation tibble
}
\keyword{internal}
