% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{lbm_boer}
\alias{lbm_boer}
\title{Lean body mass (Boer formula)}
\usage{
lbm_boer(weight, height, sex)
}
\arguments{
\item{weight}{Body weight in kg, as numeric.}

\item{height}{Body height in cm, as numeric.}

\item{sex}{Sex encoded as number (male is 0) or character (male is "M").}
}
\value{
Lean body mass in kg, as numeric. Returns NA for invalid inputs.
}
\description{
Source: Caruso D, et al., Lean Body Weight-Tailored Iodinated
Contrast Injection in Obese Patient: Boer versus James Formula. Biomed Res
Int. 2018 Aug 13;2018:8521893.
\href{https://doi.org/10.1155/2018/8521893}{doi: 10.1155/2018/8521893}.
PMID: 30186869; PMCID: PMC6110034.
}
\details{
The Boer formula is one of the most commonly used formulas for estimating
lean body mass. It provides separate equations for males and females based on
height and weight measurements.
}
\seealso{
\code{\link[=lbm_hume]{lbm_hume()}}

\code{\link[=lbm_peters]{lbm_peters()}}
}
