% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{lbm_hume}
\alias{lbm_hume}
\title{Lean body mass (Hume formula)}
\usage{
lbm_hume(weight, height, sex)
}
\arguments{
\item{weight}{Body weight in kg, as numeric.}

\item{height}{Body height in cm, as numeric.}

\item{sex}{Sex encoded as number (male is 0) or character (male is "M").}
}
\value{
Lean body mass in kg, as numeric. Returns NA for invalid inputs.
}
\description{
Source: Hume R. Prediction of lean body mass from height and weight. J Clin
Pathol. 1966 Jul;19(4):389-91, \url{doi:10.1136/jcp.19.4.389}.
}
\seealso{
\code{\link[=lbm_boer]{lbm_boer()}}

\code{\link[=lbm_peters]{lbm_peters()}}
}
