% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{lbm_peters}
\alias{lbm_peters}
\title{Lean body mass (Peters formula)}
\usage{
lbm_peters(weight, height, sex)
}
\arguments{
\item{weight}{Body weight in kg, as numeric.}

\item{height}{Body height in cm, as numeric.}

\item{sex}{Sex encoded as number (male is 0) or character (male is "M").}
}
\value{
Lean body mass in kg, as numeric.
}
\description{
Source: Peters AM, Snelling HL, Glass DM, Bird NJ. Estimation of lean body
mass in children. Br J Anaesth. 2011 May;106(5):719-23.
\href{https://doi.org/10.1093/bja/aer057}{doi: 10.1093/bja/aer057}.
PMID: 21498495.
}
\seealso{
\code{\link[=lbm_boer]{lbm_boer()}}

\code{\link[=lbm_hume]{lbm_hume()}}
}
