% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{limit}
\alias{limit}
\title{Sort nif object and add REF field}
\usage{
limit(obj, individual = TRUE, keep_no_obs_sbs = FALSE)
}
\arguments{
\item{obj}{A nif object.}

\item{individual}{Apply by ID, as logical.}

\item{keep_no_obs_sbs}{Retain subjects without observations.}
}
\value{
A nif object.
}
\description{
The input data format expected by NONMEM requires all rows ordered by ID and
TIME, and indexed sequentially on a subject level with a REF field.
Re-indexing may be required if a NIF object is extended, e.g., by merging in
further data.
}
\details{
Subset nif to rows with DTC before the last individual or global observation
}
