% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{lubrify_dates}
\alias{lubrify_dates}
\title{Convert all DTC fields into POSIXct}
\usage{
lubrify_dates(obj)
}
\arguments{
\item{obj}{A data frame.}
}
\value{
A data frame.
}
\description{
Convert all DTC fields into POSIXct
}
