% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_admin}
\alias{make_admin}
\title{Make administration data set from EX domain}
\usage{
make_admin(ex, dm, drug_mapping, cut_off_date, silent = FALSE)
}
\arguments{
\item{ex}{EX domain as data frame.}

\item{dm}{DM domain as data frame.}

\item{drug_mapping}{A data frame with the columns of EXTRT and PCTESTCD
that associate both.}

\item{cut_off_date}{The cut-off date to be used where no EXENDTC is recorded,
in POSIX format.}

\item{silent}{Switch to disable message output.}
}
\value{
A tibble with individual administrations
}
\description{
This function expands the administration ranges specified by EXSTDTC and
EXENDTC in each record of EX to that the result has individual records for
all administrations, with the time point in the TIME column.
}
\section{Specific imputations}{

If the end date (EXENDTC) is missing, i.e., the administration is ongoing
At the time of the data cutoff of the SDTM data set, EXENDTC is replaced by
cut_off_date.
}

\keyword{internal}
