% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_administration}
\alias{make_administration}
\title{Compile administration data frame}
\usage{
make_administration(
  sdtm,
  extrt,
  analyte = NA,
  cmt = 1,
  cut_off_date = NULL,
  subject_filter = {
     ACTARMCD != "SCRNFAIL"
 },
  silent = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{cleanup}{}
}
\value{
A data frame.
}
\description{
Compile administration data frame
}
\keyword{internal}
