% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{make_dm}
\alias{make_dm}
\title{Synthesize a fictional DM domain}
\usage{
make_dm(studyid = "2000950044", nsubs = 10, nsites = 5, duration = 7)
}
\arguments{
\item{studyid}{The study ID.}

\item{nsubs}{The number of subjects to synthesize.}

\item{nsites}{The number of clinical sites.}

\item{duration}{The duration of the study.}
}
\value{
The DM data as data frame.
}
\description{
Currently geared toward a food effect study, but can be adapted easily to
generate other study types.
}
