% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_drug_mapping}
\alias{make_drug_mapping}
\title{Create the drug mapping data frame for PC observations}
\usage{
make_drug_mapping(sdtm_data)
}
\arguments{
\item{sdtm_data}{The sdtm data as SDTM object.}
}
\value{
A data frame.
}
\description{
Create the drug mapping data frame for PC observations
}
\keyword{internal}
