% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{make_examplinib_food_effect_nif}
\alias{make_examplinib_food_effect_nif}
\title{#' Synthesize the package data for \code{examplinib_fe_sdtm}
#'
#' @return The sdtm object.
make_examplinib_food_effect_sdtm <- function() {
examplinib <- synthesize_sdtm() \%>\%
add_analyte_mapping("EXAMPLINIB", "RS2023")
return(examplinib)
}
Synthesize the package data for \code{examplinib_fe_nif}}
\usage{
make_examplinib_food_effect_nif()
}
\value{
The nif object.
}
\description{
#' Synthesize the package data for \code{examplinib_fe_sdtm}
#'
#' @return The sdtm object.
make_examplinib_food_effect_sdtm <- function() {
examplinib <- synthesize_sdtm() \%>\%
add_analyte_mapping("EXAMPLINIB", "RS2023")
return(examplinib)
}
Synthesize the package data for \code{examplinib_fe_nif}
}
