% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{make_examplinib_nif}
\alias{make_examplinib_nif}
\title{Synthesize the package data for \code{examplinib}}
\usage{
make_examplinib_nif()
}
\value{
The nif object.
}
\description{
Synthesize the package data for \code{examplinib}
}
