% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{make_examplinib_sad_nif}
\alias{make_examplinib_sad_nif}
\title{Make NIF data set for examplinib SAD study}
\usage{
make_examplinib_sad_nif()
}
\value{
A NIF data set.
}
\description{
Make NIF data set for examplinib SAD study
}
