% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{make_md_ex}
\alias{make_md_ex}
\title{Synthesize a fictional EX domain for single dose administration}
\usage{
make_md_ex(
  dm,
  drug = "RS2023",
  dose = 500,
  treatment_duration = 50,
  missed_prob = 0.15,
  missed_doses = T
)
}
\arguments{
\item{dm}{The DM including the subject info for whom to synthesize EX.}

\item{drug}{The name of the drug to be administered.}

\item{dose}{The dose.}

\item{treatment_duration}{The treatment duration in days.}

\item{missed_prob}{Probability to miss doses.}

\item{missed_doses}{Switch whether to include randomly missed doses as boolean.}
}
\value{
The EX domain as data frame.
}
\description{
Synthesize a fictional EX domain for single dose administration
}
\keyword{internal}
