% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_plot.R
\name{make_mean_plot_data_set}
\alias{make_mean_plot_data_set}
\title{Create mean plot data set from plot data set}
\usage{
make_mean_plot_data_set(data_set)
}
\arguments{
\item{data_set}{A list.}
}
\value{
A list.
}
\description{
Create mean plot data set from plot data set
}
\seealso{
\code{\link[=make_plot_data_set]{make_plot_data_set()}}
}
