% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{make_ntime_from_dy}
\alias{make_ntime_from_dy}
\title{Make ntime lookup table from DY field}
\usage{
make_ntime_from_dy(obj, domain = "PC")
}
\arguments{
\item{obj}{A data frame.}

\item{domain}{the domain as}
}
\value{
A data frame.
}
\description{
Make ntime lookup table from DY field
}
