% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{make_ntime_from_visitdy}
\alias{make_ntime_from_visitdy}
\title{Make ntime lookup table from VISITDY field}
\usage{
make_ntime_from_visitdy(obj, domain = "PC")
}
\arguments{
\item{obj}{A data frame.}

\item{domain}{the domain as character.}
}
\value{
A data frame.
}
\description{
Make ntime lookup table from VISITDY field
}
