% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_obs}
\alias{make_obs}
\title{Make observation data set from PC}
\usage{
make_obs(pc, time_mapping = NULL, spec = NULL, silent = F)
}
\arguments{
\item{pc}{The SDTM PC domain as a data.frame.}

\item{time_mapping}{The time mapping.}

\item{spec}{The specimen to be represented in the NIF data set as string
(e.g., "BLOOD", "PLASMA", "URINE", "FECES"). When spec is an empty string
(""), which is the default setting, the most likely specimen, i.e., "BLOOD"
or "PLASMA" is selected, depending what is found in the PC data.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
A tibble with individual observations with certain NONMEM input variables set
}
\description{
Note that the DV is converted into mg/l assuming that PCSTRESN is provided
in mg/ml
}
