% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_observation}
\alias{make_observation}
\title{Compile observation data frame}
\usage{
make_observation(
  sdtm,
  domain,
  testcd,
  analyte = NULL,
  parent = NA,
  DTC_field = NULL,
  DV_field = NULL,
  TESTCD_field = NULL,
  cmt = NA,
  observation_filter = {
     TRUE
 },
  subject_filter = {
     !ACTARMCD \%in\% c("SCRNFAIL", "NOTTRT")
 },
  NTIME_lookup = NULL,
  silent = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{domain}{The domain as character.}

\item{testcd}{The xxTESTCD entry that corresponds to the analyte of interest,
as character.}

\item{analyte}{The name for the analyte. Defaults to the 'testcd', if NULL.}

\item{parent}{The name of the parent analyte for the observation as
character.}

\item{DTC_field}{The field to use as the date-time code for the observation.
Defaults to the two-character domain name followed by 'DTC', if NULL.}

\item{DV_field}{the field to use as the dependent variable. Defaults to the
two-character domain name followed by 'STRESN', if NULL.}

\item{TESTCD_field}{The xxTESTCD field. Defaults to the two-character domain
name followed by 'TESTCD', if NULL.}

\item{cmt}{The compartment for the observation as numeric.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{NTIME_lookup}{A data frame with two columns, a column that defines the
custom nominal time information in the target domain (e.g., 'PCELTM'), and
'NTIME'. This data frame is left_join()ed into the observation data frame
to provide the NTIME field.}

\item{silent}{Suppress messages as logical.}

\item{cleanup}{Keep only necessary fields, as logical.}
}
\value{
A data frame.
}
\description{
Create a data frame of observations from a SDTM domain specified by 'domain'
where the dependent variable comes from the 'DV_field' parameter and the
timing information from the 'DTC_field' parameter.

The 'TIME' in the output is \code{NA} throughout and needs to be calculated based
on administration time point information provided separately.

If the 'NTIME_lookup' parameter is provided, 'NTIME' can be derived from a
field contained in the input data set, e.g., 'PCELTM' (see the code
examples). Otherwise, 'NTIME' will be \code{NA}.
}
