% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{make_pc}
\alias{make_pc}
\title{Synthesize a fictional PC domain.}
\usage{
make_pc(ex, dm, vs, sampling_scheme)
}
\arguments{
\item{ex}{The EX domain to syntesize PC data for.}

\item{dm}{The DM domain to syntesize PC data for.}

\item{vs}{The VS domain to syntesize PC data for.}

\item{sampling_scheme}{The sampling scheme to be used.}
}
\value{
PC data as data frame.
}
\description{
Synthesize a fictional PC domain.
}
