% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{make_sd_ex}
\alias{make_sd_ex}
\title{Synthesize a fictional EX domain for single dose administration}
\usage{
make_sd_ex(dm, admindays = c(1, 14), drug = "RS2023", dose = 500)
}
\arguments{
\item{dm}{The DM including the subject info for whom to synthesize EX.}

\item{admindays}{The treatment days as vector.}

\item{drug}{The name of the drug to be administered.}

\item{dose}{The dose.}
}
\value{
EX data as data frame.
}
\description{
Synthesize a fictional EX domain for single dose administration
}
\keyword{internal}
