% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{make_sd_pc}
\alias{make_sd_pc}
\title{Title}
\usage{
make_sd_pc(ex, dm, vs, sampling_scheme)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{dm}{The DM domain as data frame.}

\item{vs}{The VS domain as data frame.}

\item{sampling_scheme}{The PK sampling scheme as data frame.}
}
\value{
The PC domain as data frame.
}
\description{
Title
}
