% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{make_subjects.sdtm}
\alias{make_subjects.sdtm}
\title{Title}
\usage{
make_subjects.sdtm(obj, ...)
}
\arguments{
\item{...}{Further arguments.}

\item{sdtm}{A sdtm object.}
}
\value{
A data frame.
}
\description{
Title
}
