% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{make_subjects_sdtm}
\alias{make_subjects_sdtm}
\title{Make subject data frame from SDTM object}
\usage{
make_subjects_sdtm(obj, ...)
}
\arguments{
\item{...}{Further arguments.}

\item{sdtm}{A sdtm object.}
}
\value{
A data frame.
}
\description{
Make subject data frame from SDTM object
}
