% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_time}
\alias{make_time}
\title{Calculate time fields based on DTC}
\usage{
make_time(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
This function generates the following time fields:

'TIME' is the time in hours relative to the subject's first record, be it an
administration or observation event.

'TAFD' is the time in hours relative to the subject's first administration of
the respective parent. Note that if a subject has received multiple drugs
(parents), the 'TAFD' field refers to the respective first administration.
}
\keyword{internal}
