% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_time_from_TIME}
\alias{make_time_from_TIME}
\title{Calculate time fields based on TIME}
\usage{
make_time_from_TIME(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
Calculate time fields based on TIME
}
\examples{
make_time_from_TIME(examplinib_poc_min_nif)
}
