% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{max_admin_time}
\alias{max_admin_time}
\title{Maximal administration time}
\usage{
max_admin_time(obj, analytes = NULL)
}
\arguments{
\item{obj}{The NIF data set}

\item{analytes}{The analyte or analytes to filter for.}
}
\value{
A scalar representing the time in hours.
}
\description{
This function returns the time in hours of the last administration within
the data set.
}
\examples{
max_admin_time(examplinib_nif)

}
