% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{max_observation_time}
\alias{max_observation_time}
\title{Maximal observation time}
\usage{
max_observation_time(obj, analyte = NULL)
}
\arguments{
\item{obj}{The NIF object}

\item{analytes}{The analyte or analytes to filter for.}
}
\value{
A scalar representing the time in hours.
}
\description{
This function returns the time in hours of the last observation relative to
the first observation within the data set.
}
\examples{
max_observation_time(examplinib_fe_nif)
max_observation_time(examplinib_poc_nif)
max_observation_time(examplinib_poc_min_nif)
max_observation_time(examplinib_poc_min_nif, analyte = "CMT4")
}
