% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{mdrd_crea}
\alias{mdrd_crea}
\title{Serum creatinine estimation from eGFR (MDRD)}
\usage{
mdrd_crea(egfr, age, sex, race = "")
}
\arguments{
\item{egfr}{EGFR in ml/min/1.73 m^2.}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (female is 1) or character (female is "F").}

\item{race}{Race as per CDISC nomenclature. Black race is identified as the
occurrence of 'black' in the value.}
}
\value{
Serum creatinine in mg/dl.
}
\description{
Inverse of the function published in
\href{https://www.kidney.org/content/mdrd-study-equation}{National Kidney Foundation}
}
\details{
To convert crea from mg/dl to umol/l, multiply by 88.4.#'
}
