% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{mdrd_egfr}
\alias{mdrd_egfr}
\title{Glomerular filtration rate estimation from serum creatinine (MDRD)}
\usage{
mdrd_egfr(crea, age, sex, race = "")
}
\arguments{
\item{crea}{Serum creatinine in mg/dl.}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (female is 1) or character (female is "F").}

\item{race}{Race as per CDISC nomenclature. Black race is identified as the
occurrence of 'black' in the value.}
}
\value{
Estimated GFR in ml/min/1.73 m^2.
}
\description{
Reference:
\href{https://www.kidney.org/content/mdrd-study-equation}{National Kidney Foundation}
}
