% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{mean_dose_plot}
\alias{mean_dose_plot}
\title{Mean dose plot}
\usage{
mean_dose_plot(obj, analyte = NULL, title = NULL)
}
\arguments{
\item{obj}{A NIF object.}

\item{analyte}{The compound as character (i.e., the ANALYTE within the data
set).}

\item{title}{The plot title as character.}
}
\value{
A ggplot object.
}
\description{
This function plots the mean dose per day over time
}
\examples{
mean_dose_plot(examplinib_poc_nif)
}
