% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\docType{data}
\name{minimal_nif_fields}
\alias{minimal_nif_fields}
\title{Minimal nif fields}
\format{
An object of class \code{character} of length 7.
}
\usage{
minimal_nif_fields
}
\value{
A character vector of the minimal NIF fields
}
\description{
Minimal nif fields
}
\keyword{datasets}
