% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{miss_admins}
\alias{miss_admins}
\title{Create administration schedule with randomly missed doses}
\usage{
miss_admins(
  start_dtc,
  end_dtc,
  dose = 500,
  dose_red = 250,
  missed_prob = 0.15,
  red_prob = 0.2
)
}
\arguments{
\item{start_dtc}{Starting DTC.}

\item{end_dtc}{End DTC.}

\item{dose}{The normal dose.}

\item{dose_red}{The reduced dose.}

\item{missed_prob}{Probability of missing administration.}

\item{red_prob}{The probability that a subject has a dose reduction. Dose
reductions, if any, occur at a random day after day 7.}
}
\value{
EXSTDTC and EXENDTC as data frame.
}
\description{
Create administration schedule with randomly missed doses
}
\keyword{internal}
