% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_analytics.R
\name{missing_time}
\alias{missing_time}
\title{Number of DTC entries with missing time information}
\usage{
missing_time(sdtm, fields)
}
\arguments{
\item{sdtm}{A stm object.}

\item{fields}{The fields to test, as character.}
}
\value{
A data frame.
}
\description{
Number of DTC entries with missing time information
}
\examples{
missing_time(examplinib_poc, c("EXSTDTC", "EXENDTC"))
}
