% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{n_observations_per_dosing_interval}
\alias{n_observations_per_dosing_interval}
\title{Number of observations per dosing interval}
\usage{
n_observations_per_dosing_interval(obj, analyte = NULL)
}
\arguments{
\item{obj}{The NIF data set.}

\item{analyte}{The analyte to filter for as character. Optional.}
}
\value{
A data frame.
}
\description{
This function returns the number of observations by subject and analyte per
dosing interval.
}
