% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca1}
\alias{nca1}
\title{Non-compartmental analysis of NIF data}
\usage{
nca1(
  nif,
  analyte = NULL,
  parent = NULL,
  keep = NULL,
  group = NULL,
  time = "TIME",
  average_duplicates = TRUE
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte as character.}

\item{parent}{The parent as character.}

\item{keep}{Columns to keep, as character.}

\item{group}{Columns to group by, as character.}

\item{time}{The time field as character.}

\item{average_duplicates}{Average duplicate concentration values, as logical.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function is a wrapper around the NCA functions provided by the
\href{https://CRAN.R-project.org/package=PKNCA}{PKNCA} package.
}
\examples{
head(nca1(examplinib_sad_nif, time = "TAD"))
head(nca1(examplinib_fe_nif, time = "TAD", group = "FASTED"))
}
