% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca2}
\alias{nca2}
\title{Non-compartmental analysis with analyte-based grouping}
\usage{
nca2(
  obj,
  analytes = NULL,
  parameters = NULL,
  keep = NULL,
  average_duplicates = TRUE
)
}
\arguments{
\item{obj}{A NIF object containing concentration-time data}

\item{analytes}{Optional vector of analytes to analyze. If NULL, all analytes
will be analyzed.}

\item{parameters}{Optional vector of PK parameters to calculate. If NULL,
default parameters will be used.}

\item{keep}{Optional vector of additional columns to keep in the output.}

\item{average_duplicates}{Boolean to indicate whether duplicate entries
should be averaged.}
}
\value{
A data frame containing NCA results
}
\description{
This function performs NCA analysis using the PKNCA package for multiple
analytes in a NIF object. It uses analyte-specific grouping in the
concentration and dose formulas.
}
