% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_nif.R
\name{nca_overview}
\alias{nca_overview}
\title{PK parameter summary statistics by dose}
\usage{
nca_overview(nca, parameters = c("aucinf.obs", "cmax"))
}
\arguments{
\item{nca}{The NCA results as provided by \code{nca}, as data frame.}

\item{parameters}{The NCA paramters to be tabulated as character,}
}
\value{
A data frame
}
\description{
PK parameter summary statistics by dose
}
\examples{
nca_overview(nca(examplinib_sad_nif, analyte="RS2023"))
}
