% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_nif.R
\name{nca_overview_table}
\alias{nca_overview_table}
\title{Primary PK parameter summary statistics by dose.}
\usage{
nca_overview_table(nca, caption = "PK parameter overview")
}
\arguments{
\item{nca}{The NCA results as provided by \code{nca}, as data frame.}

\item{caption}{The table title}
}
\value{
Markdown-formatted table as character.
}
\description{
Primary PK parameter summary statistics by dose.
}
\examples{
nca_overview_table(nca(examplinib_sad_nif, analyte="RS2023"))
}
