% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca_power_model}
\alias{nca_power_model}
\title{Power fit for PK parameters}
\usage{
nca_power_model(nca, parameter = "aucinf.obs")
}
\arguments{
\item{nca}{PK parameters as data frame.}

\item{parameter}{The PK parameter as character.}
}
\value{
A ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Plot the PK parameter defined by \code{parameter} over \code{DOSE} and overlay with
a linear model of the log-transformed parameter and the log-transformed dose.
The slope is printed in the caption line.
}
\examples{
nca_power_model(nca(examplinib_sad_nif, analyte = "RS2023"), "aucinf.obs")
}
