% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca_summary_table}
\alias{nca_summary_table}
\title{PK parameter summary statistics table by grouping variables}
\usage{
nca_summary_table(
  nca,
  parameters = c("auclast", "cmax", "tmax", "half.life", "aucinf.obs", "AUCLST", "CMAX",
    "TMAX", "LAMZHL", "AUCIFP"),
  digits = 2,
  group = NULL
)
}
\arguments{
\item{nca}{The NCA results as provided by \code{nca}, as data frame.}

\item{parameters}{The NCA parameters to be tabulated as character.}

\item{digits}{The number of significant digits to be displayed.}

\item{group}{The grouping variable, defaults to DOSE.}
}
\value{
A data frame
}
\description{
PK parameter summary statistics table by grouping variables
}
\examples{
nca_summary_table(nca(examplinib_sad_nif, analyte = "RS2023"))
}
