% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{new_domain}
\alias{new_domain}
\title{Class constructor for domain objects}
\usage{
new_domain(domain_data)
}
\arguments{
\item{domain_data}{A data.frame.}
}
\value{
A domain object.
}
\description{
Class constructor for domain objects
}
