% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{new_nif}
\alias{new_nif}
\title{nif class constructor}
\usage{
new_nif(obj = NULL, ..., silent = NULL)
}
\arguments{
\item{obj}{A data frame containing the actual NIF data or a sdtm object.}

\item{...}{Further arguments.}

\item{silent}{Suppress messages.}
}
\value{
A nif object from the input data set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
