% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif-package.R
\docType{package}
\name{nif-package}
\alias{nif-package}
\title{nif: Build and Explore NONMEM Input Format Data Sets from Clinical Study Data}
\description{
A system to convert clinical study data from the Study Data Tabulation Model (SDTM) format into the 'NONMEM' Input Format (NIF) structure for downstream nonlinear mixed effects modeling, e.g., using 'NONMEM' or 'nlmixr2'. The core functions support the aggregation of drug administration, observations and covariate data in a stepwise manner. A further set of functions support exploratory data analysis of SDTM and NIF data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstrotmann/nif}
  \item \url{https://rstrotmann.github.io/nif/reference/index.html}
}

}
\author{
\strong{Maintainer}: Rainer Strotmann \email{rainer.strotmann@mailbox.org} (\href{https://orcid.org/0000-0003-0543-7984}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rainer Strotmann \email{rainer.strotmann@mailbox.org} (\href{https://orcid.org/0000-0003-0543-7984}{ORCID}) [copyright holder]
}

}
\keyword{internal}
