% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{nif_cleanup}
\alias{nif_cleanup}
\title{Remove non-essential fields}
\usage{
nif_cleanup(nif, keep = NULL)
}
\arguments{
\item{nif}{A nif object.}

\item{keep}{Fields to explicitly keep, as character.}
}
\value{
A nif object
}
\description{
Remove non-essential fields
}
