% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_options.R
\name{nif_disclaimer}
\alias{nif_disclaimer}
\title{Disclaimer statement}
\usage{
nif_disclaimer(disclaimer_text = NA)
}
\arguments{
\item{disclaimer_text}{Additional disclaimer text.}
}
\value{
Character.
}
\description{
Disclaimer statement
}
\examples{
nif_disclaimer()
}
