% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_options.R
\name{nif_option_value}
\alias{nif_option_value}
\title{Get the value of global options}
\usage{
nif_option_value(option)
}
\arguments{
\item{option}{The option as character.}
}
\value{
The option value, if existent, or NA.
}
\description{
Get the value of global options
}
\examples{
nif_option_get("silent")
}
\keyword{internal}
