% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_plot_id}
\alias{nif_plot_id}
\title{Plot DV time course data from individual subject}
\usage{
nif_plot_id(
  nif,
  id,
  analyte = NULL,
  y_scale = "lin",
  max_time = NA,
  lines = TRUE,
  point_size = 2,
  tad = FALSE,
  imp = "none"
)
}
\arguments{
\item{nif}{The NIF object}

\item{id}{The subject ID to be plotted}

\item{analyte}{The analytes to be displayes. Defaults to NULL (all).}

\item{y_scale}{Y-scale. Use 'scale="log"' for a logarithmic y scale. Default
is "lin".}

\item{max_time}{The right limit of the time scale}

\item{lines}{Boolean to define whether lines should be plotted.}

\item{point_size}{Point size as numeric.}

\item{tad}{Boolean to select whether time after dose (TAD) rather than TIME
should be plotted.}

\item{imp}{The IMP for which administrations are to be indicated by vertical
lines. Defaults to 'none'.}
}
\value{
the plot object
}
\description{
This function plots DV over TIME for an individual subject, id. Id can be
either the ID or the USUBJID. Administration time points are indicated with
vertical lines.
}
\examples{
nif_plot_id(examplinib_poc_nif, 1)
nif_plot_id(examplinib_poc_nif, 1, analyte="RS2023")
nif_plot_id(examplinib_poc_nif, "20230000221010001", analyte="RS2023")
nif_plot_id(examplinib_poc_nif, 8, analyte="RS2023", imp="RS2023")
nif_plot_id(examplinib_poc_nif, 8, analyte=c("RS2023", "RS2023487A"))
}
