% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_spaghetti_plot1}
\alias{nif_spaghetti_plot1}
\title{Plot nif object per subject}
\usage{
nif_spaghetti_plot1(
  x,
  analyte = NULL,
  dose = NULL,
  group = NULL,
  time_field = "TIME",
  min_time = 0,
  max_time = NULL,
  min_x = NULL,
  max_x = NULL,
  limit_time_to_observations = TRUE,
  points = FALSE,
  lines = TRUE,
  point_size = 2,
  alpha = 1,
  tad = FALSE,
  nominal_time = FALSE,
  cfb = FALSE,
  log = FALSE,
  admin = FALSE,
  summary_function = median,
  title = ""
)
}
\arguments{
\item{x}{A nif object.}

\item{analyte}{The analyte.}

\item{dose}{The dose(s) to select as numeric.}

\item{group}{The grouping variable(s).}

\item{time_field}{The field to use as the time metric, as character.}

\item{min_time}{The minimum time as numeric.}

\item{max_time}{The maximum time as numeric.}

\item{min_x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use min_time instead.}

\item{max_x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use max_time instead. values
are present.}

\item{limit_time_to_observations}{Truncate time axis to last observation, as
logical.}

\item{points}{Plot points as logical.}

\item{lines}{Plot lines as logical.}

\item{point_size}{The point size as numeric.}

\item{alpha}{The alpha value for point plotting.}

\item{tad}{Time after dose as logical.}

\item{nominal_time}{Plot NTIME on x axis, as logical.}

\item{cfb}{Change from baseline as logical.}

\item{log}{Logarithmic y scale as logical.}

\item{admin}{Plot administrations as logical.}

\item{summary_function}{The summary function to apply if multiple baseline}

\item{title}{The plot title as character.}
}
\value{
A ggplot object.
}
\description{
Plot nif object per subject
}
\examples{
nif_spaghetti_plot1(examplinib_poc_nif)
nif_spaghetti_plot1(examplinib_poc_min_nif)
nif_spaghetti_plot1(examplinib_poc_min_nif, tad = TRUE, points = TRUE,
  lines = FALSE)
nif_spaghetti_plot1(examplinib_poc_nif, dose = 500, analyte = "RS2023",
  tad = TRUE, points = TRUE, point_size = 4, alpha = 0.3, lines = FALSE,
  max_time = 12)
nif_spaghetti_plot1(examplinib_poc_nif, analyte = "RS2023", title="Test")
nif_spaghetti_plot1(examplinib_poc_nif, analyte = "RS2023",
  limit_time_to_observations = FALSE)
nif_spaghetti_plot1(examplinib_poc_nif, analyte = "RS2023", admin = TRUE)
nif_spaghetti_plot1(examplinib_poc_nif, analyte = "RS2023", , tad = TRUE,
  admin = TRUE)
nif_spaghetti_plot1(examplinib_poc_nif, analyte = "RS2023", group = "SEX")
nif_spaghetti_plot1(examplinib_poc_nif, analyte = "RS2023", tad = TRUE,
  points = TRUE, lines = FALSE, dose = 500)
nif_spaghetti_plot1(examplinib_fe_nif, analyte = "RS2023", group = c("SEX", "FASTED"))
}
\keyword{internal}
