% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_viewer.R
\name{nif_viewer}
\alias{nif_viewer}
\title{NIF viewer}
\usage{
nif_viewer(nif)
}
\arguments{
\item{nif}{The NIF data set.}
}
\value{
None
}
\description{
This function creates an interactive viewer for NIF data sets.
}
\seealso{
\code{\link[=plot]{plot()}}
}
