% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{normalize_nif}
\alias{normalize_nif}
\title{Normalize nif object}
\usage{
normalize_nif(obj, cleanup = TRUE, keep = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{cleanup}{Remove non-essential fields, as logical.}

\item{keep}{Fields to explicitly keep, as character.}
}
\value{
A nif object.
}
\description{
Order nif object, index and fill missing fields, and reduce to essential
columns.
}
