% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{obs_per_dose_level}
\alias{obs_per_dose_level}
\title{Observations per dose level}
\usage{
obs_per_dose_level(obj, analyte = NULL, group = NULL)
}
\arguments{
\item{obj}{A NIF object.}

\item{analyte}{The analyte as character. If \code{NULL} (default), all analytes
are selected.}

\item{group}{An (optional) grouping variable as character. Defaults to
'NULL'.}
}
\value{
A data frame.
}
\description{
The total number of observations, per dose level and analyte.
}
\examples{
obs_per_dose_level(examplinib_poc_nif)
obs_per_dose_level(examplinib_sad_nif)
obs_per_dose_level(examplinib_poc_nif, group = "SEX", analyte = "RS2023")
}
