% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{order_nif_columns}
\alias{order_nif_columns}
\title{Establish standard order of nif object columns}
\usage{
order_nif_columns(obj)
}
\arguments{
\item{obj}{A data frame.}
}
\value{
A data.frame.
}
\description{
Establish standard order of nif object columns
}
\examples{
order_nif_columns(examplinib_poc_nif)
}
