% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{parents}
\alias{parents}
\title{Parent compounds within a NIF object}
\usage{
parents(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
The parent compounds as character.
}
\description{
Parent compounds within a NIF object
}
\examples{
parents(examplinib_poc_nif)
parents(examplinib_poc_min_nif)
}
