% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pinboard.R
\name{pb_list_object}
\alias{pb_list_object}
\title{List contents of pinboard}
\usage{
pb_list_object(board = NULL, object_type)
}
\arguments{
\item{board}{The path to the board object. Defaults to the 'pinboard' value
of nif_option, or the value of the 'NIF_PINBOARD' key in the .Renviron
setting if the 'nif_option' is not set.}

\item{object_type}{The object to filter for (sdtm or nif)}
}
\value{
The board folder, defaults to the respective nif_option setting.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
