% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pinboard.R
\name{pb_read_nif}
\alias{pb_read_nif}
\title{Read nif object from pinboard}
\usage{
pb_read_nif(name, board = NULL)
}
\arguments{
\item{name}{The pin name, as character.}

\item{board}{The path to the board object. Defaults to the 'pinboard' value
of nif_option, or the value of the 'NIF_PINBOARD' key in the .Renviron
setting if the 'nif_option' is not set.}
}
\value{
A pinboard object.

A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
