% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pinboard.R
\name{pin_list_nif}
\alias{pin_list_nif}
\title{List nif objects in pinboard}
\usage{
pin_list_nif(board = NULL)
}
\arguments{
\item{board}{The path to the board object. Defaults to the 'pinboard' value
of nif_option, or the value of the 'NIF_PINBOARD' key in the .Renviron
setting if the 'nif_option' is not set.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated, please use \code{\link[=pb_list_nif]{pb_list_nif()}} instead.
}
\keyword{internal}
