% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{pin_list_sdtm}
\alias{pin_list_sdtm}
\title{List sdtm objects in pinboard}
\usage{
pin_list_sdtm(board = NULL)

pin_list_sdtm(board = NULL)
}
\arguments{
\item{board}{A pinboard path, as character.}
}
\value{
A data frame.

A data frame.
}
\description{
List sdtm objects in pinboard

List nif objects in pinboard
}
