% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{pin_read_nif}
\alias{pin_read_nif}
\title{Read nif object from pinboard}
\usage{
pin_read_nif(name, board = NULL)
}
\arguments{
\item{name}{The pin name, as character.}

\item{board}{The board folder, defaults to the respective nif_option setting.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
