% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{pin_read_sdtm}
\alias{pin_read_sdtm}
\title{Read sdtm object from pinboard}
\usage{
pin_read_sdtm(name, board = NULL)
}
\arguments{
\item{name}{The pin name, as character.}

\item{board}{The board folder, defaults to the respective nif_option setting.}
}
\value{
A sdtm object.
}
\description{
Read sdtm object from pinboard
}
