% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{pin_write.nif}
\alias{pin_write.nif}
\title{Write to pin board}
\usage{
\method{pin_write}{nif}(obj, name = NULL, board = NULL, title = NULL, silent = NULL)
}
\arguments{
\item{obj}{nif object.}

\item{name}{Name for the nif pin object, as character. Defaults to
'xx_nif' with xx the studies included.}

\item{board}{Path to pin board, as character, defaults to respective
nif_option setting if NULL.}

\item{title}{Title for the nif pin object, as character. Defaults to
the studies included.}

\item{silent}{Suppress messages. Defaults to nif_option setting if NULL.}
}
\value{
Nothing.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
