% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{pin_write}
\alias{pin_write}
\title{Generic function}
\usage{
pin_write(obj, name = NULL, board = NULL, title = NULL, silent = NULL)
}
\arguments{
\item{obj}{The object to pin.}

\item{name}{Name for the pin object, as character. Defaults to
'xx_sdtm' or 'xx_nif' with xx the study name.}

\item{board}{Path to pin board, as character, defaults to respective
nif_option setting if NULL.}

\item{title}{Title for the pin object, as character. Defaults to
the study name(s).}

\item{silent}{Suppress messages. Defaults to nif_option setting if NULL.}
}
\value{
Nothing.
}
\description{
Generic function
}
