% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{pin_write.sdtm}
\alias{pin_write.sdtm}
\title{Write to pin board}
\usage{
\method{pin_write}{sdtm}(obj, name = NULL, board = NULL, title = NULL, silent = NULL)
}
\arguments{
\item{obj}{sdtm object.}

\item{name}{Name for the sdtm pin object, as character. Defaults to
'xx_sdtm' with xx the study name.}

\item{board}{Path to pin board, as character, defaults to respective
nif_option setting if NULL.}

\item{title}{Title for the sdtm pin object, as character. Defaults to
the study name.}

\item{silent}{Suppress messages. Defaults to nif_option setting if NULL.}
}
\value{
Nothing.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
