% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pivot.R
\name{pivot_analytes}
\alias{pivot_analytes}
\title{Convert nif object to wide data frame by NTIME}
\usage{
pivot_analytes(obj, analyte = NULL, silent = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{Analytes to include, defaults to all, if NULL.}
}
\value{
A data frame.
}
\description{
Convert nif object to wide data frame by NTIME
}
