% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pivot.R
\name{pivot_analytes1}
\alias{pivot_analytes1}
\title{correlate observations by their actual observation time}
\usage{
pivot_analytes1(
  obj,
  indep_analyte,
  dep_analyte,
  window = 10/60,
  duplicate_function = mean
)
}
\arguments{
\item{obj}{A nif object.}

\item{indep_analyte}{The independent analyte as character.}

\item{dep_analyte}{The dependent analyte as character.}

\item{window}{The allowed time window between the independent and dependent
analyte observations in hours.}

\item{duplicate_function}{A function to resolve duplicate values for
observations of the dependent analyte. Defaults to \code{mean}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
