% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{pk_sim}
\alias{pk_sim}
\title{Simulate PK data for fictional subjects, using fictional popPK model}
\usage{
pk_sim(sbs, sampling_scheme)
}
\arguments{
\item{sbs}{The subject baseline information as data frame. Needs at least
ID, PERIOD, SEX, AGE, HEIGHT, WEIGHT, and FOOD fields.}

\item{sampling_scheme}{The PK sampling scheme as data frame. The field \code{time}
is to provide the sampling times in numerical form.}
}
\value{
The simulated PK data as data frame.
}
\description{
Based on fictional two-compartment model with sequential zero-order and
first-order absorption, elimination from central compartment.
}
