% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{plot}
\alias{plot}
\title{#' @export
plot <- function(x, y_scale="lin", max_x=NULL, analyte=NULL, mean=FALSE,
doses=NULL, points=F, id=NULL, usubjid=NULL,
group=NULL, administrations=F, nominal_time=F) {
UseMethod("plot")
}
#' @export
plot <- function(x, ...) UseMethod("plot")}
\usage{
plot(
  obj,
  dummy,
  y_scale = "lin",
  max_x = NULL,
  analyte = NULL,
  mean = FALSE,
  doses = NULL,
  points = F,
  id = NULL,
  usubjid = NULL,
  group = NULL,
  administrations = F,
  nominal_time = F
)
}
\description{
#' @export
plot <- function(x, y_scale="lin", max_x=NULL, analyte=NULL, mean=FALSE,
doses=NULL, points=F, id=NULL, usubjid=NULL,
group=NULL, administrations=F, nominal_time=F) {
UseMethod("plot")
}
#' @export
plot <- function(x, ...) UseMethod("plot")
}
