% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{plot.sdtm_domain}
\alias{plot.sdtm_domain}
\title{Plot domain object}
\usage{
\method{plot}{sdtm_domain}(
  x,
  testcd = NULL,
  points = TRUE,
  lines = FALSE,
  legend = TRUE,
  color = NULL,
  ...
)
}
\arguments{
\item{x}{A domain object.}

\item{testcd}{Testcd field to filter for, defaults to all if NULL.}

\item{points}{Plot points, as logical.}

\item{lines}{Plot lines, as logical.}

\item{legend}{Plot legend, as logical.}

\item{color}{Field to color points and lines by.}

\item{...}{Further parameters.}
}
\value{
A ggplot2 object.
}
\description{
Plot domain object
}
\examples{
plot(domain(examplinib_sad, "lb"))
}
