% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{plot_hash}
\alias{plot_hash}
\title{Plot hash as QR code}
\usage{
plot_hash(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A plot object.
}
\description{
Plot hash as QR code
}
\examples{
plot_hash(examplinib_sad_nif)
}
