% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{plot_z}
\alias{plot_z}
\title{Title}
\usage{
plot_z(
  nif,
  analyte = NULL,
  dose = NULL,
  log = TRUE,
  time = "TIME",
  group = NULL,
  min_time = 0,
  max_time = NULL,
  points = FALSE,
  lines = TRUE,
  admin = FALSE,
  cfb = FALSE,
  silent = TRUE,
  mean = FALSE,
  title = "",
  caption = "",
  ...
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte as character.}

\item{dose}{The dose as numeric.}

\item{log}{Logarithmic y scale}

\item{time}{The time field to use, as character. Can be 'TIME', 'NTIME',
'TAFD' or 'TAD'.}

\item{group}{The grouping variable as character.}

\item{min_time}{The minimal time as numeric.}

\item{max_time}{The maximal time as numeric.}

\item{points}{Show data points as logical.}

\item{lines}{Show lines as logical.}

\item{admin}{Show vertical lines for administrations, as logical.}

\item{cfb}{Show change from baseline, as logical.}

\item{silent}{Suppress messages as logical.}

\item{mean}{Show a mean plot, as logical.}

\item{title}{The plot title as character.}

\item{caption}{The plot caption line as character.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Title
}
\examples{
plot_z(examplinib_fe_nif)
plot_z(examplinib_fe_nif, nominal_time = TRUE, group = "FASTED")
plot_z(examplinib_sad_nif)
plot_z(examplinib_sad_nif, cfb = TRUE)
plot_z(examplinib_poc_nif, analyte="RS2023", admin = TRUE)
plot_z(examplinib_poc_nif)
plot_z(examplinib_poc_nif, analyte="RS2023", nominal_time = TRUE,
  group = "SEX", points = TRUE, lines = FALSE)
plot_z(examplinib_poc_nif, analyte="RS2023", tad = TRUE,
  dose = 500, log = FALSE, points = TRUE, lines = FALSE)
plot_z(examplinib_sad_min_nif)
plot_z(examplinib_poc_min_nif, dose = 500, cmt = 2)
plot_z(examplinib_poc_min_nif, dose = 500, cmt = 2, tad = TRUE,
  points = TRUE, lines = FALSE)
}
