% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plural}
\alias{plural}
\title{Return singular or plural form of word}
\usage{
plural(word, plural)
}
\arguments{
\item{word}{Source word in singular form, as character.}

\item{plural}{Return plural form, as character.}
}
\value{
Character.
}
\description{
Return singular or plural form of word
}
\examples{
plural("subject", F)
plural("subject", T)
plural("study", F)
plural("study", T)
}
\keyword{internal}
