% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pos_diff}
\alias{pos_diff}
\title{Difference, or zero, if difference is negative}
\usage{
pos_diff(a, b)
}
\arguments{
\item{a}{A as numeric.}

\item{b}{B as numeric.}
}
\value{
Numeric.
}
\description{
Difference, or zero, if difference is negative
}
\keyword{internal}
