% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{positive_or_zero}
\alias{positive_or_zero}
\title{Positive value or zero if negative}
\usage{
positive_or_zero(x)
}
\arguments{
\item{x}{Numeric.}
}
\value{
Numeric.
}
\description{
Positive value or zero if negative
}
\examples{
positive_or_zero(2)
positive_or_zero(-2)
positive_or_zero(c(2, 1, 0, -1, -2))
}
\keyword{internal}
