% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_dataset.R
\name{print.adam_dataset}
\alias{print.adam_dataset}
\title{Print method for ADaM dataset objects}
\usage{
\method{print}{adam_dataset}(x, ...)
}
\arguments{
\item{x}{A dataset object.}

\item{...}{Further arguments.}
}
\value{
Noting.
}
\description{
Print method for ADaM dataset objects
}
