% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{print.nif}
\alias{print.nif}
\title{print() implementation for nif objects}
\usage{
\method{print}{nif}(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\description{
print() implementation for nif objects
}
