% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{print}
\alias{print}
\title{print() implementation for nif objects}
\usage{
print(x, ...)
}
\arguments{
\item{x}{A nif object.}

\item{...}{Additional parameters}
}
\description{
print() implementation for nif objects
}
