% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{print.sdtm}
\alias{print.sdtm}
\title{print() implementation for SDTM}
\usage{
\method{print}{sdtm}(obj)
}
\arguments{
\item{obj, }{a sdtm class object}
}
\description{
print() implementation for SDTM
}
