% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{print.subject_info}
\alias{print.subject_info}
\title{Implementation of print for subject info}
\usage{
\method{print}{subject_info}(x, ...)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Optional further parameters.}
}
\description{
Implementation of print for subject info
}
