% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_analysis.R
\name{print.summary_nif}
\alias{print.summary_nif}
\title{Print NIF summary}
\usage{
\method{print}{summary_nif}(x, ...)
}
\arguments{
\item{x}{A summary_nif object.}

\item{...}{Further parameters.}
}
\value{
Nothing.
}
\description{
Print NIF summary
}
