% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{print.summary_sdtm}
\alias{print.summary_sdtm}
\title{print SDTM summary}
\usage{
\method{print}{summary_sdtm}(x, ...)
}
\arguments{
\item{x}{SDTM object}

\item{...}{Further parameters}
}
\value{
none
}
\description{
print SDTM summary
}
