% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_debug}
\alias{print_debug}
\title{Debug output}
\usage{
print_debug(obj)
}
\arguments{
\item{obj}{An object.}
}
\value{
Nothing
}
\description{
Print an object only if the global option 'debug' is \code{TRUE}.
}
\seealso{
\code{\link[=nif_option]{nif_option()}}
}
