% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pt_to_hours}
\alias{pt_to_hours}
\title{Convert ISO 8601-formatted duration to hours}
\usage{
pt_to_hours(iso)
}
\arguments{
\item{iso}{The duration as ISO 8601-formatted string.}
}
\value{
Duration in hours.
}
\description{
Convert ISO 8601-formatted duration to hours
}
\examples{
pt_to_hours(c("PT1H15M", "PT1.5H"))
}
