% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{raynaud_crea}
\alias{raynaud_crea}
\title{Serum creatinine estimation from eGFR (Raynaud method)}
\usage{
raynaud_crea(egfr, age, sex, race = "")
}
\arguments{
\item{egfr}{EGFR in ml/min/1.73 m^2.}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (male is 0) or character (male is "M").}

\item{race}{Race. Dummy variable for compatibility, race is not used by this
method.}
}
\value{
Serum creatinine in mg/dl.
}
\description{
Inverse of the function published in
\href{https://doi.org/10.1136/bmj-2022-073654}{Raynaud, BMJ 2023;381:e073654}
}
\details{
To convert crea from mg/dl to umol/l, multiply by 88.4.
}
