% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{raynaud_egfr}
\alias{raynaud_egfr}
\title{Glomerular filtration rate estimation from serum creatinine (Raynaud method)}
\usage{
raynaud_egfr(crea, age, sex, race = "")
}
\arguments{
\item{crea}{Serum creatinine in mg/dl.}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (male is 0) or character (male is "M").}

\item{race}{Race. Dummy variable for compatibility, race is not used by this
method.}
}
\value{
Estimated GFR in ml/min/1.73 m^2.
}
\description{
Reference:
\href{https://doi.org/10.1136/bmj-2022-073654}{Raynaud, BMJ 2023;381:e073654}
}
