% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_load.R
\name{read_adam}
\alias{read_adam}
\title{Read ADaM data}
\usage{
read_adam(data_path, dataset = NULL, format = "sas", delim = ",", ...)
}
\arguments{
\item{data_path}{The file system path to the source folder as character.}

\item{dataset}{The ADaM data sets to load. Defaults to all, if NULL.}

\item{format}{The format of the source files as character, either 'sas'
(default), 'xpt', or 'csv'.}

\item{delim}{delim Deliminator as character.}

\item{...}{Further parameters to \code{readr::read_csv}.}
}
\value{
An \code{adam} object.
}
\description{
Read ADaM data
}
