% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{read_nif}
\alias{read_nif}
\title{Load NIF file}
\usage{
read_nif(filename, format = "csv")
}
\arguments{
\item{filename}{The filename as character.}

\item{format}{The data format, as character. Can be only "csv".}
}
\value{
A nigitggf object.
}
\description{
Load NIF file
}
