% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{read_sdtm}
\alias{read_sdtm}
\title{Read SDTM data}
\usage{
read_sdtm(
  data_path,
  domain = NULL,
  format = "sas",
  delim = ",",
  silent = NULL,
  ...
)
}
\arguments{
\item{data_path}{The file system path to the source folder as character.}

\item{domain}{The domain name(s) as character, defaults to
all domains found in the folder.}

\item{format}{The format of the source files as character, either 'sas'
(default), 'xpt', or 'csv'.}

\item{delim}{Deliminator.}

\item{silent}{Suppress the progress bar. If \code{NULL}, uses the package
\code{silent} option.}

\item{...}{Further parameters, refer to readr::read_csv}
}
\value{
A \code{sdtm} object.
}
\description{
This function reads SDTM-formatted data as SAS or XPT files from a folder
location.
}
