% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recode_race}
\alias{recode_race}
\title{Recode RACE columns in nif object}
\usage{
recode_race(obj, coding_table = NULL)
}
\arguments{
\item{obj}{A nif object with RACE as character field.}

\item{coding_table}{A data frame with the columns RACE and RACEN. Uses
default coding, if NULL.}
}
\value{
A nif object with the original RACE replaced by the numerical race
code.
}
\description{
Recode RACE columns in nif object
}
\examples{
recode_race(examplinib_sad_nif)
}
