% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{reformat_date}
\alias{reformat_date}
\title{Reformat date to SDTM conventions}
\usage{
reformat_date(x)
}
\arguments{
\item{x}{The date in POSIX format}
}
\value{
Date-time formatted as xxxx-xx-xxTxx:xx:xx
}
\description{
Reformat date to SDTM conventions
}
