% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{rel_dose_intensity}
\alias{rel_dose_intensity}
\title{Title}
\usage{
rel_dose_intensity(nif, analyte = NULL)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte as character. If \code{NULL} (default), all analytes
are selected.}
}
\description{
reference: https://www.lexjansen.com/pharmasug/2024/ST/PharmaSUG-2024-ST-297.pdf
}
