% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{rename_by_formula}
\alias{rename_by_formula}
\title{Rename columns in data frame based on function terms}
\usage{
rename_by_formula(obj, f)
}
\arguments{
\item{obj}{a data frame.}

\item{f}{a function.}
}
\value{
a data frame.
}
\description{
Rename columns in data frame based on function terms
}
