% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{rich_sampling_sbs}
\alias{rich_sampling_sbs}
\title{Identify subjects with rich sampling}
\usage{
rich_sampling_sbs(obj, analyte = NA, max_time = NA, n = 4)
}
\arguments{
\item{obj}{The NIF dataset.}

\item{analyte}{The analyte. If the analyte is NA, the most likely will be
selected.}

\item{max_time}{The end of the target interval across which the number of
samples is determined. If NA, the full treatment interval is selected.}

\item{n}{The sample number cut-off.}
}
\value{
A list of IDs in numeric format.
}
\description{
Identify subjects with rich sampling
}
\examples{
rich_sampling_sbs(examplinib_poc_nif, n = 6)
}
