% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_mean}
\alias{safe_mean}
\title{Mean derivative that works with vectors containing NaN}
\usage{
safe_mean(x, ...)
}
\arguments{
\item{x}{The input as numeric.}
}
\value{
The mean as numeric.
}
\description{
Mean derivative that works with vectors containing NaN
}
\keyword{internal}
