% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_min}
\alias{safe_min}
\title{Min derivative that works with vectors containing NaN}
\usage{
safe_min(x)
}
\arguments{
\item{x}{The input as numeric.}
}
\value{
The mean as numeric.
}
\description{
Min derivative that works with vectors containing NaN
}
\keyword{internal}
