% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_sd}
\alias{safe_sd}
\title{SD derivative that works with vectors containing NaN}
\usage{
safe_sd(x)
}
\arguments{
\item{x}{The input as numeric.}
}
\value{
The mean as numeric.
}
\description{
SD derivative that works with vectors containing NaN
}
\keyword{internal}
