% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{sampling_summary}
\alias{sampling_summary}
\title{Sampling scheme overview}
\usage{
sampling_summary(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A data frame.
}
\description{
Sampling scheme overview
}
\examples{
sampling_summary(examplinib_poc_nif)
}
