% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{sd_event_table}
\alias{sd_event_table}
\title{Make RxODE event table for single dose administrations}
\usage{
sd_event_table(ex, dm, vs, sampling_scheme)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{dm}{The DM domain as data frame.}

\item{vs}{The VS domain as data frame.}

\item{sampling_scheme}{The sampling scheme domain as data frame.}
}
\value{
An event table object.
}
\description{
Make RxODE event table for single dose administrations
}
