% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{sdtm_domain}
\alias{sdtm_domain}
\title{Constructor for sdtm_domain objects}
\usage{
sdtm_domain(data, trial_title = "")
}
\arguments{
\item{data}{A data frame.}

\item{trial_title}{The trial title. Defaults to "".}
}
\value{
A sdtm_domain object.
}
\description{
Constructor for sdtm_domain objects
}
