% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_domain_model}
\alias{sdtm_domain_model}
\title{SDTM Domain Model variables}
\format{
A data frame with 1969 rows:
\describe{
\item{DOMAIN}{Domain name}
\item{VARNAM}{Variable name}
\item{VARLABEL}{Long-form variable name}
\item{DEFTYPE}{Variable type}
\item{LENGTH}{Length of variable}
\item{CORE}{Level of requiredness}
}
}
\usage{
sdtm_domain_model
}
\description{
An overview on the variables included in the SDTM specification.
}
\keyword{datasets}
