% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_analytics.R
\name{sdtm_missing_times}
\alias{sdtm_missing_times}
\title{Number of DTC entries with missing time information}
\usage{
sdtm_missing_times(sdtm, fields)
}
\arguments{
\item{sdtm}{A stm object.}

\item{fields}{The fields to test, as character.}
}
\value{
A data frame.
}
\description{
Number of DTC entries with missing time information
}
\examples{
sdtm_missing_times(examplinib_poc, c("EXSTDTC", "EXENDTC"))
}
