% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\docType{data}
\name{standard_nif_fields}
\alias{standard_nif_fields}
\title{#' Convert three-letter country name to ISO 3166 codes
#'
#' @param country_abbr The Country name in three-letter format according to
#' ISO 3166.
#'
#' @return The numerical country code according to ISO 3166.
#' @import DescTools
#' @export
country_code <- function(country_abbr) {
DescTools::d.countries \%>\%
filter(a3 \%in\% country_abbr) \%>\%
pull(code)
}}
\format{
An object of class \code{character} of length 29.
}
\usage{
standard_nif_fields
}
\value{
A character vector of the standard NIF fields
}
\description{
#' Convert ISO 3166 country code to three-letter country name.
#'
#' @param country_code The numerical country code according to ISO 3166.
#'
#' @return The three-letter country abbreviation.
#' @export
country_abbr <- function(country_code) {
DescTools::d.countries \%>\%
filter(code \%in\% country_code) \%>\%
pull(a3)
}
Standard nif fields
}
\keyword{datasets}
