% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_extension.R
\name{stat_admin}
\alias{stat_admin}
\title{ggplot stat for treatment administrations}
\usage{
stat_admin(
  mapping = NULL,
  data = NULL,
  geom = "vline",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  xintercept = NULL,
  ...
)
}
\arguments{
\item{mapping}{The mapping.}

\item{data}{The data as data frame.}

\item{geom}{The geom.}

\item{position}{The position.}

\item{na.rm}{Remove NA, as logical.}

\item{show.legend}{Show legend as logical.}

\item{inherit.aes}{Inherit aesthetics as logical.}

\item{xintercept}{The x intercept.}

\item{...}{Further parameters.}
}
\value{
A ggplot layer object.
}
\description{
ggplot stat for treatment administrations
}
