% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{studies.nif}
\alias{studies.nif}
\title{Studies within a nif object}
\usage{
\method{studies}{nif}(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A character vector of all STUDYIDs in the data set.
}
\description{
Studies within a nif object
}
