% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{subject_info.nif}
\alias{subject_info.nif}
\title{Subject information}
\usage{
\method{subject_info}{nif}(obj, id)
}
\arguments{
\item{obj}{A NIF object.}

\item{id}{The USUBJID.}
}
\description{
Subject information
}
