% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{subjects.nif}
\alias{subjects.nif}
\title{Unique subjects within a NIF object}
\usage{
\method{subjects}{nif}(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A data frame of all ID - USUBJID pairs in the data set.
}
\description{
Unique subjects within a NIF object
}
\examples{
subjects(examplinib_fe_nif)

}
