% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{subjects}
\alias{subjects}
\title{Unique subjects within a data set (generic method)}
\usage{
subjects(obj)
}
\arguments{
\item{obj}{The data set.}
}
\value{
A data frame.
}
\description{
Unique subjects within a data set (generic method)
}
