% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subjects.sdtm}
\alias{subjects.sdtm}
\title{Get the USUBJID from an sdtm object}
\usage{
\method{subjects}{sdtm}(obj)
}
\arguments{
\item{obj}{The sdtm object.}
}
\description{
Get the USUBJID from an sdtm object
}
