% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{suggest}
\alias{suggest}
\title{Suggest data programming steps for a sdtm object}
\usage{
suggest(obj)
}
\arguments{
\item{obj}{A sdtm object}
}
\description{
Suggest data programming steps for a sdtm object
}
\examples{
suggest(examplinib_poc)
}
