% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{suggest.sdtm}
\alias{suggest.sdtm}
\title{Suggest common manual data programming steps for a sdtm data set}
\usage{
\method{suggest}{sdtm}(obj)
}
\arguments{
\item{obj}{A sdtm object}
}
\description{
Suggest common manual data programming steps for a sdtm data set
}
\seealso{
\code{\link[=read_sdtm_sas]{read_sdtm_sas()}}
}
