% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_class.R
\name{summary.adam}
\alias{summary.adam}
\title{Summary method for adam objects}
\usage{
\method{summary}{adam}(object, ...)
}
\arguments{
\item{object}{An adam object.}

\item{...}{Further parameters.}
}
\value{
A summary_adam object.
}
\description{
Summary method for adam objects
}
