% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{summary}
\alias{summary}
\title{NIF or SDTM object overview}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{The NIF or SDTM object.}

\item{...}{Further arguments.}
}
\value{
A summary_nif or summary_sdtm object.
}
\description{
NIF or SDTM object overview
}
\examples{
summary(examplinib_poc)
summary(examplinib_poc_nif)
}
