% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{summary.sdtm_domain}
\alias{summary.sdtm_domain}
\title{Summary for domain object}
\usage{
\method{summary}{sdtm_domain}(object, ..., silent = NULL)
}
\arguments{
\item{object}{Domain object.}

\item{...}{Further parameters.}

\item{silent}{Suppress messages, defaults to nif_options, if NULL.}
}
\value{
A summary object
}
\description{
Summary for domain object
}
\keyword{internal}
