% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{synthesize_examplinib}
\alias{synthesize_examplinib}
\title{Synthesize package data}
\usage{
synthesize_examplinib()
}
\value{
None.
}
\description{
This function can be called during development to re-created the package
data. The data include SDTM and NIF objects for the following fictional
studies with the imaginary drug \code{examplinib}:
}
\details{
\itemize{
\item A single ascending dose (SAD) study
\item A single-arm multiple-dose proof-of-concept study
\item A cross-over food effect (FE) study
}

All SDTM data are entirely synthetic and have no relation whatsoever with
real clinical study data. This also applies to the pharmacokinetic data in
the SDTM/PC domain. These data are created using a population PK model that
is parametrized with arbitrary parameters.
}
