% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{synthesize_lb}
\alias{synthesize_lb}
\title{Synthesize baseline LB data}
\usage{
synthesize_lb(dm)
}
\arguments{
\item{dm}{The DM domain as data frame.}
}
\value{
The LB domain as data frame.
}
\description{
Synthesize baseline LB data
}
\keyword{internal}
