% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{synthesize_pp}
\alias{synthesize_pp}
\title{Synthesize PP domain}
\usage{
synthesize_pp(nif)
}
\arguments{
\item{nif}{A nif object.}
}
\value{
The PP domain as data frame.
}
\description{
Synthesize PP domain
}
\keyword{internal}
