% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{synthesize_sdtm}
\alias{synthesize_sdtm}
\title{Synthesize SDTM data for a fictional study}
\usage{
synthesize_sdtm()
}
\value{
The SDTM data as sdtm object.
}
\description{
Synthesize SDTM data for a fictional study
}
