% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{synthesize_sdtm_sad_study}
\alias{synthesize_sdtm_sad_study}
\title{Synthesize SDTM data for a 3 + 3 dose escalation study}
\usage{
synthesize_sdtm_sad_study()
}
\value{
A sdtm object.
}
\description{
Synthesize SDTM data for a 3 + 3 dose escalation study
}
