% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{time_by_ntime}
\alias{time_by_ntime}
\title{Plot TIME vs NTIME}
\usage{
time_by_ntime(obj, max_time = NULL, ...)
}
\arguments{
\item{obj}{A nif object.}

\item{max_time}{The maximum time as numeric.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Plot TIME vs NTIME
}
\keyword{internal}
