% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_check.R
\name{time_plot}
\alias{time_plot}
\title{Plot time metrics against each other for observations}
\usage{
time_plot(
  obj,
  xtime = "TIME",
  ytime = "TAD",
  analyte = NULL,
  max_time = NULL,
  color = "CHECK",
  ...
)
}
\arguments{
\item{obj}{A nif object.}

\item{max_time}{The maximum time as numeric.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Plot time metrics against each other for observations
}
\keyword{internal}
