% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{treatments.nif}
\alias{treatments.nif}
\title{Treatments in a nif object}
\usage{
\method{treatments}{nif}(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
Character.
}
\description{
Treatments in a nif object
}
\examples{
treatments(examplinib_poc_nif)
}
