% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{treatments}
\alias{treatments}
\title{Treatments in a nif or sdtm object}
\usage{
treatments(obj)
}
\arguments{
\item{obj}{A nif or sdtm object.}
}
\value{
Character.
}
\description{
Treatments in a nif or sdtm object
}
