% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{treatments.sdtm}
\alias{treatments.sdtm}
\title{Treatments in a sdtm object}
\usage{
\method{treatments}{sdtm}(obj)
}
\arguments{
\item{obj}{A sdtm object.}
}
\value{
Character.
}
\description{
Treatments in a sdtm object
}
\examples{
treatments(examplinib_poc)
}
\keyword{internal}
