% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{trial_dco}
\alias{trial_dco}
\title{Trial data cut-off date as reported in TS domain}
\usage{
trial_dco(obj)
}
\arguments{
\item{obj}{A sdtm object.}
}
\value{
A POSIXct object or NULL.
}
\description{
Trial data cut-off date as reported in TS domain
}
