% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{trial_title}
\alias{trial_title}
\title{Extract trial title from SDTM data, if possible.}
\usage{
trial_title(obj)
}
\arguments{
\item{obj}{A sdtm object.}
}
\value{
The title or NULL.
}
\description{
Extract trial title from SDTM data, if possible.
}
\examples{
trial_title(examplinib_sad)
}
