% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trialday_to_day}
\alias{trialday_to_day}
\title{Convert trial day to elapsed days}
\usage{
trialday_to_day(x)
}
\arguments{
\item{trial_day}{The trial day as numeric.}
}
\value{
The number of elapsed days as numeric.
}
\description{
This function corrects for trial day 1 actually indicating zero elapsed days
since the treatment start.
}
