% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{usubjid}
\alias{usubjid}
\title{Get the USUBJID of subject}
\usage{
usubjid(obj, id)
}
\arguments{
\item{obj}{A NIF object.}

\item{id}{A subject ID in numeric form.}
}
\value{
The USUBJID as character.
}
\description{
Get the USUBJID of subject
}
\examples{
usubjid(examplinib_fe_nif)
}
