% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_analyte}
\alias{validate_analyte}
\title{Ensure that analytes are present in nif object}
\usage{
validate_analyte(
  nif,
  analyte,
  allow_multiple = TRUE,
  allow_null = FALSE,
  allow_empty = FALSE
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{Required analyte(s) as character.}

\item{allow_multiple}{Allow vector of the specified type, as logical.}

\item{allow_null}{Allow NULL values, as logical.}

\item{allow_empty}{Allow empty parameter, as logical.}
}
\value{
Nothing or stop.
}
\description{
Ensure that analytes are present in nif object
}
