% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{validate_domain}
\alias{validate_domain}
\title{Check whether fields in domain are compliant with SDTM}
\usage{
validate_domain(domain, silent = NULL)
}
\arguments{
\item{domain}{The SDTM domain as data frame.}

\item{silent}{Suppress optional messages, as logical. Defaults to global
nif_options if NULL.}
}
\value{
nothing
}
\description{
Check whether fields in domain are compliant with SDTM
}
\examples{
validate_domain(examplinib_sad$dm)
}
