% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_domain_param}
\alias{validate_domain_param}
\title{Validate domain object parameter}
\usage{
validate_domain_param(obj)
}
\arguments{
\item{obj}{A domain object.}
}
\value{
Nothing or stop.
}
\description{
Validate domain object parameter
}
