% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{validate_lbw_parameters}
\alias{validate_lbw_parameters}
\title{Validate inputs to lbm functions}
\usage{
validate_lbw_parameters(weight, height, sex)
}
\arguments{
\item{weight}{weight.}

\item{height}{height.}

\item{sex}{sex.}
}
\value{
vector of valid inputs
}
\description{
Validate inputs to lbm functions
}
