% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_logical_param}
\alias{validate_logical_param}
\title{Validate logical parameter}
\usage{
validate_logical_param(
  param,
  param_name,
  allow_null = FALSE,
  allow_empty = FALSE,
  allow_multiple = FALSE
)
}
\arguments{
\item{param}{The parameter to be tested.}

\item{param_name}{The parameter name as character.}

\item{allow_null}{Allow NULL values, as logical.}

\item{allow_empty}{Allow empty parameter, as logical.}

\item{allow_multiple}{Allow vector of the specified type, as logical.}
}
\value{
Nothing or stop.
}
\description{
Validate logical parameter
}
