% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_min_nif}
\alias{validate_min_nif}
\title{Validate nif object with minimally required fields}
\usage{
validate_min_nif(obj, additional_fields = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{additional_fields}{Additional required fields as character.}
}
\value{
Nothing or stop.
}
\description{
Validate nif object with minimally required fields
}
