% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_nif}
\alias{validate_nif}
\title{Validate nif object parameter}
\usage{
validate_nif(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
Nothing or stop.
}
\description{
Validate nif object parameter
}
