% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_param}
\alias{validate_param}
\title{Generic function parameter validation}
\usage{
validate_param(
  type = c("string", "logical", "numeric"),
  param,
  param_name,
  allow_null = FALSE,
  allow_empty = FALSE,
  allow_multiple = FALSE,
  allow_na = FALSE
)
}
\arguments{
\item{type}{Parameter type (string, logical or numeric)}

\item{param}{The parameter to be tested.}

\item{param_name}{The parameter name as character.}

\item{allow_null}{Allow NULL values, as logical.}

\item{allow_empty}{Allow empty parameter, as logical.}

\item{allow_multiple}{Allow vector of the specified type, as logical.}

\item{allow_na}{Allow NA value, as logical.}
}
\value{
Nothing or stop.
}
\description{
Generic function parameter validation
}
