% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{validate_sdtm}
\alias{validate_sdtm}
\title{Validate sdtm object}
\usage{
validate_sdtm(obj)
}
\arguments{
\item{obj}{sdtm object.}
}
\value{
The unchanged sdtm object.
}
\description{
Validate sdtm object
}
