% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_testcd}
\alias{validate_testcd}
\title{Validate that one or multiple testcd are included in a sdtm object}
\usage{
validate_testcd(sdtm, testcd, domain = NULL)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{testcd}{Testcode to validate as character.}

\item{domain}{Domain as character}
}
\value{
Validated testcode(s) as character.
}
\description{
Validate that one or multiple testcd are included in a sdtm object
}
