% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_extension.R
\name{watermark}
\alias{watermark}
\title{Add a watermark annotation layer for a ggplot2 object}
\usage{
watermark(watermark_text = NULL, cex = 1.5, fontface = "bold")
}
\arguments{
\item{watermark_text}{The watermark as character.}

\item{cex}{The cex as numeric.}

\item{fontface}{Font face ("bold" by default)}
}
\value{
A grob.
}
\description{
Add a watermark annotation layer for a ggplot2 object
}
