% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{write_csv}
\alias{write_csv}
\title{Export a nif object as csv file}
\usage{
write_csv(obj, filename)
}
\arguments{
\item{obj}{A nif object.}

\item{filename}{The filename for the exported file.}
}
\description{
Export a nif object as csv file
}
