% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_write.R
\name{write_monolix}
\alias{write_monolix}
\title{Write as comma-separated file, complying with the format used by Monolix}
\usage{
write_monolix(obj, filename = NULL, fields = NULL)
}
\arguments{
\item{obj}{The NIF object.}

\item{filename}{The filename as string. If not filename is specified, the
file is printed only.}

\item{fields}{The fields to export. If NULL (default), all fields will be
exported.}
}
\value{
Nothing.
}
\description{
Write as comma-separated file, complying with the format used by Monolix
}
\examples{
head(write_monolix(examplinib_fe_nif))
}
