% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_write.R
\name{write_nif}
\alias{write_nif}
\title{Write NONMEM input formatted nif object to file}
\usage{
write_nif(obj, filename = NULL, fields = NULL, sep = NULL)
}
\arguments{
\item{obj}{The NIF object.}

\item{filename}{The filename as string. If not filename is specified, the
file is printed only.}

\item{fields}{The fields to export. If NULL (default), all fields will be
exported.}

\item{sep}{The separating character, e.g. ',' or ';'. If NULL (default), the
output has a space-separated, fixed-width format.}
}
\value{
Nothing.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
