% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_write.R
\name{write_nonmem}
\alias{write_nonmem}
\title{Write NONMEM input formatted nif object to file}
\usage{
write_nonmem(
  obj,
  filename = NULL,
  fields = NULL,
  sep = NULL,
  numeric_fields_only = FALSE,
  dot_columns = NULL
)
}
\arguments{
\item{obj}{The NIF object.}

\item{filename}{The filename as string. If not filename is specified, the
file is printed only.}

\item{fields}{The fields to export in addition to the minimally required
fields. If NULL (default), all fields will be exported.}

\item{sep}{The separating character, e.g. ',' or ';'. If NULL (default), the
output has a space-separated, fixed-width format.}

\item{numeric_fields_only}{Include only numerical fields.}

\item{dot_columns}{Fields in which zeros will be replaced with ".".}
}
\value{
Nothing.
}
\description{
Write as space-delimited, fixed-width or a character-separated file. If the
'numeric_fields_only' argument is TRUE, only columns that are numeric will be
exported.
}
\details{
All numeric fields are reduced to 4 decimal places. For this, IEC 60559
is applied, i.e., rounding for a last digit of 5 is to the next even number.
All fields are converted to character, and NA-values are converted to '.'.

Further, in the columns specified by the 'dot_columns' argument, zero values
are also converted to '.'.
}
\examples{
write_nonmem(examplinib_fe_nif)
write_nonmem(examplinib_fe_nif, numeric_fields_only = TRUE)
write_nonmem(examplinib_fe_nif, dot_columns = c("DV", "AMT"))
}
