% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{wt_by_ht}
\alias{wt_by_ht}
\title{Weight by height scatterplot}
\usage{
wt_by_ht(obj, alpha = 0.7)
}
\arguments{
\item{obj}{The NIF object.}

\item{alpha}{The alpha for the point rendering.}
}
\value{
A plot object.
}
\description{
Weight by height scatterplot
}
\examples{
wt_by_ht(examplinib_poc_nif)
}
