% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{wt_by_race}
\alias{wt_by_race}
\title{Weight by race diagram}
\usage{
wt_by_race(obj)
}
\arguments{
\item{obj}{The NIF file object.}
}
\value{
A plot object.
}
\description{
Weight by race diagram
}
\examples{
wt_by_race(examplinib_poc_nif)
}
\keyword{internal}
