% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{wt_by_sex}
\alias{wt_by_sex}
\title{Weight by sex diagram}
\usage{
wt_by_sex(obj)
}
\arguments{
\item{obj}{The NIF file object.}
}
\value{
A ggplot object.
}
\description{
Weight by sex diagram
}
\examples{
wt_by_sex(examplinib_poc_nif)
}
\keyword{internal}
