% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_class.R
\name{adam}
\alias{adam}
\title{adam object class constructor}
\usage{
adam(adam_data)
}
\arguments{
\item{adam_data}{The ADaM datasets as data frames.}
}
\value{
An adam object.
}
\description{
adam object class constructor
}
