% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_administrations.R
\name{add_administration}
\alias{add_administration}
\title{Add administration events}
\usage{
add_administration(
  nif,
  sdtm,
  extrt,
  analyte = NULL,
  pctestcd = NULL,
  cmt = 1,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  cut_off_date = NULL,
  keep = NULL,
  debug = FALSE,
  imputation = NULL,
  duration = NULL,
  iv_admin = NULL,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{extrt}{The EXTRT for the administration, as character.}

\item{analyte}{The name of the analyte as character.}

\item{pctestcd}{The PCTESTCD of the pharmacokinetic analyte corresponding to
the administered drug. This is needed when administration times are imputed
from the PCRFTDTC field from the PC domain.}

\item{cmt}{The compartment for the administration as numeric.}

\item{subject_filter}{The filtering to apply to the DM domain, as string,}

\item{cut_off_date}{The data cut-off date as Posix date-time or character.}

\item{keep}{Columns to keep after cleanup, as character.}

\item{debug}{Include debug fields, as logical.}

\item{imputation}{The imputation rule set as a list, or \code{NULL} to use
\code{attr(nif, "imputation_rules")}.}

\item{duration}{The duration of the iv administration as numeric. If NULL,
will be determined automatically based on the EXDUR field, if available.}

\item{iv_admin}{Administration is iv, as boolean. If NULL, will be tested
automatically based on the EXROUTE field, if available.}

\item{silent}{Suppress messages, defaults to nif_option standard, if NULL.}
}
\value{
A nif object.
}
\description{
Add rows to a \link{nif} object that represent drug administration events
(EVID of 1) This is usually the first step in the stepwise creation of NIF
data tables.
}
\details{
Drug administration data is taken from the EX domain of the source
\link{sdtm} object. The \code{extrt} argument specifies the drug name as
represented in EX. By default, this will also be the value of the 'ANALYTE'
column for the respective rows in the resulting \link{nif} object.
Alternatively, a custom \code{analyte} name can be explicitly provided, e.g., to
match with the 'ANALYTE' name of the corresponding pharmacokinetic
observations.

For administrations, a model compartment of 1 is selected by default and will
be the corresponding value of the 'CMT' column. A different compartment can
be explicitly specified by the \code{cmt} argument.

For an overview on the representation of administration events in NONMEM
Input Format compliant data sets, see: Bauer, R.J. (2019), NONMEM Tutorial
Part I: Description of Commands and Options, With Simple Examples of
Population Analysis. CPT Pharmacometrics Syst. Pharmacol., 8: 525-537
\doi{10.1002/psp4.12404}.

To add observation events to the \link{nif} object, see
\code{\link[=add_observation]{add_observation()}}.
}
\examples{
add_administration(nif(), examplinib_sad, "EXAMPLINIB")

}
