% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_administration}
\alias{add_administration}
\title{Add adminisration}
\usage{
add_administration(nif, ...)
}
\arguments{
\item{nif}{A nif object.}

\item{...}{
  Arguments passed on to \code{\link[=make_administration]{make_administration}}
  \describe{
    \item{\code{sdtm}}{A sdtm object.}
    \item{\code{subject_filter}}{The filtering to apply to the DM domain.}
    \item{\code{cleanup}}{}
  }}
}
\value{
A nif object.
}
\description{
Add adminisration
}
