% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_administration}
\alias{add_administration}
\title{Append drug administration events}
\usage{
add_administration(
  nif,
  sdtm,
  extrt,
  analyte = NA,
  cmt = 1,
  cut_off_date = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  silent = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{extrt}{The EXTRT for the administration, as character.}

\item{analyte}{The name of the analyte as character.}

\item{cmt}{The compartment for the administration as numeric.}

\item{cut_off_date}{The data cut-off date as Posix date-time.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{silent}{Suppress messages, as logical.}

\item{cleanup}{Remove unnecessary fields, as logical.}
}
\value{
A nif object.
}
\description{
Drug administration information is taken from the EX domain of the sdtm
object. The 'extrt' field specifies the drug name as represented in 'EX',
however, a different 'analyte' name can be assigned to match with
pharmacokinetic observations for the parent drug in plasma.
}
\examples{
add_administration(new_nif(), examplinib_sad, "EXAMPLINIB")

}
