% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_administrations.R
\name{add_administration}
\alias{add_administration}
\title{Append administration events}
\usage{
add_administration(
  nif,
  sdtm,
  extrt,
  analyte = NULL,
  cmt = 1,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  cut_off_date = NULL,
  keep = NULL,
  debug = FALSE,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{extrt}{The EXTRT for the administration, as character.}

\item{analyte}{The name of the analyte as character.}

\item{cmt}{The compartment for the administration as numeric.}

\item{subject_filter}{The filtering to apply to the DM domain, as string,}

\item{cut_off_date}{The data cut-off date as Posix date-time.}

\item{keep}{Columns to keep after cleanup, as character.}

\item{debug}{Include debug fields, as logical.}

\item{silent}{Suppress messages, defaults to nif_option standard when NULL.}
}
\value{
A nif object.
}
\description{
Drug administration data is taken from the EX domain of the sdtm object. The
'extrt' field specifies the drug name as represented in 'EX', however, a
different 'analyte' name can be assigned to match with that of the
pharmacokinetic observations for the parent drug in plasma.
}
\examples{
add_administration(new_nif(), examplinib_sad, "EXAMPLINIB")

}
