% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ae.R
\name{add_ae_observation}
\alias{add_ae_observation}
\title{Add AE observation}
\usage{
add_ae_observation(
  nif,
  sdtm,
  ae_term,
  ae_field = "AEDECOD",
  parent = NULL,
  cmt = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  coding_table = NULL,
  keep = NULL,
  debug = FALSE,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{ae_term}{The AE term as character.}

\item{ae_field}{The field in which the AE term is specified in. Can be:
\itemize{
\item AEDECOD:  Dictionary-derived term
\item AELLT:    Lowest-level term
\item AEHLT:    High-level term
\item AEHLGT:   High-level group term
\item AESOC:    System organ class
\item AEBODSYS: Body system or organ class
}

or any other field from the 'AE' domain}

\item{parent}{The parent compound as character.}

\item{cmt}{The compartment as numeric.}

\item{subject_filter}{A subject filter term.}

\item{observation_filter}{An observation filter term.}

\item{coding_table}{A DV coding table as data frame. The coding table translates
columns included in the AE domain into a DV field.}

\item{keep}{Columns to keep, as character.}

\item{debug}{Include debug fields, as logical.}

\item{silent}{Suppress messages, defaults to nif_option setting, if NULL.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function adds AE events as observations with the time of the AE onset
as the observation time point and the severity grade as the dependent
variable if the field AETOXGR is included. If AETOXGR is not included in the
AE domain, a coding table must be provided to translate any other field into
a numeric DV column, e.g.,
\verb{coding_table = data.frame( AESEV = c("MILD", "MODERATE", "SEVERE"), DV = c(1, 2, 3)}
}
