% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_mappings.R
\name{add_analyte_mapping}
\alias{add_analyte_mapping}
\title{Attach a treatment-analyte mapping to SDTM object}
\usage{
add_analyte_mapping(obj, extrt, pctestcd, analyte = NULL)
}
\arguments{
\item{obj}{A SDTM object.}

\item{extrt}{The treatment as defined in EX.}

\item{pctestcd}{The analyte as character (as in 'PCTESTCD' for PK observations}

\item{analyte}{The analyte name to be used in the nif object, as character.}
}
\value{
A SDTM object.
}
\description{
In some studies, multiple drugs are co-administered, and there may be analyte
data related to different parent drugs. In order to appropriately correlate
observations with administrations, the \code{\link[=nif_auto]{nif_auto()}} function needs to know
which analyte (\code{PCTESTCD}) belongs to which drug (\code{EXTRT}). Multiple mappings
can be provided.
}
\keyword{internal}
