% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_baseline}
\alias{add_baseline}
\title{Title}
\usage{
add_baseline(
  nif,
  sdtm,
  domain,
  testcd,
  DV_field = NULL,
  TESTCD_field = NULL,
  observation_filter = {
     TRUE
 },
  subject_filter = {
     TRUE
 },
  baseline_filter = NULL,
  summary_function = mean,
  silent = FALSE
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{domain}{The domain as character.}

\item{testcd}{The xxTESTCD to filter for, as character.}

\item{DV_field}{the field to use as the dependent variable. Defaults to the
two-character domain name followed by 'STRESN'.}

\item{TESTCD_field}{The xxTESTCD field as character. Defaults to "xxTESTCD"
where xx is the domain name.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{baseline_filter}{The filter to identify the baseline value. Defaults to
{xxBLFL == 'Y'} where xx is the domain name.}

\item{summary_function}{The summary funcition to apply if there are multiple
baseline values, defaults to \code{mean}.}

\item{silent}{Suppress messages as logical.}
}
\value{
A nif object.
}
\description{
Title
}
