% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_baseline}
\alias{add_baseline}
\title{Attach baseline covariate column}
\usage{
add_baseline(
  nif,
  sdtm,
  domain,
  testcd,
  DV_field = NULL,
  TESTCD_field = NULL,
  observation_filter = "TRUE",
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  baseline_filter = NULL,
  summary_function = mean,
  silent = FALSE
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{domain}{The domain as character.}

\item{testcd}{The covariate variable name as character.}

\item{DV_field}{The name of the DV field as character.}

\item{TESTCD_field}{The name of the TESTCD field. defaults to xxTESTCD with
xx the domain name, as character.}

\item{observation_filter}{A filter term for the \code{domain}, as character.}

\item{subject_filter}{A filter term for the DM domain, as character.}

\item{baseline_filter}{A filter term to identify the baseline condition
within the \code{domain}. Defaults to "xxBLFL == 'Y'" (with xx the domain
code).}

\item{summary_function}{The summary function to summarize multiple baseline
values. Defaults to \code{mean}.}

\item{silent}{Suppress messages, as logical.}
}
\value{
A nif object.
}
\description{
Baseline covariates, specified by the 'testcd' field, can come from any SDTM
domain. By default, the baseline value is identified by xxBLFL having a value
of "Y" in the respective SDTM domain. Alternatively, a custom observation
filter can be defined. The name of the baseline covariate is 'testcd',
prefixed with 'BL_'.
}
\examples{
add_baseline(examplinib_sad_nif, examplinib_sad, "vs", "WEIGHT")
add_baseline(examplinib_sad_nif, examplinib_sad, "vs", "WEIGHT",
  baseline_filter = "VSBLFL == 'Y'")
}
