% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_baseline}
\alias{add_baseline}
\title{Add a baseline covariate}
\usage{
add_baseline(
  nif,
  sdtm,
  domain,
  testcd,
  name = NULL,
  dv_field = NULL,
  testcd_field = NULL,
  observation_filter = "TRUE",
  cat = NULL,
  scat = NULL,
  baseline_filter = NULL,
  coding_table = NULL,
  summary_function = mean,
  factor = 1,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{domain}{The domain as character.}

\item{testcd}{The covariate variable name as character.}

\item{name}{The column label, as character.}

\item{dv_field}{The name of the DV field as character.}

\item{testcd_field}{The name of the TESTCD field. defaults to xxTESTCD with
xx the domain name, as character.}

\item{observation_filter}{A filter term for the \code{domain}, as character. Note:
if the filter term includes date comparisons, make sure to represent the
date as, datetime object e.g.,  \code{lubridate::as_datetime()}.}

\item{cat}{xxCAT filter to apply, as character.}

\item{scat}{xxSCAT filter to apply, as character.}

\item{baseline_filter}{A filter term to identify the baseline condition.
within the \code{domain}. Defaults to either "xxBLFL == 'Y'" or
"xxLOBXFL == 'Y'" (with xx the domain code), whichever is found first in
the domain data.}

\item{coding_table}{A recoding table as data frame, or NULL. If present, the
table needs to have a field that matches a column in the domain, and a
field 'DV' that provides the re-coded value.}

\item{summary_function}{The summary function to summarize multiple baseline
values. Defaults to \code{mean}.}

\item{factor}{A multiplier for the baseline value, defaults to 1.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A nif object.
}
\description{
Add a column to a \link{nif} object that represents the baseline value for a
subject-level covariate.
}
\details{
The source of the baseline covariate is specified by the \code{domain} and
\code{testcd} arguments. The baseline condition is defined by the
\code{baseline_filter} argument. If none is provided, the baseline filter defaults
to \code{xxBLFL == "Y"} where 'xx' is the domain code. In addition, a custom
\code{observation_filter} can be defined to further specify the observation. This
may be necessary, when observations defined by the \code{testcd} alone are
ambiguous, e.g., when for pharmacokinetic baseline observations, both BLOOD
and URINE observations are included in the PC domain data.

The name of the baseline column defaults to the 'testcd', prefixed with
'BL_', e.g., BL_WEIGHT. A specific name can be defined by the \code{name}
argument. Note that baseline WEIGHT, HEIGHT and BMI (if applicable) are
automatically included during the generation of a nif object.
}
\examples{
add_baseline(examplinib_sad_nif, examplinib_sad, "vs", "WEIGHT")
add_baseline(examplinib_sad_nif, examplinib_sad, "vs", "WEIGHT",
  baseline_filter = "VSBLFL == 'Y'"
)
}
