% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_covariates.R
\name{add_baseline}
\alias{add_baseline}
\title{Attach baseline covariate}
\usage{
add_baseline(
  nif,
  sdtm,
  domain,
  testcd,
  name = NULL,
  DV_field = NULL,
  TESTCD_field = NULL,
  observation_filter = "TRUE",
  baseline_filter = NULL,
  coding_table = NULL,
  summary_function = mean,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{domain}{The domain as character.}

\item{testcd}{The covariate variable name as character.}

\item{name}{The column label, as character.}

\item{DV_field}{The name of the DV field as character.}

\item{TESTCD_field}{The name of the TESTCD field. defaults to xxTESTCD with
xx the domain name, as character.}

\item{observation_filter}{A filter term for the \code{domain}, as character.}

\item{baseline_filter}{A filter term to identify the baseline condition.
within the \code{domain}. Defaults to either "xxBLFL == 'Y'" or
"xxLOBXFL == 'Y'" (with xx the domain code), whichever is found first in
the domain data.}

\item{coding_table}{A recoding table as data frame, or NULL. If present, the
table needs to have a field that matches a column in the domain, and a
field 'DV' that provides the re-coded value.}

\item{summary_function}{The summary function to summarize multiple baseline
values. Defaults to \code{mean}.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A nif object.
}
\description{
Baseline covariates, as specified by the 'testcd' field, can come from any
SDTM domain. By default, the baseline value is identified by \code{xxBLFL == "Y"}
in the respective SDTM domain. Alternatively, a custom observation filter can
be defined. The name of the baseline covariate is the 'testcd', prefixed with
'BL_'.
}
\examples{
add_baseline(examplinib_sad_nif, examplinib_sad, "vs", "WEIGHT")
add_baseline(examplinib_sad_nif, examplinib_sad, "vs", "WEIGHT",
  baseline_filter = "VSBLFL == 'Y'")
}
