% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_qtime.R
\name{add_bintime}
\alias{add_bintime}
\title{Make BINTIME field}
\usage{
add_bintime(obj, method = "fisher", time = "TAFD")
}
\arguments{
\item{obj}{A nif object.}

\item{method}{Univariate class intervals method, can be one of jenks, kmeans,
pretty, quantile, hclust, sd, bclust or fisher. See classInt::classInterval()
for details. Default is fisher.}

\item{time}{The time field to use.}
}
\value{
A nif object with the BINTIME, BIN_LEFT and BIN_RIGHT fields added.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
