% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_bl_crcl}
\alias{add_bl_crcl}
\title{Add baseline creatinine clearance field.}
\usage{
add_bl_crcl(obj, method = egfr_cg, molar = FALSE)
}
\arguments{
\item{obj}{A NIF object.}

\item{method}{The function to calculate eGFR (CrCL) from serum creatinine.}

\item{molar}{Convert to molar units, as logical.
Currently either: egfr_mdrd, egfr_cg or egfr_raynaud}
}
\value{
A NIF object with the baseline creatinine clearance (BL_EGFR) field
added,
}
\description{
Add baseline creatinine clearance field.
}
\examples{
head(add_bl_crcl(examplinib_poc_nif))
}
\seealso{
\code{\link[=egfr_mdrd]{egfr_mdrd()}}

\code{\link[=egfr_cg]{egfr_cg()}}

\code{\link[=egfr_raynaud]{egfr_raynaud()}}
}
