% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_bl_creat}
\alias{add_bl_creat}
\title{Add baseline creatinine}
\usage{
add_bl_creat(
  obj,
  sdtm,
  baseline_filter = NULL,
  observation_filter = "TRUE",
  silent = NULL
)
}
\arguments{
\item{obj}{A nif data set.}

\item{sdtm}{Source sdtm data object.}

\item{baseline_filter}{The filter term to identify baseline conditions.}

\item{silent}{Suppress messages.}
}
\value{
A nif object with the BL_CREAT column added, if possible. Otherwise
the unchanged input.
}
\description{
Add baseline creatinine
}
