% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif_deprecated.R
\name{add_bl_lab}
\alias{add_bl_lab}
\title{Add baseline lab covariate}
\usage{
add_bl_lab(obj, lb, lbtestcd, lbspec = NULL, silent = FALSE)
}
\arguments{
\item{obj}{A NIF dataset.}

\item{lb}{SDTM LB domain as data frame.}

\item{lbtestcd}{Lab parameter(s) as encoded by LBTESTCD.}

\item{lbspec}{The specimen, e.g., "BLOOD", "SERUM" or "URINE".}

\item{silent}{Disable messages.}
}
\value{
A NIF dataset.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Lab parameters not found in LB will be reported in a warning message.
}
