% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_bl_lab}
\alias{add_bl_lab}
\title{Add baseline covariates to a NIF object}
\usage{
add_bl_lab(obj, lb, lbtestcd, lbspec = "", silent = FALSE)
}
\arguments{
\item{obj}{NIF dataset.}

\item{lb}{SDTM LB domain as data frame.}

\item{lbtestcd}{Lab parameters as encoded by LBTESTCD, as strings.}

\item{lbspec}{The specimen, usually "BLOOD" or "URINE".}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
A NIF dataset
}
\description{
Lab parameters not found in LB will be reported in a warning message.
}
