% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_bl_odwg}
\alias{add_bl_odwg}
\title{Add baseline hepatic function class}
\usage{
add_bl_odwg(
  obj,
  sdtm,
  observation_filter = NULL,
  baseline_filter = NULL,
  summary_function = mean,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{sdtm}{The corresponding sdtm object.}

\item{observation_filter}{The filter term for the observation source data, as
character.}

\item{baseline_filter}{A filter term to identify the baseline condition, as
character.}

\item{summary_function}{The summary function to summarize multiple baseline
values. Defaults to \code{mean}}

\item{silent}{Suppress messages.}
}
\value{
A nif object.
}
\description{
Based on the NCI ODWG criteria with TB the total (direct and indirect) serum
bilirubin, and AST aspartate aminotransferase.
}
\details{
\itemize{
\item normal: TB & AST <= upper limit of normal (ULN)
\item mild hepatic dysfunction: TB > ULN to 1.5 x ULN or AST > ULN
\item moderate hepatic dysfunction: TB >1.5–3 x ULN, any AST
\item severe hepatic dysfunction: TB >3 - 10 x ULN, any AST
}
}
