% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_bl_renal}
\alias{add_bl_renal}
\title{Add baseline renal function class}
\usage{
add_bl_renal(obj, method = egfr_cg)
}
\arguments{
\item{obj}{A NIF object.}

\item{method}{The function to calculate eGFR (CrCL) from serum creatinine.
Currently either: egfr_mdrd, egfr_cg or egfr_raynaud}
}
\value{
A NIF object.
}
\description{
Add baseline renal function class
}
\examples{
as.data.frame(add_bl_renal(examplinib_poc_nif))
}
