% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_bl_renal}
\alias{add_bl_renal}
\title{Add baseline renal function class}
\usage{
add_bl_renal(obj, method = egfr_cg, molar = FALSE)
}
\arguments{
\item{obj}{A NIF object.}

\item{method}{The function to calculate eGFR (CrCL) from serum creatinine.}

\item{molar}{Use molar concentrations.
Currently either: egfr_mdrd, egfr_cg or egfr_raynaud}
}
\value{
A NIF object.
}
\description{
If baseline creatinine clearance (BL_CRCL) is not included in the input, it
will be calculated first.
}
\examples{
head(add_bl_renal(examplinib_poc_nif), 5)
}
