% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_baseline.R
\name{add_cfb}
\alias{add_cfb}
\title{Add baseline and change from baseline fields}
\usage{
add_cfb(
  obj,
  baseline_filter = "TIME <= 0",
  summary_function = median,
  silent = NULL
)
}
\arguments{
\item{obj}{A NIF object.}

\item{baseline_filter}{A filter term to identify the baseline condition.}

\item{summary_function}{The function to derive the baseline. This function is
applied over the DV values identified by the 'baseline_filter' term. The
default function is \code{median}. Alternatively, \code{mean}, \code{min} or \code{max} can be
considered.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A NIF object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Pleas use \code{\link[=derive_cfb]{derive_cfb()}} instead.
}
\details{
Output fields:
\itemize{
\item \code{DVBL} Baseline value for the dependent variable DV.
\item \code{DVCFB} Change from baseline for the dependent variable DV.
}

The Baseline is calculated as the median (or the summary function
output) of the DV field for all time points identified by the
baseline_filter' term.
}
