% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_cfb}
\alias{add_cfb}
\title{Add baseline and change from baseline fields}
\usage{
add_cfb(obj, summary_function = median)
}
\arguments{
\item{obj}{A NIF object.}

\item{summary_function}{The function to derive the baseline. This function is
applied over the DV values at TIME less than or equal to zero. The default is
\code{median}. Alternatively, \code{mean}, \code{min} or \code{max} can be considered.}
}
\value{
A NIF object
}
\description{
Add baseline and change from baseline fields
}
\details{
Output fields:
\itemize{
\item \code{DVBL} Baseline value for the dependent variable DV.
\item \code{DVCFB} Change from baseline for the dependent variable DV.
}

The Baseline is calculated as the median of the DV for all times lower or
equal to zero.
}
\examples{
head(add_cfb(examplinib_poc_nif))

}
