% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_covariates.R
\name{add_covariate}
\alias{add_covariate}
\title{Attach time-varying covariate}
\usage{
add_covariate(
  nif,
  sdtm,
  domain,
  testcd,
  covariate = NULL,
  DTC_field = NULL,
  DV_field = NULL,
  TESTCD_field = NULL,
  observation_filter = "TRUE",
  cat = NULL,
  scat = NULL,
  duplicate_function = mean,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{The corresponding sdtm object.}

\item{domain}{The domain as character.}

\item{testcd}{The xxTESTCD with xx the domain name, as character.}

\item{covariate}{The name of the covariate, defaults to the testcd if NULL.}

\item{DTC_field}{The field to use as the date-time code for the observation.
Defaults to the two-character domain name followed by 'DTC', if NULL.}

\item{DV_field}{The name of the DV field as character.}

\item{TESTCD_field}{The name of the TESTCD field. defaults to xxTESTCD (with
xx the domain code), as character.}

\item{observation_filter}{A filter term for the \code{domain}, as character.}

\item{cat}{xxCAT filter to apply, as character.}

\item{scat}{xxSCAT filter to apply, as character.}

\item{duplicate_function}{The function to apply if multiple covariate values
are found by day.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A nif object with a new column added that contains the time-varying covariate values.
The name of this column is determined by the \code{covariate} parameter (or defaults to
the value of \code{testcd} if not specified). The covariate values are matched to the nif
object by USUBJID and date. For each subject, missing covariate values are filled
using the last observed value (carrying forward).
}
\description{
A time-varying covariate is added as a new field with daily time granularity
and carried forward for missing entries. The name of the covariate can be
specified by 'covariate'. By default, it is set to the 'testcd' (without any
prefix).
}
\examples{
add_covariate(examplinib_poc_nif, examplinib_poc, "vs", "WEIGHT",
  covariate = "wt")
}
\seealso{
\code{\link[=add_baseline]{add_baseline()}}
}
