% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_dose_level}
\alias{add_dose_level}
\title{Add dose level column}
\usage{
add_dose_level(obj, silent = NULL)
}
\arguments{
\item{obj}{A NIF dataset.}

\item{silent}{Suppress messages.}
}
\value{
A NIF dataset.
}
\description{
Dose level is defined as the starting dose regimen for each ID.
}
