% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_generic_observation}
\alias{add_generic_observation}
\title{Add an observation from a generic source}
\usage{
add_generic_observation(
  obj,
  source,
  DTC_field,
  selector,
  DV_field,
  analyte_name,
  cmt = NULL,
  silent = FALSE
)
}
\arguments{
\item{obj}{A NIF object.}

\item{source}{The data source as data frame.}

\item{DTC_field}{The date-time information field as character.}

\item{selector}{R code to conduct filtering of the data set.}

\item{DV_field}{The DV field as data character.}

\item{analyte_name}{The name of the analyte in the output, as character.}

\item{cmt}{The comopartment for the added observation.}

\item{silent}{Suppress messages.}
}
\value{
A NIF object
}
\description{
Add an observation from a generic source
}
\examples{
add_generic_observation(examplinib_poc_nif, domain(examplinib_poc, "vs"),
  "VSDTC", {VSTESTCD == "WEIGHT"}, "VSSTRESN", "wt")
}
