% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_lab_covariate}
\alias{add_lab_covariate}
\title{Add lab covariate}
\usage{
add_lab_covariate(obj, lb, lbspec = "", lbtestcd, silent = F)
}
\arguments{
\item{obj}{The NIF data set.}

\item{lb}{The LB SDTM domain}

\item{lbspec}{The specimem, e.g., SERUM.}

\item{lbtestcd}{Lab parameters to be included as character scalar or vector.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
A NIF data set
}
\description{
Add lab covariate
}
