% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_lab_covariate}
\alias{add_lab_covariate}
\title{Add lab covariate}
\usage{
add_lab_covariate(obj, lb, lbspec = "SERUM", lbtestcd, silent = FALSE)
}
\arguments{
\item{obj}{The NIF object.}

\item{lb}{The LB SDTM domain}

\item{lbspec}{The specimen, e.g., SERUM.}

\item{lbtestcd}{Lab parameters to be included as character scalar or vector.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
A NIF object
}
\description{
This functions adds columns for the lab parameters specified in \code{lbtestcd},
in a time-varying way, i.e., the actual lab value at the time of the
observation or administration. This is in contrast to  \code{\link[=add_bl_lab]{add_bl_lab()}}.
In rows of missing lab data, the last value is carried forward.
}
\details{
Note that for some lab parameters, e.g., leukocytes, bili, etc., there may
be observations both in serum and in urine. It is therefore necessary to
specify the specimen tested. This corresponds to the \code{LBSPEC} field used in
the LB SDTM domain.
}
\seealso{
\code{\link[=add_bl_lab]{add_bl_lab()}}

\code{\link[=add_lab_observation]{add_lab_observation()}}
}
