% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_lab_observation}
\alias{add_lab_observation}
\title{Add lab value observation}
\usage{
add_lab_observation(
  obj,
  lb,
  lbtestcd,
  cmt = NULL,
  lbspec = "",
  parent = NA,
  silent = FALSE
)
}
\arguments{
\item{obj}{The NIF object.}

\item{lb}{The LB SDTM domain.}

\item{lbtestcd}{The LBTESTCD.}

\item{cmt}{A numerical value to specify the compartment this observation is
assigned to.}

\item{lbspec}{The LBSPECT.}

\item{parent}{The parent as character. If provided, the parent field defines
which parent compound administration is taken as the reference time when
time after dose (TAD) is calculated.}

\item{silent}{Switch to disable message output.}
}
\value{
The resulting NIF object.
}
\description{
Add lab value observation
}
