% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_lab_observation}
\alias{add_lab_observation}
\title{Add lab value as observation}
\usage{
add_lab_observation(obj, lb, lbtestcd, cmt = NULL, lbspec = "", silent = FALSE)
}
\arguments{
\item{obj}{The NIF object.}

\item{lb}{The LB SDTM domain.}

\item{lbtestcd}{The LBTESTCD.}

\item{cmt}{A numerical value to specify the compartment this observation is
assigned to.}

\item{lbspec}{The LBSPECT.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
The resulting NIF object.
}
\description{
Add lab value as observation
}
