% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_mapping}
\alias{add_mapping}
\title{Add mapping method for legacy compatibility}
\usage{
add_mapping(obj, extrt = "", pctestcd = "")
}
\arguments{
\item{obj}{A SDTM object.}

\item{extrt}{The treatment as defined in EX.}

\item{pctestcd}{The analyte as defined in PC.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
