% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_metabolite_mapping}
\alias{add_metabolite_mapping}
\title{Attach a parent-metabolite mapping to a SDTM object}
\usage{
add_metabolite_mapping(obj, pctestcd_parent, pctestcd_metabolite)
}
\arguments{
\item{obj}{The SDTM object.}

\item{pctestcd_parent}{The PCTESTCD of the parent compound.}

\item{pctestcd_metabolite}{The PCTESTCD of the metabolite.}
}
\value{
The SDTM object.
}
\description{
In case multiple analytes are measured for a specific administered drug, some
functions need that information to correlate plasma concentrations with
administrations.
}
\examples{
add_metabolite_mapping(examplinib_fe, "RS2023", "RS2023487A")
}
\seealso{
\code{\link[=nif_auto]{nif_auto()}}
}
