% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_observation}
\alias{add_observation}
\title{Add observation}
\usage{
add_observation(
  nif,
  sdtm,
  domain,
  DTC_field,
  DV_field,
  analyte = NA,
  cmt = NULL,
  parent = NULL,
  observation_filter = {
     TRUE
 },
  subject_filter = {
     !ACTARMCD \%in\% c("SCRNFAIL", "NOTTRT")
 },
  NTIME_lookup = NULL,
  silent = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{domain}{The domain as character.}

\item{DTC_field}{The field to use as the date-time code for the observation.}

\item{DV_field}{the field to use as the dependent variable.}

\item{analyte}{The name of the observed analyte.}

\item{cmt}{The compartment for the observation as numeric.}

\item{parent}{The name of the parent analyte for the observation as
character.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{NTIME_lookup}{A data frame with two columns, a column that defines the
custom nominal time information in the target domain (e.g., 'PCELTM'), and
'NTIME'. This data frame is left_join()ed into the observation data frame
to provide the NTIME field.}

\item{silent}{Suppress messages as logical.}

\item{cleanup}{Keep only necessary fields, as logical.}
}
\value{
A nif object.
}
\description{
Add observation
}
