% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{add_observation}
\alias{add_observation}
\title{Append observation events}
\usage{
add_observation(
  nif,
  sdtm,
  domain,
  testcd,
  analyte = NULL,
  parent = NULL,
  metabolite = FALSE,
  cmt = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  TESTCD_field = NULL,
  DTC_field = NULL,
  DV_field = NULL,
  coding_table = NULL,
  factor = 1,
  NTIME_lookup = NULL,
  ntime_method = NULL,
  keep = NULL,
  debug = FALSE,
  include_day_in_ntime = FALSE,
  silent = NULL,
  duplicates = "stop",
  duplicate_function = mean,
  na.rm = TRUE
)
}
\arguments{
\item{nif}{A nif object.}

\item{debug}{Include debug fields, as logical.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}

\item{duplicates}{Selection how to deal with duplicate observations with
respect to the USUBJID, ANALYTE and DTC fields:
\itemize{
\item 'stop': Stop execution and produce error message
\item 'ignore': Include duplicates in the data set
\item 'identify': Return a list of duplicate entries
\item 'resolve': Resolve duplicates, applying the \code{duplicate_function} to the
duplicate entries.
}}

\item{duplicate_function}{Function to resolve duplicate values, defaults to
\code{mean}.}

\item{na.rm}{Logical indicating whether to remove NA values when applying the
duplicate_function. Defaults to TRUE.}
}
\value{
A nif object.
}
\description{
Observations can be pharmacokinetic observations (i.e., from the PC domain),
or any other class of observation from any other SDTM domain. The 'testcd'
specifies the value of the respective \strong{TESTCD} field (e.g., 'PCTESTCD',
'VSTESTCD' or 'LBTESTCD') that defines the observation. Observation events
can be attached to an administered drug by specifying the 'parent' field.
This is required for, e.g., the time-after-dose ('TAD') and
time-after-first-dose ('TAFD') time calculation.
}
\details{
Observations can be further specified with the 'observation_filter' term. The
filter term can refer to any field of the respective SDTM domain.

A PK/PD model compartment can be specified with 'cmt' or will be
automatically assigned if \code{cmt = NULL}.
}
\seealso{
\code{\link[=add_administration]{add_administration()}}
}
