% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{add_observation}
\alias{add_observation}
\title{Append observation events}
\usage{
add_observation(
  nif,
  sdtm,
  domain,
  testcd,
  analyte = NULL,
  parent = NULL,
  metabolite = FALSE,
  cmt = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  TESTCD_field = NULL,
  DTC_field = NULL,
  DV_field = NULL,
  coding_table = NULL,
  factor = 1,
  NTIME_lookup = NULL,
  keep = NULL,
  debug = FALSE,
  include_day_in_ntime = FALSE,
  silent = NULL,
  duplicates = "stop",
  duplicate_function = mean
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{domain}{The domain as character.}

\item{testcd}{The observation variable, as character.}

\item{analyte}{The name for the analyte. Defaults to the 'testcd', if NULL.}

\item{parent}{The name of the parent analyte for the observation as
character. Defaults to the value of 'analyte' if NULL.}

\item{metabolite}{Observation is a metabolite, as logical.}

\item{cmt}{The compartment for the observation as numeric.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{TESTCD_field}{The xxTESTCD field. Defaults to the two-character domain
name followed by 'TESTCD', if NULL.}

\item{DTC_field}{The field to use as the date-time code for the observation.
Defaults to the two-character domain name followed by 'DTC', if NULL.}

\item{DV_field}{the field to use as the dependent variable. Defaults to the
two-character domain name followed by 'STRESN', if NULL.}

\item{coding_table}{Coding table to translate a categorical values into
numerical values, as data frame. The data frame must have at least one
column that matches a column in the domain, and a numerical 'DV' column
that provides the recoding result.}

\item{factor}{Multiplier for the DV field, as numeric.}

\item{NTIME_lookup}{A data frame with two columns, a column that defines the
custom nominal time information in the target domain (e.g., 'PCELTM'), and
'NTIME'. This data frame is left_joined into the observation data frame
to provide the NTIME field.}

\item{keep}{Columns to keep, as character.}

\item{debug}{Include debug fields, as logical.}

\item{include_day_in_ntime}{Include treatment day in the calculation of
NTIME, as logical.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}

\item{duplicates}{Selection how to deal with duplicate observations with
respect to the USUBJID, ANALYTE and DTC fields:
\itemize{
\item stop: stop execution and produce error message
\item identify: return a list of duplicate entries
\item remove: remove duplicates, applying the \code{duplicate_function} to the
duplicate entries.
}}

\item{duplicate_function}{Function to resolve duplicate values, defaults to
\code{mean}.}
}
\value{
A nif object.
}
\description{
Observations can be pharmacokinetic observations (i.e., from the PC domain),
or any other class of observation from any other SDTM domain. The 'testcd'
specifies the value of the respective \strong{TESTCD} field (e.g., 'PCTESTCD',
'VSTESTCD' or 'LBTESTCD') that defines the observation. Observation events
can be attached to an administered drug by specifying the 'parent' field.
This is required for, e.g., the time-after-dose ('TAD') and
time-after-first-dose ('TAFD') time calculation.
}
\details{
Observations can be further specified with the 'observation_filter' term. The
filter term can refer to any field of the respective SDTM domain.

A PK/PD model compartment can be specified with 'cmt' or will be
automatically assigned if \code{cmt = NULL}.
}
\examples{
add_observation(examplinib_fe_nif, examplinib_fe, "pc", "RS2023487A",
  parent = "RS2023")

}
\seealso{
\code{\link[=add_administration]{add_administration()}}
}
