% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{add_parent_mapping}
\alias{add_parent_mapping}
\title{Attach parent mapping to SDTM object}
\usage{
add_parent_mapping(obj, analyte, parent)
}
\arguments{
\item{obj}{A sdtm object.}

\item{analyte}{The analyte as character.}

\item{parent}{The parent as character.}
}
\value{
A sdtm object.
}
\description{
Attach parent mapping to SDTM object
}
