% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_tad}
\alias{add_tad}
\title{Add time-after-dose (TAD) field}
\usage{
add_tad(nif)
}
\arguments{
\item{nif}{A NIF object.}
}
\value{
The updated NIF object
}
\description{
Add time-after-dose (TAD) field
}
\examples{
add_tad(examplinib_poc_nif)
}
