% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_tad}
\alias{add_tad}
\title{Add time-after-dose (TAD) field}
\usage{
add_tad(nif)
}
\arguments{
\item{nif}{A NIF object.}
}
\value{
A NIF object with an added TAD column.
}
\description{
This function adds a time-after-dose (TAD) field to a NIF object. TAD represents
the time elapsed since the most recent administration of the parent compound.
For observations before the first dose, TAD will be negative.
}
\examples{
# Add TAD to a NIF object
add_tad(examplinib_poc_nif)
}
