% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{add_tad_non_pk}
\alias{add_tad_non_pk}
\title{Add time-after-dose (TAD) field for a non-PK analyte}
\usage{
add_tad_non_pk(nif, analyte = NULL, parent = NULL)
}
\arguments{
\item{nif}{A NIF object.}

\item{analyte}{The analyte as character.}

\item{parent}{The drug to set the administration reference to, as character.}
}
\value{
A NIF object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
