% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_window.R
\name{add_time_deviation}
\alias{add_time_deviation}
\title{Calculate time deviations for observations}
\usage{
add_time_deviation(obj, silent = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{silent}{Suppress messages.}
}
\value{
A nif object with the 'TIME_DEV' field added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate the time difference in hours between the actual time after dose
(TAD) and the scheduled time (NTIME) of observations.

If NTIME == 0, the data point is considered pre-dose, and instead of TAD, the
(negative) time difference to the next administration is calculated!
}
\examples{
library(dplyr)

examplinib_poc_nif |>
  add_time_deviation() |>
  head()

}
