% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_time.R
\name{add_trtdy}
\alias{add_trtdy}
\title{Add treatment day ('TRTDY') column}
\usage{
add_trtdy(obj)
}
\arguments{
\item{obj}{The NIF object as data frame.}
}
\value{
The updated NIF object as data frame.
}
\description{
Add treatment day ('TRTDY') column
}
\examples{
head(add_trtdy(examplinib_poc_nif))
}
