% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{add_TRTDY}
\alias{add_TRTDY}
\title{Add treatment day}
\usage{
add_TRTDY(obj)
}
\arguments{
\item{obj}{The NIF data set as data frame.}
}
\value{
The updated NIF data set as data frame.
}
\description{
Add treatment day
}
