% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_analysis.R
\name{administration_summary}
\alias{administration_summary}
\title{Overview on the number of administrations in the subjects by parent}
\usage{
administration_summary(obj)
}
\arguments{
\item{obj}{A NIF data set.}
}
\value{
A data frame.
}
\description{
Overview on the number of administrations in the subjects by parent
}
\examples{
administration_summary(examplinib_nif)

}
