% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{ae_summary}
\alias{ae_summary}
\title{AE summary}
\usage{
ae_summary(
  sdtm_data,
  level = "SOC",
  show_cd = FALSE,
  group = NULL,
  order_by_subj = F,
  ae_filter = "TRUE"
)
}
\arguments{
\item{sdtm_data}{An sdtm object.}

\item{level}{The level to summarize by, as character. Can be one or multiple
of:
\itemize{
\item 'TERM':   Reported term
\item 'LLT':    Lowest level term
\item 'DECOD':  Dictionary-derived term
\item 'HLT':    High level term
\item 'BODSYS': Body system or organ class
\item 'SOC':    System organ class
}}

\item{show_cd}{Show AE term code, as logical.}

\item{group}{Additional grouping variable, as character.}

\item{order_by_subj}{Order by number of subject, instead of by number of
event, as logical.}

\item{ae_filter}{A filter term as character.}
}
\value{
A data frame.
}
\description{
AE summary
}
