% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{analyte_overview}
\alias{analyte_overview}
\title{Analyte overview}
\usage{
analyte_overview(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A data frame.
}
\description{
Overview on the analytes included in a nif object and their respective parent
analytes.
}
\examples{
analyte_overview(examplinib_poc_nif)
}
