% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{analytes}
\alias{analytes}
\title{Analytes within a nif object}
\usage{
analytes(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A character vector of all analytes in the data set.
}
\description{
Analytes within a nif object
}
