% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_auto.R
\name{auto_mapping}
\alias{auto_mapping}
\title{Create analyte mapping table from mapping formula or list}
\usage{
auto_mapping(sdtm, mapping = NULL)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{mapping}{A formula or list of formulae. Each formula should be in the
format PCTESTCD ~ EXTRT or PCTESTCD1 + PCTESTCD2 ~ EXTRT for multiple
analytes from the same treatment.}
}
\value{
A data frame with unique combinations of EXTRT and PCTESTCD.
}
\description{
Creates a mapping between treatments (EXTRT) and analytes (PCTESTCD). Each
combination of EXTRT and PCTESTCD must be unique. If duplicate mappings are
found, an error will be raised.
}
