% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_auto.R
\name{auto_mapping}
\alias{auto_mapping}
\title{Create analyte mapping table from mapping formula or list}
\usage{
auto_mapping(sdtm, mapping = NULL)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{mapping}{A formla or list of formulae.}
}
\value{
A data frame.
}
\description{
Create analyte mapping table from mapping formula or list
}
