% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{baseline_covariates}
\alias{baseline_covariates}
\title{Extract baseline vital sign covariates from VS}
\usage{
baseline_covariates(vs, silent = FALSE)
}
\arguments{
\item{vs}{The VS domain as data frame.}

\item{silent}{Switch to disable message output.}
}
\value{
Baseline VS data as wide data frame.
}
\description{
Extract baseline vital sign covariates from VS
}
\keyword{internal}
