% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_qtime.R
\name{bintime_plot}
\alias{bintime_plot}
\title{BINTIME plot}
\usage{
bintime_plot(
  obj,
  analyte,
  method = "fisher",
  points = FALSE,
  caption = TRUE,
  size = 1.5,
  alpha = 1
)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{The analyte as character.}

\item{method}{Univariate class intervals method, can be one of jenks, kmeans,
pretty, quantile, hclust, sd, bclust or fisher. See classInt::classInterval()
for details.}

\item{points}{Plot original data points as logical.}

\item{caption}{Show caption as logical.}

\item{size}{The point size.}

\item{alpha}{The alpha parameter for the data points.}
}
\value{
A ggplot2 object.
}
\description{
BINTIME plot
}
