% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_qtime.R
\name{bintime_plot}
\alias{bintime_plot}
\title{BINTIME plot}
\usage{
bintime_plot(
  obj,
  analyte,
  method = "kmeans",
  time = "TAFD",
  min_time = NULL,
  max_time = NULL,
  points = FALSE,
  cfb = FALSE,
  caption = TRUE,
  title = NULL,
  size = 1.5,
  alpha = 1,
  refline = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{The analyte as character.}

\item{method}{Univariate class intervals method, can be one of jenks, kmeans,
pretty, quantile, hclust, sd, bclust or fisher. See classInt::classInterval()
for details.}

\item{time}{The time field.}

\item{min_time}{The minimal time in units of TAFD, as numeric.}

\item{max_time}{The minimal time in units of TAFD, as numeric.}

\item{points}{Plot original data points as logical.}

\item{cfb}{Plot change from baseline, as logical.}

\item{caption}{Show caption as logical.}

\item{title}{The plot title, as character. If none is provided, a generic
title based on the analyte will be chosen. Override this by setting
.  title = "", if needed.}

\item{size}{The point size.}

\item{alpha}{The alpha parameter for the data points.}

\item{refline}{Plot horizontal dashed reference lines at these y axis values,
defaults to NULL (no lines).}
}
\value{
A ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
