% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_check.R
\name{check}
\alias{check}
\title{Check nif or sdtm object}
\usage{
check(obj, ...)
}
\arguments{
\item{obj}{The nif or sdtm object}
}
\value{
The input with the CHECK field added.
}
\description{
Check nif or sdtm object
}
