% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{clip_nif}
\alias{clip_nif}
\title{Reduce a NIF object on the subject level by excluding all
administrations after the last observation}
\usage{
clip_nif(nif)
}
\arguments{
\item{nif}{A NIF dataset.}
}
\value{
A NIF dataset.
}
\description{
Reduce a NIF object on the subject level by excluding all
administrations after the last observation
}
\examples{
clip_nif(examplinib_poc_nif)
}
