% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{cmt_mapping}
\alias{cmt_mapping}
\title{Analyte-to-compartment mapping}
\usage{
cmt_mapping(obj)
}
\arguments{
\item{obj}{A NIF object}
}
\value{
A data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
