% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{compose_dtc}
\alias{compose_dtc}
\title{Compose DTC from date and time components}
\usage{
compose_dtc(date, time)
}
\arguments{
\item{date}{A date in POSIX or character format.}

\item{time}{A time in character format.}
}
\value{
A POSICct object
}
\description{
Compose DTC from date and time components
}
