% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{compress}
\alias{compress}
\title{Generic function for compress}
\usage{
compress(nif, fields = standard_nif_fields, debug = TRUE)
}
\arguments{
\item{nif}{A NIF data set object.}

\item{fields}{Fields to be included as character. If 'fields' is NULL
(default), the fields defined in \code{standard_nif_fields} plus all fields with
a name that starts with 'BL_' (baseline covariates) are included.}

\item{debug}{Boolean to indicate whether the debug fields, \code{PCREFID} and
\code{EXSEQ} should be included in the NIF data set.}
}
\value{
A NIF data set object.
}
\description{
Generic function for compress
}
