% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif_deprecated.R
\name{compress_nif}
\alias{compress_nif}
\title{This function removes columns from a NIF object that are not needed for
downstream analysis}
\usage{
compress_nif(nif, ...)
}
\arguments{
\item{nif}{A NIF object.}

\item{...}{Further optional parameters are fields to be included in the
output. If none are provided, the standard set will be used.}
}
\value{
A NIF dataset with only the specified fields
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{compress_nif()} has been superseded in favor of \code{\link[=compress]{compress()}}.
}
\details{
During creating of a NIF object using \code{\link[=make_nif]{make_nif()}}, multiple SDTM tables are
aggregated without deleting the original fields (columns). Many of these
fields may not be required for the final analysis. This function reduces the
fields to those typically required in the downstream analysis. Applying
\code{\link[=compress_nif]{compress_nif()}} is typically the last step in the creation of a NIF data
set, after creating the basic NIF object  with \code{\link[=make_nif]{make_nif()}} and applying
custom imputations and manually deriving convariates as needed.
}
\keyword{internal}
