% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conditional_message}
\alias{conditional_message}
\title{Issue message based on silent flag}
\usage{
conditional_message(..., silent = NULL)
}
\arguments{
\item{...}{Message components.}

\item{silent}{Logical flag to suppress messages. If NULL, will check .nif_env
environment (defaults to FALSE if not found).}
}
\value{
Nothing.
}
\description{
Issue message based on silent flag
}
\keyword{internal}
