% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{conditional_message}
\alias{conditional_message}
\title{Issue message based on silent flag}
\usage{
conditional_message(msg, ..., silent = FALSE)
}
\arguments{
\item{msg}{The message as character.}

\item{...}{Further message components.}

\item{silent}{A boolean.}
}
\value{
Nothing.
}
\description{
Issue message based on silent flag
}
