% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pivot.R
\name{correlate_obs}
\alias{correlate_obs}
\title{Correlate observations}
\usage{
correlate_obs(
  obj,
  indep_analyte,
  dep_analyte,
  window = 10/60,
  time_field = "TIME",
  duplicate_function = mean
)
}
\arguments{
\item{obj}{A nif object.}

\item{indep_analyte}{The independent analyte as character.}

\item{dep_analyte}{The dependent analyte(s) as character.}

\item{window}{The allowed time window between the independent and dependent
analyte observations in hours.}

\item{time_field}{Time variable to use for the correlation of observations,
can be 'DTC' or 'TIME' (default).}

\item{duplicate_function}{A function to resolve duplicate values for
observations of the dependent analyte. Defaults to \code{mean}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Identify observations of the dependent variables that are timely correlated
to the independent variable, i.e., are within a specified time window to the
observations of the independent variable. If multiple dependent observations
of a given analyte fall into the window, they are summarized using the
'duplicate_function'. The output data frame is a wide table based on the
independent observations, with the (summarized) dependent observations and
the respective summarized observation times for the dependent obsrvations as
additional columns.
}
\examples{
head(correlate_obs(examplinib_poc_nif, "RS2023", "RS2023487A"), 3)

}
