% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{covariate_hist}
\alias{covariate_hist}
\title{Generic covariate distribution histogram}
\usage{
covariate_hist(
  obj,
  cov,
  nbins = 11,
  group = NULL,
  alpha = 0.5,
  density = TRUE,
  title = NULL
)
}
\arguments{
\item{obj}{The NIF object.}

\item{cov}{The covariate field of the NIF object to analyze, as character.}

\item{nbins}{The number of bins to be used if no binwidth is specified.
Defaults to 11.}

\item{group}{The field(s) to group by, as character.}

\item{alpha}{The alpha as numeric.}

\item{density}{Plot density instead of N on the y axis, as logical.}

\item{title}{The title as character.}
}
\value{
A ggplot object.
}
\description{
Generic covariate distribution histogram
}
\examples{
covariate_hist(examplinib_sad_nif, "AGE")
covariate_hist(examplinib_sad_nif, "BL_CRCL")
}
