% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_administrations.R
\name{create_iv_fields}
\alias{create_iv_fields}
\title{Check and prepare ex for iv administrations}
\usage{
create_iv_fields(admin, iv_admin = NULL, duration = NULL, silent = NULL)
}
\arguments{
\item{admin}{The un-expanded EX domain}

\item{iv_admin}{Administration is iv, as boolean. If NULL, will be tested
automatically based on the EXROUTE field, if available.}

\item{duration}{The duration of the iv administration as numeric. If NULL,
will be determined automatically based on the EXDUR field, if available.}

\item{silent}{Suppress messages.}
}
\value{
The updated EX domain.
}
\description{
Check and prepare ex for iv administrations
}
