% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_dataset.R
\name{dataset}
\alias{dataset}
\title{Retrieve dataset from adam object}
\usage{
dataset(adam, name)
}
\arguments{
\item{adam}{An adam object.}

\item{name}{The dataset name as character.}
}
\value{
A data frame.
}
\description{
Overview: \href{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3}{ADaMIG v1.3}
}
