% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_class.R
\name{dataset}
\alias{dataset}
\title{Retrieve dataset from adam object}
\usage{
dataset(adam, name)
}
\arguments{
\item{adam}{An adam object.}

\item{name}{The dataset name as character.}
}
\value{
A data frame.
}
\description{
Retrieve dataset from adam object
}
