% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ddt.R
\name{ddt}
\alias{ddt}
\title{Data definition table for NIF object}
\usage{
ddt(obj, silent = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{silent}{Suppress messages, defaults to nif_option setting if NULL.}
}
\value{
A data table.
}
\description{
Data definition table for NIF object
}
\examples{
ddt(examplinib_sad_nif)
}
