% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{derive_sld}
\alias{derive_sld}
\title{Calculate SLD for SDTM.TR domain}
\usage{
derive_sld(sdtm_obj, observation_filter = "TRGRPID == 'TARGET'")
}
\arguments{
\item{sdtm_obj}{A SDTM object.}

\item{observation_filter}{A filter term, as character.}
}
\value{
A SDTM object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Calculate the sum of longest diameters (SLD) and number of lesions by subject
and time point, and add as rows. If a SDTM.TR domain is not included in the
input SDTM object, the SDTM object is returned as-is.
}
