% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_to_string}
\alias{df_to_string}
\title{Render data frame object to string}
\usage{
df_to_string(
  df,
  indent = "",
  n = NULL,
  header = TRUE,
  color = FALSE,
  show_none = FALSE
)
}
\arguments{
\item{df}{The data.frame to be rendered}

\item{indent}{A string that defines the left indentation of the rendered
output.}

\item{n}{The number of lines to be included, or all if NULL.}

\item{header}{Boolean to indicate whether the header row is to be included.}

\item{color}{Print headers in grey as logical.}

\item{show_none}{Show empty data frame as 'none', as logical.}
}
\value{
The output as string.
}
\description{
This function renders a data.frame into a string similar to its
representation when printed without line numbers
}
\keyword{internal}
