% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{domain}
\alias{domain}
\title{Get a specific domain from a SDTM object}
\usage{
domain(obj, name)
}
\arguments{
\item{obj}{The sdtm object.}

\item{name}{The domain to return.}
}
\value{
The specified domain as data.frame
}
\description{
Get a specific domain from a SDTM object
}
\examples{
domain(examplinib, "dm")
}
