% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{domain}
\alias{domain}
\title{Return a specific domain from a sdtm object}
\usage{
domain(obj, name)
}
\arguments{
\item{obj}{The sdtm object.}

\item{name}{The domain to return.}
}
\value{
The specified domain as data.frame
}
\description{
Return a specific domain from a sdtm object
}
\examples{
head(domain(examplinib_fe, "dm"))
}
