% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{domain}
\alias{domain}
\title{Return a specific domain from a sdtm object}
\usage{
domain(obj, name)
}
\arguments{
\item{obj}{The sdtm object.}

\item{name}{The domain to return as a single character string.}
}
\value{
The specified domain as data frame. Issues a warning if the domain
does not exist and returns NULL.
}
\description{
Return a specific domain from a sdtm object
}
\examples{
head(domain(examplinib_fe, "dm"), 3)
}
