% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{domain}
\alias{domain}
\title{Get a specific domain from the sdtm object}
\usage{
domain(obj, dom = "")
}
\arguments{
\item{obj}{The sdtm object.}

\item{dom}{The code of the domain to be returned.}
}
\value{
The specified domain as data.frame
}
\description{
Get a specific domain from the sdtm object
}
