% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{dose_levels}
\alias{dose_levels}
\title{Dose levels}
\usage{
dose_levels(obj, grouping = NULL)
}
\arguments{
\item{obj}{A nif object.}

\item{grouping}{Further fields to be included (and to be grouped by) in the
output.}
}
\value{
A data frame
}
\description{
In studies with multiple administrations, dose modifications may occur. The
\code{dose_levels()} function only summarizes the initial (i.e., the first)
administrations.
}
\examples{
dose_levels(examplinib_nif)
dose_levels(examplinib_nif, grouping="SEX")
dose_levels(examplinib_nif, grouping=c("SEX", "FASTED"))
}
