% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{dose_levels}
\alias{dose_levels}
\title{Dose levels within a NIF object}
\usage{
dose_levels(obj, cmt = 1, group = NULL)
}
\arguments{
\item{obj}{A rich or minimal NIF object.}

\item{cmt}{The compartment (CMT) as numeric.}

\item{group}{Further fields to be included (and to be grouped by) in the
output.}
}
\value{
A data frame.
}
\description{
This function summarizes the doses in the individual first administration by
subject and drug, and the number of subjects treated at this dose level.
Subsequent dose modifications are ignored.
}
\examples{
dose_levels(examplinib_fe_nif)
dose_levels(examplinib_fe_nif, group = "SEX")
dose_levels(examplinib_fe_nif, group = c("SEX", "FASTED"))
dose_levels(examplinib_sad_min_nif)
dose_levels(new_nif())
}
