% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{dose_plot_id}
\alias{dose_plot_id}
\title{Plot dose time course data from individual subject}
\usage{
dose_plot_id(
  obj,
  id,
  y_scale = "lin",
  max_dose = NA,
  time_field = "TIME",
  point_size = 2,
  max_time = NA,
  analyte = NULL,
  ...
)
}
\arguments{
\item{obj}{The NIF object.}

\item{id}{The subject ID to be plotted.}

\item{y_scale}{Y-scale. Use 'scale="log"' for a logarithmic y scale. Default
is "lin".}

\item{max_dose}{The upper limit of the dose scale.}

\item{time_field}{The field to use as the time metric, as character.}

\item{point_size}{The point size as numeric.}

\item{max_time}{The right limit of the time scale.}

\item{analyte}{The analyte of interest.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
This function plots AMT over TIME for an individual subject, id. Id can be
either the ID or the USUBJID. Administration time points are indicated with
vertical lines.
}
\examples{
dose_plot_id(examplinib_poc_nif, 18)
dose_plot_id(examplinib_poc_nif, dose_red_sbs(examplinib_poc_nif)[1, 1])
dose_plot_id(examplinib_poc_min_nif, 18)
}
\keyword{internal}
