% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{dose_plot_id}
\alias{dose_plot_id}
\title{Plot dose time course data from individual subject}
\usage{
dose_plot_id(
  nif,
  id,
  y_scale = "lin",
  max.dose = NA,
  max.time = NA,
  analyte = NULL
)
}
\arguments{
\item{nif}{The NIF data set.}

\item{id}{The subject ID to be plotted.}

\item{y_scale}{Y-scale. Use 'scale="log"' for a logarithmic y scale. Default
is "lin".}

\item{max.dose}{The upper limit of the dose scale.}

\item{max.time}{The right limit of the time scale.}

\item{analyte}{The analyte of interest.}
}
\value{
A ggplot object.
}
\description{
This function plots AMT over TIME for an individual subject, id. Id can be
either the ID or the USUBJID. Administration time points are indicated with
vertical lines.
}
