% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{dose_red_sbs}
\alias{dose_red_sbs}
\title{Subjects with dose reduction}
\usage{
dose_red_sbs(obj, analyte = NULL)
}
\arguments{
\item{obj}{A NIF object object.}

\item{analyte}{The analyte of interest as string. considers all analytes if
analyte is NULL (default).}
}
\value{
The IDs.
}
\description{
Subjects with dose reduction
}
\examples{
dose_red_sbs(examplinib_poc_nif)
}
