% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{doses}
\alias{doses}
\title{Doses in a nif or sdtm object}
\usage{
doses(obj)
}
\arguments{
\item{obj}{A nif or sdtm object.}
}
\value{
Numeric.
}
\description{
Doses in a nif or sdtm object
}
\examples{
doses(examplinib_poc_nif)
}
