% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{doses}
\alias{doses}
\title{Doses within a NIF object}
\usage{
doses(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A number vector of all doses (AMT) in the data set.
}
\description{
Doses within a NIF object
}
\examples{
doses(examplinib_poc_nif)
}
