% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{dtc_formats}
\alias{dtc_formats}
\title{The list of expected date/time formats as per ISO 8601}
\format{
An object of class \code{character} of length 4.
}
\usage{
dtc_formats
}
\description{
The list of expected date/time formats as per ISO 8601
}
\keyword{internal}
