% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\docType{data}
\name{dtc_formats}
\alias{dtc_formats}
\title{The list of expected date/time formats in the SDTM data}
\format{
An object of class \code{character} of length 4.
}
\usage{
dtc_formats
}
\description{
The list of expected date/time formats in the SDTM data
}
\keyword{datasets}
