% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{edish_plot}
\alias{edish_plot}
\title{Drug-induced serious hepatotoxicity (eDISH) plot}
\usage{
edish_plot(
  nif,
  sdtm,
  enzyme = "ALT",
  observation_filter = NULL,
  baseline_filter = "LBBLFL == 'Y'",
  show_labels = FALSE,
  shading = TRUE,
  title = "eDISH plot",
  size = 3,
  alpha = 0.5,
  silent = NULL,
  ...
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{enzyme}{The transaminase enzyme to be plotted on the x axis as
character, can be 'AST' or 'ALT' (default).}

\item{observation_filter}{A filter term as character.}

\item{baseline_filter}{A filter term to identify baseline conditions.}

\item{show_labels}{Show ID labels per point.}

\item{shading}{Highlight Hy's law area, as logical.}

\item{title}{The plot title as character.}

\item{size}{The point size as numeric.}

\item{alpha}{The alpha value as numeric.}

\item{silent}{Suppress messages.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Refer to the FDA guidance on DILI, https://www.fda.gov/media/116737/download
and Watkins 2011, \doi{10.2165/11586600-000000000-00000}.
}
