% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{edish_plot}
\alias{edish_plot}
\title{Drug-induced serious hepatotoxicity (eDISH) plot}
\usage{
edish_plot(
  nif,
  sdtm,
  enzyme = "ALT",
  observation_filter = "LBSPEC != 'URINE'",
  show_labels = FALSE,
  autoscale = TRUE,
  shading = TRUE,
  nominal_time = TRUE,
  ntime_method = NULL,
  time = NULL,
  parent = NULL,
  title = "eDISH plot: All time points",
  size = 3,
  alpha = 0.5,
  ...
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{enzyme}{The transaminase enzyme to be plotted on the x axis as
character, can be 'AST' or 'ALT' (default).}

\item{observation_filter}{A filter term as character.}

\item{show_labels}{Show ID labels per point.}

\item{autoscale}{Use automatic axis scaling, as logical. Defaults to
0.01-1000 for ALT/AST and 0.01-100 for bili.}

\item{shading}{Highlight Hy's law area, as logical.}

\item{nominal_time}{Use NTIME as logical.}

\item{time}{time/nominal time filter as numeric.}

\item{parent}{The parent compound as character.}

\item{title}{The plot title as character.}

\item{size}{The point size as numeric.}

\item{alpha}{The alpha value as numeric.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Refer to the \href{https://www.fda.gov/media/116737/download}{FDA guidance on Drug-induced liver injury}
and \href{https://doi.org/10.2165/11586600-000000000-00000}{Watkins 2011}.
}
