% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{egfr_cg}
\alias{egfr_cg}
\title{Glomerular filtration rate estimation from serum creatinine (Cockcroft-Gault)}
\usage{
egfr_cg(crea, age, sex, race = "", weight = NA, molar = F)
}
\arguments{
\item{crea}{Serum creatinine in mg/dl.}

\item{age}{Age in years.}

\item{sex}{Sex encocded as number (female is 1) or character (female is "F").}

\item{race}{Race as per CDISC nomenclature. Black race is identified as the
occurrence of 'black' in the value. For campatibility onla, not used in
this formula.}

\item{weight}{Body weight in kg.}

\item{molar}{Switch to select whether the creatinine value is in mg/dl
(default) or umol/l units.
#'}
}
\value{
Estimated GFR in ml/min (as body size is accounted for by the weigth
in the input).
}
\description{
Reference:
\href{https://doi.org/10.1159/000180580}{Cockcroft and Gault, Nephron 1976}
}
