% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{egfr_mdrd}
\alias{egfr_mdrd}
\title{Glomerular filtration rate estimation from serum creatinine (MDRD)}
\usage{
egfr_mdrd(crea, age, sex, race = "", weight = NA, molar = FALSE)
}
\arguments{
\item{crea}{Serum creatinine in mg/dl.}

\item{age}{Age in years.}

\item{sex}{Sex encoded as number (female is 1) or character (female is "F").}

\item{race}{Race as per CDISC nomenclature. Black race is identified as the
occurrence of 'black' in the value.}

\item{weight}{Body weight. Not used in this formula but included for
compatibility.}

\item{molar}{Switch to select whether the creatinine value is in mg/dl
(default) or umol/l units.}
}
\value{
Estimated GFR in ml/min/1.73 m^2.
}
\description{
Source: doi.org/10.7326/0003-4819-130-6-199903160-00002.
}
\details{
\deqn{
 eGFR = 175 * crea^{1.154} * age^{0.203} * 0.742 (female) * 1.212 (black)
}
}
\seealso{
\code{\link[=egfr_raynaud]{egfr_raynaud()}}

\code{\link[=egfr_cg]{egfr_cg()}}
}
