% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiological_calculations.R
\name{egfr_raynaud}
\alias{egfr_raynaud}
\title{Glomerular filtration rate estimation from serum creatinine (Raynaud method)}
\usage{
egfr_raynaud(crea, age, sex, race = "", weight = NA, molar = F)
}
\arguments{
\item{crea}{Serum creatinine in mg/dl or umol/l (if molar=TRUE).}

\item{age}{Age in years.}

\item{sex}{Sex encoded as number (male is 0) or character (male is "M").}

\item{race}{Race. Dummy variable for compatibility, race is not used by this
method.}

\item{weight}{Body weight. Not used in this formula but included for
compatibility.}

\item{molar}{Switch to select whether the creatinine value is in mg/dl
(default) or umol/l units.}
}
\value{
Estimated GFR in ml/min/1.73 m^2.
}
\description{
Source: Raynaud M, et al., Race-free estimated glomerular
filtration rate equation in kidney transplant recipients: development and
validation study. BMJ. 2023 May 31;381:e073654.
\href{https://doi.org/10.1136/bmj-2022-073654}{doi: 10.1136/bmj-2022-073654}.
PMID: 37257905; PMCID: PMC10231444.
}
\details{
\deqn{
 eGFR = e^{4.43 - 0.82 * ln(crea) - 0.012 * crea^2 - 0.006 * age + 0.18 * male}
}

with:
\itemize{
\item \code{crea} in mg/dl
\item \code{age} in years
\item \code{male} 1 for male and 0 for female patients
}
}
\seealso{
\code{\link[=efgr_mdrd]{efgr_mdrd()}}

\code{\link[=egfr_cg]{egfr_cg()}}
}
