% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_dose}
\alias{ensure_dose}
\title{Ensure that the DOSE field is present}
\usage{
ensure_dose(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
Ensure that the DOSE field is present
}
\keyword{internal}
