% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_tad}
\alias{ensure_tad}
\title{Ensure that TAD is present in the NIF object}
\usage{
ensure_tad(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A NIF object.
}
\description{
Ensure that TAD is present in the NIF object
}
\keyword{internal}
