% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{ensure_tafd}
\alias{ensure_tafd}
\title{Ensure TAFD is present in the NIF object}
\usage{
ensure_tafd(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object.
}
\description{
Ensure TAFD is present in the NIF object
}
\keyword{internal}
