% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib}
\alias{examplinib}
\title{Sample SDTM data}
\format{
A SDTM object
}
\usage{
examplinib
}
\description{
Synthetic SDTM data from a fictional clinical study with 'examplinib'
}
\keyword{datasets}
