% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_iv}
\alias{examplinib_iv}
\title{Synthetic SDTM data from a fictional clinical iv administration study for
'examplinib'.}
\format{
A sdtm object
}
\usage{
examplinib_iv
}
\description{
Synthetic SDTM data from a fictional clinical iv administration study for
'examplinib'.
}
\keyword{datasets}
