% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{exclude_EXSTDTC_after_RFENDTC}
\alias{exclude_EXSTDTC_after_RFENDTC}
\title{Title}
\usage{
exclude_EXSTDTC_after_RFENDTC(ex, dm, silent = FALSE)
}
\arguments{
\item{ex}{The EX domain as data.frame.}

\item{dm}{The DM domain as data.frame}

\item{silent}{A boolean.}
}
\value{
A data frame.
Filter out EX events after the last dose as specified in RFENDTC

The modified EX domain as data.frame.
}
\description{
Title
}
