% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_date}
\alias{extract_date}
\title{Extract thendate component of a POSICct object}
\usage{
extract_date(dtc)
}
\arguments{
\item{dtc}{The POSIX-formatted datetime.}
}
\value{
The date as character.
}
\description{
Extract thendate component of a POSICct object
}
\keyword{internal}
