% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{filter_correct_date_format}
\alias{filter_correct_date_format}
\title{Filter out any rows containing DTC rows with incomplete data format}
\usage{
filter_correct_date_format(obj, verbose = T, silent = F)
}
\arguments{
\item{obj}{The SDTM domain as data frame}

\item{verbose}{Boolean to indicate whether to include details.}

\item{silent}{Boolean to indicate whether message output is produced.}
}
\value{
The filtered SDTM domain as data frame.
}
\description{
Filter out any rows containing DTC rows with incomplete data format
}
