% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{filter_EXSTDTC}
\alias{filter_EXSTDTC}
\title{Remove administrations with EXSTDTC after RFENDTC}
\usage{
filter_EXSTDTC(ex, dm, silent = FALSE)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{dm}{the DM domain as data frame.}

\item{silent}{Suppress messages as logical.}
}
\value{
A data frame.
}
\description{
Remove administrations with EXSTDTC after RFENDTC
}
\keyword{internal}
