% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{filter_EXSTDTC_after_EXENDTC}
\alias{filter_EXSTDTC_after_EXENDTC}
\title{Remove administrations with EXSTDTC after EXENDTC}
\usage{
filter_EXSTDTC_after_EXENDTC(ex, dm, silent = NULL)
}
\arguments{
\item{ex}{The ex domain as data frame.}

\item{dm}{The dm domain as data frame.}

\item{silent}{Suppress messages, defaults to nif_option setting, if NULL.}
}
\value{
A data frame.
}
\description{
Remove administrations with EXSTDTC after EXENDTC
}
