% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{filter_subject}
\alias{filter_subject}
\title{Keep only selected USUBJID in the data set}
\usage{
filter_subject(obj, usubjid)
}
\arguments{
\item{obj}{The input, either a \code{nif} or \code{sdtm} object.}

\item{usubjid}{The USUBJID as character.}
}
\value{
The filtered object.
}
\description{
Keep only selected USUBJID in the data set
}
\examples{
filter_subject(examplinib_poc, subjects(examplinib_poc)[1, "USUBJID"])
filter_subject(examplinib_poc_nif, subjects(examplinib_poc_nif)[1, "USUBJID"])
}
