% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_duplicates.R
\name{gather_duplicates}
\alias{gather_duplicates}
\title{Consolidate multiplicate observations}
\usage{
gather_duplicates(
  obj,
  id_field = "NTIME",
  duplicate_function = mean,
  na_rm = TRUE,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{id_field}{The field(s) over which to identify duplicates (in addition
to ID ANALYTE and CMT which are automatically considered).}

\item{duplicate_function}{A function by which to consolidate the
multiplicates}

\item{na_rm}{Remove NA values.}

\item{silent}{Suppress messages.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Identify multiplicate observations over ID, CMT and 'fields' and integrate
the DV value for them using the 'duplicate_function'. A common application is
the averaging of triplicate ECG parameter observations by NTIME.

Observations with MDV == 1 are excluded from the process.
}
