% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_extension.R
\name{geom_admin}
\alias{geom_admin}
\title{Administration geom layer for ggplot}
\usage{
geom_admin(
  mapping = NULL,
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  color = "grey",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If \code{NULL}, the default, the data is inherited from the plot data as
specified in the call to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. A \code{data.frame}, or other object, will
override the plot data. A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and will be used as
the layer data.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a
warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never
includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than
combining with them. This is most useful for helper functions that define
both data and aesthetics and shouldn't inherit behavior from the default
plot specification.}

\item{color}{The color of the vertical lines. Defaults to "grey".}

\item{linewidth}{The width of the lines. Defaults to 0.5.}

\item{linetype}{The type of the lines. Defaults to 1 (solid).}

\item{alpha}{The transparency of the lines. Defaults to NA (fully opaque).}

\item{...}{Additional parameters passed to the layer.}
}
\value{
A ggplot layer object.
}
\description{
Administration geom layer for ggplot
}
