% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{guess_analyte}
\alias{guess_analyte}
\title{Guess the most likely analyte based on its prevalence in the NIF object}
\usage{
guess_analyte(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
The analyte as character.
}
\description{
Guess the most likely analyte based on its prevalence in the NIF object
}
\examples{
guess_analyte(examplinib_poc_nif)

}
