% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{guess_lbspec}
\alias{guess_lbspec}
\title{Guess the most likely LBSPEC}
\usage{
guess_lbspec(lb, silent = TRUE)
}
\arguments{
\item{lb}{The LB SDTM domain as data frame.}

\item{silent}{Switch to disable message output.}
}
\value{
the imputed LBSPEC as character.
}
\description{
Guess the most likely LBSPEC
}
\examples{
guess_lbspec(domain(examplinib_poc, "lb"))
}
\keyword{internal}
