% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{has_domain}
\alias{has_domain}
\title{Check whether a domain is present in an SDTM object}
\usage{
has_domain(obj, name)
}
\arguments{
\item{obj}{The sdtm object.}

\item{name}{The domain name(s) to check as a character string or vector.}
}
\value{
Logical indicating whether all specified domains exist in the SDTM
object.
}
\description{
Check whether a domain is present in an SDTM object
}
\examples{
# Check if DM domain exists
has_domain(examplinib_fe, "dm")

# Check if a non-existent domain exists
has_domain(examplinib_fe, "xyz")

# Check if multiple domains exist
has_domain(examplinib_fe, c("dm", "vs"))
}
\keyword{internal}
