% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{has_time}
\alias{has_time}
\title{Check whether POSIX datetime object includes time information}
\usage{
has_time(obj)
}
\arguments{
\item{datetime}{The datetime object as POSIX or as ISO 8601-formatted string.}
}
\value{
A Boolean value.
}
\description{
Check whether POSIX datetime object includes time information
}
\keyword{internal}
