% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{has_time}
\alias{has_time}
\title{Check whether POSIX datetime object includes time information}
\usage{
has_time(datetime)
}
\arguments{
\item{datetime}{The datetime object in POSIX format.}
}
\value{
A boolean value.
}
\description{
Check whether POSIX datetime object includes time information
}
