% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{hash}
\alias{hash}
\title{Hash function for nif, sdtm or domain objects}
\usage{
hash(obj)
}
\arguments{
\item{obj}{A nif, sdtm or domain object.}
}
\value{
The XXH128 hash of the object as character.
}
\description{
Hash function for nif, sdtm or domain objects
}
