% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R, R/sdtm_domain.R, R/utilities.R
\name{hash.sdtm}
\alias{hash.sdtm}
\alias{hash.domain}
\alias{hash}
\alias{hash.nif}
\title{XXH128 hash}
\usage{
\method{hash}{sdtm}(x)

\method{hash}{domain}(x)

hash(x)

\method{hash}{nif}(x)
}
\arguments{
\item{x}{An object.}
}
\value{
XXH128 hash as string.
}
\description{
XXH128 hash
}
