% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{head.nif}
\alias{head.nif}
\title{Implementation of the head function}
\usage{
\method{head}{nif}(obj, ...)
}
\arguments{
\item{obj}{A nif object}

\item{...}{Further parameters}
}
\value{
None
}
\description{
Implementation of the head function
}
