% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{head}
\alias{head}
\title{#' Implementation of the head function
#'
#' @param obj A nif object
#' @import dplyr
#' @return None
#' @import utils
#' @export
head <- function(obj, ...) {
UseMethod("head")
}
Implementation of the head function}
\usage{
head(obj, ...)
}
\arguments{
\item{obj}{A nif object}

\item{...}{Further parameters}
}
\value{
None
}
\description{
#' Implementation of the head function
#'
#' @param obj A nif object
#' @import dplyr
#' @return None
#' @import utils
#' @export
head <- function(obj, ...) {
UseMethod("head")
}
Implementation of the head function
}
