% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{head}
\alias{head}
\title{Implementation of the head function}
\usage{
head(obj, n = 6)
}
\arguments{
\item{obj}{A nif object}

\item{n}{Number of rows to return.}
}
\value{
None
}
\description{
Implementation of the head function
}
