% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{ht_by_wt}
\alias{ht_by_wt}
\title{Weight by height scatterplot}
\usage{
ht_by_wt(obj)
}
\arguments{
\item{obj}{The NIF object.}
}
\value{
A plot object.
}
\description{
Weight by height scatterplot
}
