% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{import_nif}
\alias{import_nif}
\title{Read NIF data}
\usage{
import_nif(filename, format = "NONMEM")
}
\arguments{
\item{filename}{File name as character.}

\item{format}{Format as character. Must be 'NONMEM'.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
