% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{import_observation}
\alias{import_observation}
\title{Add observation from non-SDTM-formatted data table}
\usage{
import_observation(
  nif,
  raw,
  analyte,
  parent = NULL,
  cmt = NULL,
  observation_filter = "TRUE",
  USUBJID_field = "USUBJID",
  dtc_field = NULL,
  ntime_field = NULL,
  dv_field = NULL,
  keep = NULL,
  debug = FALSE,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{raw}{The raw observation data frame.}

\item{analyte}{The analyte name as character.}

\item{parent}{The parent as character.}

\item{cmt}{The compartment for the analyte as numeric.}

\item{observation_filter}{Filter term, as character.}

\item{USUBJID_field}{The field specifying the USUBJID, as character.}

\item{dtc_field}{The field specifying the DTC, as character.}

\item{ntime_field}{The field specifying the NTIME, as character.}

\item{dv_field}{The field specifying the dependent variable, as character.}

\item{keep}{Columns to keep, as character.}

\item{debug}{Keep debug information.}

\item{silent}{Suppress messages, defaults to nif_option setting, if NULL.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
