% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputation_rules.R
\docType{data}
\name{imputation_1}
\alias{imputation_1}
\title{Alternative imputation rule set}
\format{
An object of class \code{list} of length 3.
}
\usage{
imputation_1
}
\description{
\itemize{
\item No administration time imputations
\item TAFD is set to 0 for predose observations
}
}
\keyword{datasets}
