% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputation_rules.R
\docType{data}
\name{imputation_rules_standard}
\alias{imputation_rules_standard}
\title{Standard imputation rule set}
\format{
A list with the following elements:
\describe{
\item{admin_pre_expansion}{A function to conduct imputations on the EX domain
before expansion of the administration episodes.}
\item{admin_post_expansion}{A function to conduct imputations on the
administration data table after expansion of the administration episodes}
\item{obs_raw}{A function to conduct imputations on the observation data.
At this stage, time fields other than NTIME are not derived yet.}
\item{obs_final}{A function to conduct imputations on the nif data table
including the new observations, before finalization. New observations are
flagged by .current_observation == TRUE.}
}
}
\usage{
imputation_rules_standard
}
\description{
A list containing the standard imputation functions to be used in
add_administration().
}
\keyword{datasets}
