% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\docType{data}
\name{imputation_standard}
\alias{imputation_standard}
\title{Standard imputation rules}
\format{
A list with the following elements:
\describe{
\item{admin_pre_expansion}{A function to conduct imputations on the EX domain
before expansion of the administration episodes.}
\item{admin_post_expansion}{A function to conduct imputations on the EX domain
after expansion of the administration episodes}
}
}
\usage{
imputation_standard
}
\description{
A list containing the standard imputation functions to be used in
add_administration().
}
\keyword{datasets}
