% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{imputation_standard}
\alias{imputation_standard}
\title{standard imputation function}
\usage{
imputation_standard(step = c("ex_pre_expansion", "ex_post_expansion"))
}
\arguments{
\item{step}{Imputation step}
}
\value{
A function
}
\description{
standard imputation function
}
