% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_admin_times_from_pcrftdtc}
\alias{impute_admin_times_from_pcrftdtc}
\title{Derive administration time from the PCRFTDTC}
\usage{
impute_admin_times_from_pcrftdtc(obj, pc, analyte, pctestcd)
}
\arguments{
\item{obj}{A data frame.}

\item{pc}{The corresponding PC domain as data frame.}

\item{analyte}{The analyte as string.}

\item{pctestcd}{The PCTESTCD corresponding to the analyte.}
}
\value{
A data frame.
}
\description{
Derive administration time from the PCRFTDTC
}
\keyword{internal}
