% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{impute_admin_times_from_pcrftdtc}
\alias{impute_admin_times_from_pcrftdtc}
\title{Derive administration time from PCRFTDTC}
\usage{
impute_admin_times_from_pcrftdtc(obj, pc, analyte, pctestcd)
}
\arguments{
\item{obj}{A data frame.}

\item{pc}{The corresponding PC domain as data frame.}

\item{analyte}{The analyte as character.}

\item{pctestcd}{The PCTESTCD corresponding to the analyte as character.}
}
\value{
A data frame.
}
\description{
Derive administration time from PCRFTDTC
}
\keyword{internal}
