% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{impute_exendtc_to_cutoff}
\alias{impute_exendtc_to_cutoff}
\title{Impute last EXENDTC per subject and treatment to cutoff date when absent.}
\usage{
impute_exendtc_to_cutoff(ex, cut_off_date = NA, silent = NULL)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{cut_off_date}{The cut-off date.}

\item{silent}{Suppress messages. Defaults to nif_option setting if NULL.}
}
\value{
The updated EX domain as data frame.
}
\description{
In some instances, particularly when analyzing non-cleaned SDTM data from
ongoing clinical studies with multiple-dose administrations, the last
administration epoch in EX may have an empty EXENDTC field. Often, the
underlying reason is that the respective subjects are still on treatment.
This function replaces the missing EXENDTC with the global data cut-off
date, \code{cut_off_date}.
}
