% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_imputations.R
\name{impute_missing_baseline}
\alias{impute_missing_baseline}
\title{Impute missing basline values}
\usage{
impute_missing_baseline(nif, baseline_fields = NULL, summary_function = median)
}
\arguments{
\item{nif}{The infput nif object.}

\item{baseline_fields}{Baseline fields to impute, as character. Will be set
to default values (WEIGHT, HEIGHT, BMI, other fields starting with BL_), if
NULL.}

\item{summary_function}{A function to determine the population center value
of each baseline. Defaults tom median.}
}
\value{
The nif object with missing baseline values imputed to the
respective population center value.
}
\description{
Impute missing basline values
}
