% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_missing_EXENDTC}
\alias{impute_missing_EXENDTC}
\title{Impute missing EXENDTC to the day before the next EXSTDTC.}
\usage{
impute_missing_EXENDTC(ex, silent = F)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{silent}{A boolean.}
}
\value{
The updated EX domain as data frame.
}
\description{
This imputation does not apply to the last administration per subject and
EXTRT. For these cases, missing EXENDT can be imputed to the global cut off
date using \code{impute_missing_EXENDTC_to_cutoff}.
}
