% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_missing_EXENDTC_to_cutoff}
\alias{impute_missing_EXENDTC_to_cutoff}
\title{Impute last EXENDTC per subject and treatment to cutoff date when absent.}
\usage{
impute_missing_EXENDTC_to_cutoff(ex, cut.off.date = NA, silent = F)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{cut.off.date}{The cut-off date.}

\item{silent}{Boolean.}
}
\value{
The EX domain as data frame.
}
\description{
Impute last EXENDTC per subject and treatment to cutoff date when absent.
}
