% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{impute_missing_EXENDTC_to_RFENDTC}
\alias{impute_missing_EXENDTC_to_RFENDTC}
\title{Impute very last EXENDTC for a subject and EXTRT to RFENDTC if absent}
\usage{
impute_missing_EXENDTC_to_RFENDTC(ex, dm, silent = F)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{dm}{The DM domain as data frame.}

\item{silent}{Boolean.}
}
\value{
The updated EX domain as data frame.
}
\description{
Impute very last EXENDTC for a subject and EXTRT to RFENDTC if absent
}
