% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{index_dosing_interval}
\alias{index_dosing_interval}
\title{Index dosing intervals}
\usage{
index_dosing_interval(obj, parent = NULL)
}
\arguments{
\item{obj}{The NIF object.}

\item{parent}{The treatments to filter for. Defaults to all parents, if NULL.}
}
\value{
A NIF object with the DI column added.
}
\description{
This function adds a column 'DI' that indicates the dosing interval per
parent. All baseline observations before the first dosing interval get
assigned to the first dosing interval.
}
\examples{
head(index_dosing_interval(examplinib_fe_nif))
head(index_dosing_interval(examplinib_poc_nif))
head(index_dosing_interval(examplinib_poc_min_nif))
}
