% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{index_dosing_interval}
\alias{index_dosing_interval}
\title{Index dosing invervals}
\usage{
index_dosing_interval(obj)
}
\arguments{
\item{obj}{The NIF object.}
}
\value{
A new NIF object.
}
\description{
This function adds a column \code{DI} that indicates the dosing interval. All
baseline observations before the first dosing interval get assigned to the
first dosing interval, too.
In addition to \code{DI}, the function also calls \code{index_nif()}, thus creating
the field \code{REF} as a side effect.
}
\examples{
as.data.frame(index_dosing_interval(examplinib_fe_nif))

}
