% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{index_nif}
\alias{index_nif}
\title{This function orders a NIF object and adds a REF field}
\usage{
index_nif(nif)
}
\arguments{
\item{nif}{NIF object, e.g., as created by \code{\link[=make_nif]{make_nif()}} and manually
modified.}
}
\value{
The updated NIF dataset including an updated REF field.
}
\description{
The input data format expected by NONMEM requires all rows ordered by ID
and TIME, and indexed sequentially on a subject level with a REF field.
Re-indexing may be required if a NIF object is extended, e.g., by merging
in further data.
}
\examples{
index_nif(examplinib_fe_nif)

}
