% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{index_rich_sampling_intervals}
\alias{index_rich_sampling_intervals}
\title{Identify and index rich PK sampling intervals}
\usage{
index_rich_sampling_intervals(obj, min_n = 4)
}
\arguments{
\item{obj}{The NIF object.}

\item{min_n}{The minimum number of PK samples per analyte to qualify as rich
sampling.}
}
\value{
A new NIF object.
}
\description{
Currently experimental. Don't use in production!
}
\details{
Adds the fields 'DI' (dosing interval per analyte) and 'RICH_N' (index of the
rich sampling interval by analyte).

This function identifies rich sampling intervals by the number of
observations that follow an administration. A number of 'min_n' or more
observations before the next administration is interpreted as a rich sampling
interval. The index of the rich sampling intervals per subject and analyte is
reported in the 'RICH_N' field.
}
\examples{
head(index_rich_sampling_intervals(examplinib_poc_nif))

}
