% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_nif.R
\name{index_rich_sampling_intervals}
\alias{index_rich_sampling_intervals}
\title{Identify and index rich PK sampling intervals}
\usage{
index_rich_sampling_intervals(obj, min_n = 4)
}
\arguments{
\item{obj}{The NIF data set.}

\item{min_n}{The minimum number of PK samples per analyte to qualify as rich
sampling.}
}
\value{
A new NIF data set.
}
\description{
Currently experimental. Don't use in production!
}
