% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{index_rich_sampling_intervals}
\alias{index_rich_sampling_intervals}
\title{Identify and index rich PK sampling intervals}
\usage{
index_rich_sampling_intervals(obj, analyte = NA, min_n = 4)
}
\arguments{
\item{obj}{The NIF object.}

\item{analyte}{The analyte as character. If \code{NA} (default), the most likely
will be selected automatically.}

\item{min_n}{The minimum number of PK samples per analyte to qualify as rich
sampling.}
}
\value{
A new NIF object.
}
\description{
Currently experimental. Don't use in production!
}
