% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{last_dtc}
\alias{last_dtc}
\title{Last recorded date-time}
\usage{
last_dtc(obj)
}
\arguments{
\item{obj}{A nif, sdtm or domain domain object.}
}
\value{
A POSIXct scalar.
}
\description{
Last recorded date-time
}
