% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_analysis.R
\name{lbm_boer}
\alias{lbm_boer}
\title{Lean body mass (Boer formula)}
\usage{
lbm_boer(weight, height, sex)
}
\arguments{
\item{weight}{Body weight in kg, as numeric.}

\item{height}{Body height in cm, as numeric.}

\item{sex}{Sex encoded as number (male is 0) or character (male is "M").}
}
\value{
Lean body mass in kg, as numeric.
}
\description{
Source: Caruso D, De Santis D, Rivosecchi F, Zerunian M, Panvini N, Montesano
M, Biondi T, Bellini D, Rengo M, Laghi A. Lean Body Weight-Tailored Iodinated
Contrast Injection in Obese Patient: Boer versus James Formula. Biomed Res
Int. 2018 Aug 13;2018:8521893. doi: 10.1155/2018/8521893. PMID: 30186869;
PMCID: PMC6110034.
}
