% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lubrify_dates}
\alias{lubrify_dates}
\title{Convert all DTC fields from ISO 8601 into POSIXct}
\usage{
lubrify_dates(obj, col = NULL)
}
\arguments{
\item{obj}{A data frame.}

\item{col}{Columns to convert. Defaults to all columns ending in "DTC", if
NULL.}
}
\value{
A data frame.
}
\description{
Change all columns in the input data frame that end with 'DTC' to standard
POSIXct format.
}
\keyword{internal}
