% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lubrify_dates}
\alias{lubrify_dates}
\title{Convert all DTC fields from ISO 8601 into POSIXct}
\usage{
lubrify_dates(obj)
}
\arguments{
\item{obj}{A data frame.}
}
\value{
A data frame.
}
\description{
Change all columns in the input data frame that end with 'DTC' to standard
POSIXct format.
}
\seealso{
\code{\link[=isofy_dates]{isofy_dates()}}
}
