% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_admin}
\alias{make_admin}
\title{Make administration data set from EX domain}
\usage{
make_admin(
  ex,
  drug_mapping,
  cut.off.date,
  impute.missing.end.time = TRUE,
  silent = F
)
}
\arguments{
\item{ex}{EX domain}

\item{drug_mapping}{A data frame with the columns of EXTRT and PCTESTCD
that associate both.}

\item{cut.off.date}{The cut-off date to be used where no EXENDTC is recorded,
in "\%Y-\%m-\%dT\%H:\%M" format.}

\item{impute.missing.end.time}{A logic value to indicate whether in rows
in EX where EXENDTC does not include a time, the time should be copied from
EXSTDTC.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}
}
\value{
A tibble with individual administrations
}
\description{
This function expands the administration ranges specified by EXSTDTC and
EXENDTC in each record of EX to that the result has individual records for
all administrations, with the time point in the TIME column.
}
\section{Specific imputations}{

If the end date (EXENDTC) is missing, i.e., the administration is ongoing
At the timeof the data cutoff of the SDTM data set, EXENDTC is replaced by
cut.off.date (which needs to be supplied in the format used by SDTM)
}

