% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_administration}
\alias{make_administration}
\title{Compile administration data frame}
\usage{
make_administration(
  sdtm,
  extrt,
  analyte = NA,
  cmt = 1,
  cut_off_date = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  silent = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{extrt}{The EXTRT for the administration, as character.}

\item{analyte}{The name of the analyte as character.}

\item{cmt}{The compartment for the administration as numeric.}

\item{cut_off_date}{The data cut-off date as Posix date-time.}

\item{subject_filter}{The filtering to apply to the DM domain, as string,}

\item{silent}{Suppress messages, as logical.}

\item{cleanup}{Remove unnecessary fields, as logical.}
}
\value{
A data frame.
}
\description{
Compile administration data frame
}
\keyword{internal}
