% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_ae.R
\name{make_ae}
\alias{make_ae}
\title{Make AE observation}
\usage{
make_ae(
  sdtm,
  ae_term,
  ae_field = "AEDECOD",
  parent = NA,
  cmt = NA,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  coding_table = NULL,
  keep = character(0)
)
}
\arguments{
\item{sdtm}{A sdtm object.}

\item{ae_term}{The AE term as character.}

\item{ae_field}{The field in which the AE term is specified in. Can be:
\itemize{
\item AEDECOD:  Dictionary-derived term
\item AELLT:    Lowest-level term
\item AEHLT:    High-level term
\item AEHLGT:   High-level group term
\item AESOC:    System organ class
\item AEBODSYS: Body system or organ class
}

or any other field from the 'AE' domain}

\item{parent}{The parent compound as character.}

\item{cmt}{The compartment as numeric.}

\item{subject_filter}{A subject filter term.}

\item{observation_filter}{An observation filter term.}

\item{coding_table}{A DV coding table as data frame. The coding table translates
columns included in the AE domain into a DV field. A typical case is when
AETOXGR is not present but only AESEV. A coding table of
\verb{data.frame( AESEV = c("MILD", "MODERATE", "SEVERE"), DV = c(1, 2, 3)} can
then be used to convert the AESEV into a numeric DV variable.}

\item{keep}{Columns to keep, as character.}
}
\value{
A nif object.
}
\description{
This function adds AE events as observations with the time of the AE onset
as the observation time point and the severity grade as the dependent
variable.
}
\keyword{internal}
