% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_EXSTDY_EXENDY}
\alias{make_EXSTDY_EXENDY}
\title{Title}
\usage{
make_EXSTDY_EXENDY(ex, dm)
}
\arguments{
\item{ex}{The EX domain as data frame.}

\item{dm}{The DM domain as data frame.}

\item{silent}{Boolean.}
}
\value{
A data frame.
Make EXSTDY and EXENDY by USUBJID, EXTRT

Treatment days are calculated relative to RFSTDTC from DM.

The enhanced EX domain as data frame.
}
\description{
Title
}
\details{
Caution: Currently, the function works only with treatment days after Day 1.
If this be ever used for days before Day 1, a term must be
implemented to correct for the missing Day 0 in the time nomenclature.
}
