% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_nif}
\alias{make_nif}
\title{Make raw NIF data set from list of SDTM domains}
\usage{
make_nif(
  sdtm.data,
  spec = NULL,
  impute.missing.end.time = TRUE,
  impute.administration.time = TRUE,
  silent = FALSE,
  truncate.to.last.observation = TRUE,
  use_pctptnum = TRUE
)
}
\arguments{
\item{sdtm.data}{A list of SDTM domains as data tables, e.g., as loaded using
read_sas_sdtm(). As a minimum, dm, vs, pc and ex are needed.}

\item{spec}{The specimen to be represented in the NIF data set as string
(e.g., "BLOOD", "PLASMA", "URINE", "FECES"). When spec is an empty string
(""), which is the default setting, the most likely specimen, i.e., "BLOOD"
or "PLASMA" is selected, depending what is found in the PC data.}

\item{impute.missing.end.time}{A logic value to indicate whether in rows
in EX where EXENDTC does not include a time, the time should be copied from
EXSTDTC.}

\item{impute.administration.time}{A boolean value to indicate whether the
time of administration is to be imputed from the PCRFDTC field from the
PC domain. This field is 'permissible' and may not be present in certain
SDTM data.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}

\item{truncate.to.last.observation}{Boolean to indicate whether the data set
should be truncated to the last observation. In this case, administrations
after the last observation time point will deleted.}

\item{use_pctptnum}{Boolean to indicate whether to derive nominal time
(\code{NTIME}) from \code{PCTPTNUM}.}
}
\value{
A NIF object.
}
\description{
This function makes a basic NONMEM input file (NIF) data set,
following the conventions summarized in
\href{https://doi.org/10.1002/psp4.12404}{Bauer, CPT Pharmacometrics Syst. Pharmacol. (2019)}.

For SDTM data sets in which the administered drugs (EXTRT in domain EX) and
the corresponding analyte (PCTESTCD in domain PC) have different names,
treatment-to-analyte mappings must be added to the sdtm object using
\code{\link[=add_mapping]{add_mapping()}}.
}
\section{Imputations}{

Subjects with administration but no observations for the respective
analyte are deleted from the data set. For further imputations, see the
vignette "nif-imputations".
}

\section{Output fields}{

\itemize{
\item \code{ID} Subject identification number
\item \code{TIME} Recorded time of administration or observation events in hours
relative to the first individual event.
\item \code{AMT} Dose administered for dosing record, or zero for observations.
\item \code{DOSE} Dose in mg for administrations and post-dose observations.
\item \code{DV} The dependent variable, i.e., observed concentration, or zero for
administration records, in mg/l.
\item \code{LNDV} The natural Log of DV.
\item \code{RATE} Rate of infusion of drug or zero if drug is given as a bolus.
\item \code{MDV} One for missing DV, else zero.
\item \code{EVID} Event ID: 0 for observations, 1 for administrations.
\item \code{CMT} Pharmacokinetic compartment. Will be set to 1 for administrations
and 2 for observations. Should be changed afterwards, if needed.
\item \code{DTC} The date-time of the data record.
\item \code{FIRSTDTC} Date and time of first event per subject. This field is used
internally for the calculation of \code{TIME}. Although it is not needed
for NONMEM analysis, it is provided for subsequent NIF file building
steps, e.g., addition of further time-dependent endpoints.
\item \code{FIRSTADMINDTC} The date-time of the first administration of the
respective parent drug for the respective subject.
\item \code{FIRSTTRTDTC} The date-time of the first administration of any parent
drug for the respective subject.
\item \code{ANALYTE} The analyte or drug in the data record.
\item \code{TRTDY} The treatment day, i.e., the relative day after the first
treatment for the respective subject.
}
}

\examples{
make_nif(examplinib_fe)

}
\seealso{
\code{\link[=add_analyte_mapping]{add_analyte_mapping()}}

\code{\link[=add_time_mapping]{add_time_mapping()}}
}
