% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_nif}
\alias{make_nif}
\title{Make raw NIF data set from list of SDTM domains}
\usage{
make_nif(
  sdtm.data,
  spec = NULL,
  impute.missing.end.time = TRUE,
  silent = F,
  truncate.to.last.observation = FALSE,
  use_pctptnum = TRUE
)
}
\arguments{
\item{sdtm.data}{A list of SDTM domains as data tables, e.g., as loaded using
read_sas_sdtm(). As a minimum, dm, vs, pc and ex are needed.}

\item{spec}{The specimen to be represented in the NIF data set as string
(e.g., "BLOOD", "PLASMA", "URINE", "FECES"). When spec is an empty string
(""), which is the default setting, the most likely specimen, i.e., "BLOOD"
or "PLASMA" is selected, depending what is found in the PC data.}

\item{impute.missing.end.time}{A logic value to indicate whether in rows
in EX where EXENDTC does not include a time, the time should be copied from
EXSTDTC.}

\item{silent}{Boolean value to indicate whether warnings should be printed.}

\item{truncate.to.last.observation}{Boolean to indicate whether the data set
should be truncated to the last observation. In this case, administrations
after the last observation time point will deleted.}

\item{use_pctptnum}{Boolean to indicate whether to derive nominal time
(\code{NTIME}) from \code{PCTPTNUM}.}
}
\value{
A NIF data set as nif object.
}
\description{
This function makes a basic NONMEM input file (NIF) data set, following the
conventions summarized in
\href{https://doi.org/10.1002/psp4.12404}{Bauer, CPT Pharmacometrics Syst. Pharmacol. (2019)}.

For SDTM data sets in which the administered drugs (EXTRT in domain EX) and
the corresponding analyte (PCTESTCD in domain PC) have different names,
treatment-to-analyte mappings must be added to the sdtm object using
\code{\link[=add_mapping]{add_mapping()}}.
}
\section{Imputations}{

'make_nif()' uses a cut-off time that is equal to the last observation
time. All later administrations will not be in the data set.

Subjects-administration pairs that have no observations for the respective
analyte will be deleted from the data set.
}

\section{Output fields}{

\itemize{
\item \code{ID} Subject identification number
\item \code{TIME} Recorded time of administration or observation events in hours
relative to the first individual event.
\item \code{AMT} Dose administered for dosing record, or zero for observations.
\item \code{DOSE} Dose in mg for administrations and post-dose observations.
\item \code{DV} The dependent variable, i.e., observed concentration, or zero for
administration records.
\item \code{RATE} Rate of infusion of drug or zero if drug is given as a bolus.
\item \code{MDV} One for missing DV, else zero.
\item \code{EVID} Event ID: 0 for observations, 1 for administrations.
\item \code{CMT} Pharmacokinetic compartment. Will be set to 1 for administrations and 2 for
observations. Should be changed afterwards, if needed.
\item \code{FIRSTDTC} Date and time of first event per subject. This field is used
internally for the calculation of \code{TIME}. Although it is not needed
for NONMEM analysis, it is provided for subsequent NIF file building
steps, e.g., addition of further time-dependent endpoints.
}
}

\examples{
make_nif(examplinib)

}
\seealso{
\code{\link[=add_analyte_mapping]{add_analyte_mapping()}}

\code{\link[=add_time_mapping]{add_time_mapping()}}
}
