% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{make_ntime}
\alias{make_ntime}
\title{Make nominal time}
\usage{
make_ntime(obj, include_day = FALSE)
}
\arguments{
\item{obj}{The input as data table.}

\item{include_day}{include_day Include time component of treatment day, as
logical.}
}
\value{
A data frame.
}
\description{
Return NTIME lookup table or NULL if the xxELTM field is not included in
the input data frame.
}
\examples{
make_ntime(examplinib_sad_nif)
make_ntime(examplinib_fe_nif)
}
