% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{make_ntime}
\alias{make_ntime}
\title{Make nominal time}
\usage{
make_ntime(obj, domain = NULL, include_day = FALSE, silent = NULL)
}
\arguments{
\item{obj}{The input as data table.}

\item{domain}{The domain name as character.}

\item{include_day}{include_day Include time component of treatment day, as
logical.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}
}
\value{
A data frame.
}
\description{
Return NTIME lookup table or NULL if the xxELTM field is not included in
the input data frame.
}
