% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{make_ntime_from_tptnum}
\alias{make_ntime_from_tptnum}
\title{Make ntime lookup table from TPTNUM field}
\usage{
make_ntime_from_tptnum(obj, domain = NULL)
}
\arguments{
\item{obj}{A data frame containing time point text descriptions.}

\item{domain}{The domain code as character (default: "PC" for
pharmacokinetic). Used to determine the column name containing time point
descriptions.}
}
\value{
A data frame with a column representing the unique values of the
xxTPT variable and a NTIME column with the time in hours.
}
\description{
Make ntime lookup table from TPTNUM field
}
