% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_observations.R
\name{make_observation}
\alias{make_observation}
\title{Compile observation data frame}
\usage{
make_observation(
  sdtm,
  domain,
  testcd,
  analyte = NULL,
  parent = NULL,
  metabolite = FALSE,
  cmt = NA,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  cat = NULL,
  scat = NULL,
  testcd_field = NULL,
  dtc_field = NULL,
  dv_field = NULL,
  coding_table = NULL,
  factor = 1,
  ntime_lookup = NULL,
  ntime_method = "TPT",
  keep = NULL,
  include_day_in_ntime = FALSE,
  omit_not_done = TRUE,
  silent = NULL,
  na_to_zero = FALSE
)
}
\arguments{
\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{domain}{The domain as character.}

\item{testcd}{The observation variable, as character.}

\item{analyte}{The name for the analyte. Defaults to the 'testcd', if NULL.}

\item{parent}{The name of the parent analyte for the observation as
character. Defaults to the respective treatment administered before the
observation, if NULL.}

\item{metabolite}{Observation is a metabolite, as logical.}

\item{cmt}{The compartment for the observation as numeric.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{cat}{xxCAT filter to apply, as character.}

\item{scat}{xxSCAT filter to apply, as character.}

\item{testcd_field}{The xxTESTCD field. Defaults to the two-character domain
name followed by 'TESTCD', if NULL.}

\item{dtc_field}{The field to use as the date-time code for the observation.
Defaults to the two-character domain name followed by 'DTC', if NULL.}

\item{dv_field}{the field to use as the dependent variable. Defaults to the
two-character domain name followed by 'STRESN', if NULL.}

\item{coding_table}{Coding table to translate a categorical values into
numerical values, as data frame. The data frame must have at least one
column that matches a column in the domain, and a numerical 'DV' column
that provides the recoding result.}

\item{factor}{Multiplier for the DV field, as numeric.}

\item{ntime_lookup}{A data frame with two columns, a column that defines the
custom nominal time information in the target domain (e.g., 'PCELTM'), and
'NTIME'. This data frame is left_joined into the observation data frame
to provide the NTIME field.}

\item{ntime_method}{the field to derive the nominal time from. Allowed values
are 'TPT', 'TPTNUM', 'ELTM', 'VISITDY' and 'DY'. Defaults to xxTPT where xx
is the domain name.}

\item{keep}{Columns to keep, as character.}

\item{include_day_in_ntime}{as logical.}

\item{omit_not_done}{Delete rows where xxSTAT is "NOT DONE, as logical.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}

\item{na_to_zero}{Set all NA values of DV to 0, as logical.}
}
\value{
A data frame.
}
\description{
Create a data frame of observations from a SDTM domain specified by the \code{domain} argument
where the dependent variable comes from the \code{dv_field} argument and the
timing information from the \code{dtc_field} argument.

The 'TIME' in the output is \code{NA} throughout and needs to be calculated based
on administration time point information provided separately.

If the 'ntime_lookup' parameter is provided, 'NTIME' can be derived from a
field contained in the input data set, e.g., 'PCELTM' (see the code
examples). Otherwise, 'NTIME' will be \code{NA}.
}
\keyword{internal}
