% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_observation}
\alias{make_observation}
\title{Compile observation data frame}
\usage{
make_observation(
  sdtm,
  domain,
  testcd,
  analyte = NULL,
  parent = NA,
  cmt = NA,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  TESTCD_field = NULL,
  DTC_field = NULL,
  DV_field = NULL,
  coding_table = NULL,
  factor = 1,
  NTIME_lookup = NULL,
  keep = NULL
)
}
\arguments{
\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{domain}{The domain as character.}

\item{testcd}{The observation variable, as character.}

\item{analyte}{The name for the analyte. Defaults to the 'testcd', if NULL.}

\item{parent}{The name of the parent analyte for the observation as
character.}

\item{cmt}{The compartment for the observation as numeric.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{TESTCD_field}{The xxTESTCD field. Defaults to the two-character domain
name followed by 'TESTCD', if NULL.}

\item{DTC_field}{The field to use as the date-time code for the observation.
Defaults to the two-character domain name followed by 'DTC', if NULL.}

\item{DV_field}{the field to use as the dependent variable. Defaults to the
two-character domain name followed by 'STRESN', if NULL.}

\item{coding_table}{Coding table to translate a categorical values into
numerical values, as data frame. The data frame must have at least one
column that matches a column in the domain, and a numerical 'DV' column
that provides the recoding result.}

\item{factor}{Multiplier for the DV field, as numeric.}

\item{NTIME_lookup}{A data frame with two columns, a column that defines the
custom nominal time information in the target domain (e.g., 'PCELTM'), and
'NTIME'. This data frame is left_joined into the observation data frame
to provide the NTIME field.}

\item{keep}{Columns to keep, as character.}
}
\value{
A data frame.
}
\description{
Create a data frame of observations from a SDTM domain specified by 'domain'
where the dependent variable comes from the 'DV_field' parameter and the
timing information from the 'DTC_field' parameter.

The 'TIME' in the output is \code{NA} throughout and needs to be calculated based
on administration time point information provided separately.

If the 'NTIME_lookup' parameter is provided, 'NTIME' can be derived from a
field contained in the input data set, e.g., 'PCELTM' (see the code
examples). Otherwise, 'NTIME' will be \code{NA}.
}
\keyword{internal}
