% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_plot.R
\name{make_plot_data_set}
\alias{make_plot_data_set}
\title{Prepare data set for plotting from NIF object}
\usage{
make_plot_data_set(
  nif,
  analyte = NULL,
  dose = NULL,
  time = "TAFD",
  color = NULL,
  min_time = NULL,
  max_time = NULL,
  cfb = FALSE,
  dose_norm = FALSE,
  facet = "DOSE",
  na_value = NA
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte(s) to be plotted, as character.}

\item{dose}{The dose(s) to be filtered for.}

\item{time}{The time field to be plotted on the x-axis. Can be any of 'TIME',
'NTIME', 'TAFD' or 'TAD'.}

\item{color}{The column(s) to be used for coloring.}

\item{min_time}{The minimal time in units of the selected time field, as
numeric.}

\item{max_time}{The minimal time in units of the selected time field, as
numeric.}

\item{cfb}{Plot change from baseline, as logical.}

\item{dose_norm}{Dose-normalized values, as logical.}

\item{facet}{The column(s) to be used for faceting.}

\item{na_value}{Value to replace NA values with, as numeric.}
}
\value{
A list with:
\itemize{
\item 'data', the data table for plotting
\item 'group', the grouping variable
\item 'color', the coloring variable
\item 'facet', the faceting variable
}
}
\description{
Prepare data set for plotting from NIF object
}
\examples{
make_plot_data_set(examplinib_sad_nif)
make_plot_data_set(examplinib_sad_nif, color = "RACE")
make_plot_data_set(examplinib_sad_nif, facet = "RACE")
make_plot_data_set(examplinib_poc_nif, color = c("DI", "RACE"))
}
\keyword{internal}
