% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{make_sd_pc}
\alias{make_sd_pc}
\title{Create PC based on single-dose admninistration}
\usage{
make_sd_pc(dm, ex, vs, lb, sampling_scheme)
}
\arguments{
\item{dm}{The DM domain as data frame.}

\item{ex}{The EX domain as data frame.}

\item{vs}{The VS domain as data frame.}

\item{lb}{The LB domain as data frame.}

\item{sampling_scheme}{The PK sampling scheme as data frame.}
}
\value{
The PC domain as data frame.
}
\description{
Create PC based on single-dose admninistration
}
\keyword{internal}
