% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{make_subjects}
\alias{make_subjects}
\title{Compile subject information}
\usage{
make_subjects(
  dm,
  vs,
  silent = FALSE,
  subject_filter = {
     !ACTARMCD \%in\% c("SCRNFAIL", "NOTTRT")
 },
  cleanup = TRUE
)
}
\arguments{
\item{dm}{The DM domain as data table.}

\item{vs}{The VS domain as data table.}

\item{silent}{No messages as logical.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{cleanup}{Keep only required fields, as logical.}
}
\value{
A data table.
}
\description{
Compile subject information
}
\keyword{internal}
