% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_subjects.R
\name{make_subjects}
\alias{make_subjects}
\title{Compile subject information}
\usage{
make_subjects(
  dm,
  vs = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  keep = ""
)
}
\arguments{
\item{dm}{The DM domain as data table.}

\item{vs}{The VS domain as data table.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{keep}{Columns to keep, as character.}
}
\value{
A data table.
}
\description{
Compile subject information
}
\keyword{internal}
