% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_time.R
\name{make_time}
\alias{make_time}
\title{Calculate time fields based on DTC}
\usage{
make_time(obj)
}
\arguments{
\item{obj}{A nif object.}
}
\value{
A nif object with TIME, TAFD, and TAD fields added.
}
\description{
This function generates the following time fields:
\itemize{
\item 'TIME' is the time in hours relative to the subject's first record, be it
an administration or observation event.
\item 'TAFD' is the time in hours relative to the subject's first administration
of the respective parent. Note that if a subject has received multiple drugs
(parents), the 'TAFD' field refers to the respective first administration.
\item 'TAD' is the time in hours relative to the most recent administration of
the parent compound.
}
}
