% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{make_vs}
\alias{make_vs}
\title{Synthesize fictional baseline vital sign data (VS domain)}
\usage{
make_vs(dm)
}
\arguments{
\item{dm}{The DM domain to provide the subject data.}
}
\value{
The VS domain data as data frame.
}
\description{
Synthesize fictional baseline vital sign data (VS domain)
}
\keyword{internal}
