% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{mean_dose_plot}
\alias{mean_dose_plot}
\title{Mean dose plot}
\usage{
mean_dose_plot(obj, analyte = NULL)
}
\arguments{
\item{obj}{A NIF data set.}

\item{analyte}{The compound as character (i.e., the ANALYTE within the data
set).}
}
\value{
A ggplot object.
}
\description{
This function plots the mean dose per day over time
}
