% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_test_data.R
\name{miss_admins}
\alias{miss_admins}
\title{Create administration schedule with randomly missed doses}
\usage{
miss_admins(start_dtc, treatment_duration = 50, missed_prob = 0.15)
}
\arguments{
\item{start_dtc}{Starting DTC.}

\item{treatment_duration}{Number of treatment days.}

\item{missed_prob}{Probability of missing administration.}
}
\value{
EXSTDTC and EXENDTC as data frame.
}
\description{
Create administration schedule with randomly missed doses
}
