% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{n_administrations}
\alias{n_administrations}
\title{Number of administrations per subject}
\usage{
n_administrations(obj)
}
\arguments{
\item{obj}{A NIF object.}
}
\value{
A data frame.
}
\description{
This function returns the number of administrations per \code{ID} and \code{PARENT}.
}
\examples{
n_administrations(examplinib_poc_nif)
n_administrations(examplinib_poc_min_nif)
}
