% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{n_administrations}
\alias{n_administrations}
\title{Number of administrations per subject}
\usage{
n_administrations(obj)
}
\arguments{
\item{obj}{A NIF data set.}
}
\value{
A data frame.
}
\description{
Number of administrations per subject
}
\examples{
n_administrations(examplinib_nif)

}
