% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_nif.R
\name{nca}
\alias{nca}
\title{Non-compartmental analysis}
\usage{
nca(obj, group = "SEX", analyte = NULL)
}
\arguments{
\item{obj}{The source NIF object.}

\item{group}{The grouping variable. If none specified, defaults to SEX.}

\item{analyte}{The analyte. If none specified and multiple analytes are in
the dataset, defaults to the first analyte.}
}
\value{
A PKNCA results object
}
\description{
Non-compartmental analysis
}
