% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_nif.R
\name{nca}
\alias{nca}
\title{Non-compartmental analysis}
\usage{
nca(obj, analyte = NULL, keep = NULL)
}
\arguments{
\item{obj}{The source NIF object.}

\item{analyte}{The analyte. If none specified and multiple analytes are in
#'   the dataset, defaults to the first analyte.}

\item{keep}{A vector of fields to keep in the output.}
}
\value{
A data frame.
}
\description{
Non-compartmental analysis
}
