% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca}
\alias{nca}
\title{Non-compartmental analysis}
\usage{
nca(
  obj,
  analyte = NULL,
  parent = NULL,
  keep = "DOSE",
  group = NULL,
  nominal_time = F,
  silent = F,
  average_duplicates = T
)
}
\arguments{
\item{obj}{The source NIF object.}

\item{analyte}{The analyte. If none specified and multiple analytes are in
the dataset, defaults to the first analyte.}

\item{parent}{The parent compound to derive the administration information
from. By default, equal to the analyte.}

\item{keep}{A vector of fields to keep in the output.}

\item{group}{The grouping variable as string.}

\item{nominal_time}{A boolean to indicate whether nominal time rather than
actual time should be used for NCA.}

\item{silent}{No message output.}

\item{average_duplicates}{Boolean to indicate whether duplicate entries
should be averaged.}
}
\value{
A data frame.
}
\description{
Non-compartmental analysis
}
\examples{
nca(examplinib_nif)
nca(examplinib_nif, group = c("FASTED", "SEX"), analyte = "RS2023")

}
