% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca}
\alias{nca}
\title{Non-compartmental analysis of NIF data}
\usage{
nca(
  nif,
  analyte = NULL,
  parent = NULL,
  keep = NULL,
  group = NULL,
  time = "TIME",
  duplicates = "stop",
  duplicate_function = mean,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte as character.}

\item{parent}{The parent as character.}

\item{keep}{Columns to keep, as character.}

\item{group}{Columns to group by, as character.}

\item{time}{The time field as character.}

\item{duplicates}{Selection how to deal with duplicate observations with
respect to the USUBJID, ANALYTE and DTC fields:
\itemize{
\item 'stop': Stop execution and produce error message
\item 'identify': Return a list of duplicate entries
\item 'resolve': Resolve duplicates, applying the \code{duplicate_function} to the
duplicate entries.
}}

\item{duplicate_function}{Function to resolve duplicate values, defaults to
\code{mean}.}

\item{silent}{Suppress messages.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function is a wrapper around the NCA functions provided by the
\href{https://CRAN.R-project.org/package=PKNCA}{PKNCA} package.

NA values are set to zero!
Negative concentrations are set to zero!
}
\examples{
head(nca(examplinib_sad_nif, time = "TAD"))
head(nca(examplinib_fe_nif, time = "TIME", group = "FASTED"))
}
