% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca_power_model}
\alias{nca_power_model}
\title{Power fit for PK parameters}
\usage{
nca_power_model(
  nca,
  parameter = NULL,
  group = NULL,
  title = NULL,
  size = 2,
  alpha = 1
)
}
\arguments{
\item{nca}{PK parameters as data frame.}

\item{parameter}{The PK parameter as character.}

\item{group}{Grouping parameter as character.}

\item{title}{The title as character.}

\item{size}{The point size as numeric.}

\item{alpha}{The alpha value for the data points as numeric.}
}
\value{
A list of ggplot2 objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Plot the PK parameter defined by \code{parameter} over \code{DOSE} and overlay with
a linear model of the log-transformed parameter and the log-transformed dose.
The slope is printed in the caption line.

All zero values for the selected parameter are filtered out before analysis.
}
\examples{
nca_power_model(nca(examplinib_sad_nif, analyte = "RS2023"), "aucinf.obs")
nca_power_model(nca(examplinib_sad_nif, analyte = "RS2023"),
  c("cmax", "aucinf.obs"))
}
