% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_nca.R
\name{nca_summary}
\alias{nca_summary}
\title{PK parameter summary statistics by dose}
\usage{
nca_summary(
  nca,
  parameters = c("auclast", "cmax", "tmax", "half.life", "aucinf.obs"),
  group = "DOSE"
)
}
\arguments{
\item{nca}{The NCA results as provided by \code{nca}, as data frame.}

\item{parameters}{The NCA paramters to be tabulated as character,}

\item{group}{The grouping variable, defaults to DOSE.}
}
\value{
A data frame
}
\description{
PK parameter summary statistics by dose
}
\examples{
nca_summary(nca(examplinib_sad_nif, analyte="RS2023"))
}
