% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{new_nif}
\alias{new_nif}
\title{nif class constructor}
\usage{
new_nif(obj = NULL, ...)
}
\arguments{
\item{obj}{A data frame containing the actual NIF data or a sdtm object.}

\item{...}{Further arguments.}
}
\value{
A nif object from the input data set.
}
\description{
nif class constructor
}
