% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{new_sdtm}
\alias{new_sdtm}
\title{SDTM class constructor, creating a sdtm object from a set of SDTM domains}
\usage{
new_sdtm(
  sdtm_data,
  analyte_mapping = data.frame(),
  metabolite_mapping = data.frame(),
  time_mapping = data.frame()
)
}
\arguments{
\item{sdtm_data}{The SDTM domains as list of data frames.}

\item{analyte_mapping}{The analyte mapping as data frame.}

\item{metabolite_mapping}{The metabolite mapping as data frame.}

\item{time_mapping}{The time mapping as data frame.}
}
\value{
A sdtm object.
}
\description{
SDTM class constructor, creating a sdtm object from a set of SDTM domains
}
