% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{new_sdtm}
\alias{new_sdtm}
\title{SDTM class constructor, creating a sdtm object from a set of SDTM domains}
\usage{
new_sdtm(sdtm_data)
}
\arguments{
\item{sdtm_data}{The SDTM domains as list of data frames.}
}
\value{
A sdtm object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
