% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nice_enumeration}
\alias{nice_enumeration}
\title{Nice enumeration of multiple strings}
\usage{
nice_enumeration(items, conjunction = "and")
}
\arguments{
\item{items}{Items to enumerate as character.}

\item{conjunction}{The conjunction between the last and penultmate items.}
}
\value{
Enumeration as character.
}
\description{
Nice enumeration of multiple strings
}
\examples{
nice_enumeration("A")
nice_enumeration(c("A", "B"))
nice_enumeration(c("A", "B", "C"))
nice_enumeration(c("A", "B", "C"), conjunction = "or")
}
\keyword{internal}
