% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif-package.R
\docType{package}
\name{nif-package}
\alias{nif-package}
\title{nif: Build and Explore NONMEM Input Format Data Sets from Clinical Study Data}
\description{
A System to Convert Study Data Tabulation Model (SDTM) Formatted Data into Longitudinal Data Tables Suitable for Nonlinear Mixed Effects Modeling (NONMEM). Aggregate Drug Administrations, PK/PD Observations and Covariates in a Stepwise Manner. Explore, Summarize and Visualize SDTM and NONMEM Data Sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstrotmann/nif}
  \item \url{https://rstrotmann.github.io/nif/reference/index.html}
}

}
\author{
\strong{Maintainer}: Rainer Strotmann \email{rainer.strotmann@mailbox.org} (\href{https://orcid.org/0000-0003-0543-7984}{ORCID}) [copyright holder]

}
\keyword{internal}
