% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif-package.R
\docType{package}
\name{nif-package}
\alias{nif}
\alias{nif-package}
\title{nif: Construct NONMEM input files from SDTM data}
\description{
Functions to import SDTM, and create and inspect NIF files.
}
\author{
\strong{Maintainer}: Rainer Strotmann \email{rainer.strotmann@mailbox.org}

}
\keyword{internal}
