% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif-package.R
\docType{package}
\name{nif-package}
\alias{nif-package}
\title{nif: Build and Explore NONMEM Input Format data tables from SDTM Data}
\description{
Import and explore Study Data Tabulation Model (SDTM) data from clinical studies. Create and explore NONMEM Input Format (NIF) data sets from [SDTM](https://www.cdisc.org/standards/foundational/sdtm) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstrotmann/nif}
  \item \url{https://rstrotmann.github.io/nif/reference/index.html}
}

}
\author{
\strong{Maintainer}: Rainer Strotmann \email{rainer.strotmann@mailbox.org} (\href{https://orcid.org/0000-0003-0543-7984}{ORCID}) [copyright holder]

}
\keyword{internal}
