% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif-package.R
\docType{package}
\name{nif-package}
\alias{nif-package}
\title{nif: Construct NONMEM input files from SDTM data}
\description{
A system to import and explore Study Data Tabulation Model (SDTM) formatted data from clinical studies, and to create and explore NONMEM Input File (NIF) data sets from them that can be used for pharmacometric analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstrotmann/nif}
  \item \url{https://rstrotmann.github.io/nif/reference/index.html}
}

}
\author{
\strong{Maintainer}: Rainer Strotmann \email{rainer.strotmann@mailbox.org}

}
\keyword{internal}
