% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{nif}
\alias{nif}
\title{nif class constructor}
\usage{
nif(nif.data)
}
\arguments{
\item{nif.data}{A data frame containing the actual NIF data as generated by,
e.g, \code{\link[=make_nif]{make_nif()}}.}
}
\value{
A nif object from the input data set.
}
\description{
nif class constructor
}
