% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{nif}
\alias{nif}
\title{nif class constructor}
\usage{
nif(obj)
}
\arguments{
\item{obj}{A data frame containing the actual NIF data or a sdtm object.}
}
\value{
A nif object from the input data set.
}
\description{
nif class constructor
}
