% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{nif}
\alias{nif}
\title{nif class constructor}
\usage{
nif(obj = NULL, ..., silent = NULL)
}
\arguments{
\item{obj}{A data frame containing the actual NIF data or a sdtm object.}

\item{...}{Further arguments.}

\item{silent}{suppress messages.}
}
\value{
A nif object.
}
\description{
If no obj argument is provided, an empty, minimal nif object will be
generated. If a data frame is provided for obj, it is converted into a nif
object. Minimally expected fields are ID, TIME, AMT, CMT, EVID and DV. If ID
is missing but USUBJID is available, ID will be derived.
}
\details{
If the input is a sdtm object, a pharmacokinetic nif object is automatically
generated. Analyte mapping formulae can be supplied as the ... argument. For
details, see \code{nif_auto()}.
}
\examples{
nif()
}
\seealso{
\code{\link[=nif_auto]{nif_auto()}}
}
