% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{nif}
\alias{nif}
\title{nif class constructor}
\usage{
nif(obj = NULL, ..., silent = NULL)
}
\arguments{
\item{obj}{A data frame containing the actual NIF data or a sdtm object.}

\item{...}{Further arguments.}

\item{silent}{suppress messages.}
}
\value{
A nif object.
}
\description{
nif class constructor
}
\examples{
nif()
}
