% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_auto.R
\name{nif_auto}
\alias{nif_auto}
\title{Auto-generate nif from sdtm object}
\usage{
nif_auto(
  sdtm,
  ...,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  duplicates = "resolve",
  duplicate_function = mean,
  keep = NULL,
  silent = NULL
)
}
\arguments{
\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{...}{Formulae to define the relationships between PCTESTCD and EXTRT.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{duplicates}{Selection how to deal with duplicate observations with
respect to the USUBJID, ANALYTE and DTC fields:
\itemize{
\item 'stop': Stop execution and produce error message
\item 'ignore': Include duplicates in the data set
\item 'identify': Return a list of duplicate entries
\item 'resolve': Resolve duplicates, applying the \code{duplicate_function} to the
duplicate entries.
}}

\item{duplicate_function}{Function to resolve duplicate values, defaults to
\code{mean}.}

\item{keep}{Columns to keep, as character.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}
}
\value{
A nif object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
nif_auto(examplinib_sad, RS2023 ~ EXAMPLINIB)
nif_auto(examplinib_sad, RS2023 + RS2023487A ~ EXAMPLINIB)
}
