% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_auto.R
\name{nif_auto}
\alias{nif_auto}
\title{Auto-generate nif from sdtm object}
\usage{
nif_auto(
  sdtm,
  ...,
  subject_filter = "!ACTARMCD \%in\% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  baseline_filter = "LBBLFL == 'Y'",
  duplicates = "resolve",
  duplicate_function = mean,
  keep = NULL,
  silent = NULL
)
}
\arguments{
\item{sdtm}{A sdtm object. Needs at least the 'DM' and 'VS' domains, and the
domain the observations come from.}

\item{...}{Formulae to define the relationships between PCTESTCD and EXTRT.}

\item{subject_filter}{The filtering to apply to the DM domain.}

\item{observation_filter}{The filtering to apply to the observation source
data.}

\item{baseline_filter}{A filter term to identify the baseline condition, as
character.}

\item{duplicates}{Selection how to deal with duplicate observations with
respect to the USUBJID, ANALYTE and DTC fields:
\itemize{
\item 'stop': Stop execution and produce error message
\item 'ignore': Include duplicates in the data set
\item 'identify': Return a list of duplicate entries
\item 'resolve': Resolve duplicates, applying the \code{duplicate_function} to the
duplicate entries.
}}

\item{duplicate_function}{Function to resolve duplicate values, defaults to
\code{mean}.}

\item{keep}{Columns to keep, as character.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}
}
\value{
A nif object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
If no mapping is provided, \code{\link[=nif_auto]{nif_auto()}} will try to find matching treatment
names in EX and PCTESTCD in PC to create a basic nif object for pharmacokinetic
observations. It will fail if no EXTRT and PCTESTCD with identical names are
found.

If one or more mappings are provided as additional parameters, \code{\link[=nif_auto]{nif_auto()}}
will create respective observations. Mappings must be provided in the form:

\code{TESTCD ~ EXTRT}

where TESTCD can be any xxTESTCD from any domain xx included in the sdtm
object. See \code{\link[=testcd]{testcd()}} for an overview
on all analytes in all domains. Formulae can also specify multiple
observations to be associated with one treatment, e.g.,

\code{RS2023 + RS2023487A ~ EXAMPLINIB}.

Multiple mappings can be given, separated by commas. The analyte name in the
nif object is automatically set to the respective TESTCD.

\strong{IMPORTANT}: If multiple analytes are specified for a treatment, it is
implicitly assumed that the first analyte corresponds to the parent analyte
for the treatment, e.g., in the case above, 'RS2023' is assumed to be the
analyte for the 'EXAMPLINIB' treatment parent compound.

Note that compartments (CMT) are automatically assigned. If more control over
compartments, analyte names and parent assignment is desired, stepwise
creation of a nif object using \code{\link[=add_administration]{add_administration()}} and \code{\link[=add_observation]{add_observation()}}
is recommended.
}
\examples{
nif_auto(examplinib_sad, RS2023 ~ EXAMPLINIB, silent = TRUE)
nif_auto(examplinib_sad, RS2023 + RS2023487A ~ EXAMPLINIB, silent = TRUE)
nif_auto(examplinib_sad, RS2023 + WEIGHT ~ EXAMPLINIB, silent = TRUE)
}
\seealso{
\code{\link[=testcd]{testcd()}}

\code{\link[=add_administration]{add_administration()}}

\code{\link[=add_observation]{add_observation()}}
}
