% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_debugger.R
\name{nif_debugger}
\alias{nif_debugger}
\title{NIF file debugger}
\usage{
nif_debugger(nif_data, sdtm_data, analyte = NULL, usubjid = NULL)
}
\arguments{
\item{nif_data}{A NIF data set.}

\item{sdtm_data}{The underlying SDTM data set.}

\item{analyte}{The analyte as character. If NULL (default), the most likely
analyte will be selected.}
}
\value{
Nothing
}
\description{
This shiny app helps identifying suspicious data in the NIF data set. This
may be due to missing data in the underlying SDTM data set or bad imputations
during the generation of the NIF data set.
}
