% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_mean_plot}
\alias{nif_mean_plot}
\title{Mean plot}
\usage{
nif_mean_plot(
  obj,
  points = FALSE,
  lines = TRUE,
  group = "ANALYTE",
  dose = NULL,
  analyte = NULL,
  cmt = NULL,
  min_x = NULL,
  max_x = NULL,
  log = FALSE,
  point_size = 2
)
}
\arguments{
\item{obj}{The NIF object.}

\item{points}{Boolean to indicate whether points should be drawn.}

\item{lines}{Boolean to indicate whether lines should be drawn.}

\item{group}{The grouping covariate, defaults to 'ANALYTE'.}

\item{dose}{The dose(s) to filter for as numeric.}

\item{analyte}{The analyte as character. If NULL (default), all analytes will
be plotted.}

\item{cmt}{The compartment to plot. If NULL (default), all compartments are
plotted, unless 'analyte' is defined.}

\item{min_x}{The minimal value for the x scale as numeric.}

\item{max_x}{The maximal value for the x scale as numeric.}

\item{log}{Logarithmic y scale as logical.}

\item{point_size}{The point size as numeric, defaults to 2.}
}
\value{
A ggplot2 plot object.
}
\description{
This function expects the \code{NTIME} field in the data set and will stop
otherwise.
}
\examples{
nif_mean_plot(examplinib_sad_nif)
nif_mean_plot(examplinib_fe_nif, group = "FASTED")
nif_mean_plot(examplinib_poc_nif, dose = 500, max_x = 24)
nif_mean_plot(examplinib_poc_nif, dose = 500, max_x = 12, group = "SEX")
}
\keyword{internal}
