% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_mean_plot}
\alias{nif_mean_plot}
\title{Mean plot over a selected covariate}
\usage{
nif_mean_plot(x, points = FALSE, lines = TRUE, group = "ANALYTE")
}
\arguments{
\item{x}{The NIF data set.}

\item{points}{Boolean to indicate whether points should be drawn.}

\item{lines}{Boolean to indicate whether lines should be drawn.}

\item{group}{The grouping covariate, defaults to 'ANALYTE'.}
}
\value{
A ggplot2 plot object.
}
\description{
Mean plot over a selected covariate
}
\examples{
nif_mean_plot(examplinib_sad_nif)
nif_mean_plot(examplinib_fe_nif, group = "FASTED")
}
