% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_options.R
\name{nif_option}
\alias{nif_option}
\title{Set or get global options}
\usage{
nif_option(...)
}
\arguments{
\item{...}{Options as named values, or nothing.}
}
\value{
The global options as list, if ... is empty, or nothing.
}
\description{
Set global behavior options for all \code{nif} package functions. Currently
supported:
\itemize{
\item \code{silent} as logical: Suppress messages.
\item \code{watermark} as character: Watermark text on all figures.
}
}
\examples{
nif_option(silent = TRUE)
}
