% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_options.R
\name{nif_option}
\alias{nif_option}
\title{Set global options}
\usage{
nif_option(...)
}
\arguments{
\item{...}{Options as named values.}
}
\value{
Nothing.
}
\description{
Set global options
}
\examples{
nif_option(silent = TRUE)
}
