% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_options.R
\name{nif_option}
\alias{nif_option}
\title{Set global options}
\usage{
nif_option(...)
}
\arguments{
\item{...}{Options as named values.}
}
\value{
Nothing.
}
\description{
Set global behavior options for all \code{nif} package functions. Currently
supported:
\itemize{
\item \code{silent} as logical: Suppress messages.
}
}
\examples{
nif_option(silent = TRUE)
}
