% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_options.R
\name{nif_option}
\alias{nif_option}
\title{Set or get global options}
\usage{
nif_option(...)
}
\arguments{
\item{...}{Options as named values, or nothing.}
}
\value{
The global options as list, if ... is empty, or nothing.
}
\description{
Set global behavior options for all \code{nif} package functions. Currently
supported:
\itemize{
\item \code{silent} as logical: Suppress messages.
\item \code{watermark} as character: Watermark text on all figures.
\item \code{pinboard} as character: Pinboard path for sharing of nif/sdtm objects.
\item \code{debug} as logical: Print debug information.
\item \code{show_hash} as logical: Include dataset hash in figures.
\item \code{abbreviation_threshold} as numeric: Allowed line count in message output
without abbreviation. Set to Inf to always show all lines.
\item \code{abbreviation_maxlines} as numeric: Line number to that lists in message
output will be abbreviated if the line count threshold is exceeded.
}
}
\examples{
nif_option(silent = TRUE)
}
