% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_plot_id}
\alias{nif_plot_id}
\title{Plot DV time course data from individual subject}
\usage{
nif_plot_id(
  obj,
  id,
  analyte = NULL,
  cmt = NULL,
  max_time = NA,
  lines = TRUE,
  point_size = 2,
  log = FALSE,
  tad = FALSE,
  imp = NULL,
  y_scale = "lin"
)
}
\arguments{
\item{obj}{The NIF object}

\item{id}{The subject ID to be plotted}

\item{analyte}{The analytes to be displayes. Defaults to NULL (all).}

\item{cmt}{The compartment to plot as numeric.}

\item{max_time}{The right limit of the time scale}

\item{lines}{Plot lines as logical.}

\item{point_size}{Point size as numeric.}

\item{log}{Logarithmic y scale.}

\item{tad}{Logical value to select whether time after dose (TAD) rather than
TIME should be plotted.}

\item{imp}{The IMP for which administrations are to be indicated by vertical
lines. Defaults to NULL.}

\item{y_scale}{Y scale type. Use 'scale="log"' for a logarithmic y scale.
Default is "lin". Note that this argument is deprecated. Use 'log' instead.}
}
\value{
A ggplot2 object.
}
\description{
This function plots DV over TIME for an individual subject, id. Id can be
either the ID or the USUBJID. Administration time points are indicated with
vertical lines.
}
\examples{
nif_plot_id(examplinib_poc_nif, 1)
nif_plot_id(examplinib_poc_min_nif, 1, log = TRUE)
nif_plot_id(examplinib_poc_nif, 1, log = TRUE)
nif_plot_id(examplinib_poc_nif, 1, analyte="RS2023")
nif_plot_id(examplinib_poc_nif, 1, analyte="RS2023", tad = TRUE)
nif_plot_id(examplinib_poc_nif, "20230000221010001", analyte="RS2023")
nif_plot_id(examplinib_poc_nif, "20230000221010001", analyte="RS2023")
nif_plot_id(examplinib_poc_nif, 8, analyte="RS2023", imp="RS2023")
nif_plot_id(examplinib_poc_nif, 8, analyte=c("RS2023", "RS2023487A"))
nif_plot_id(examplinib_poc_min_nif, 1, analyte="CMT3")
nif_plot_id(examplinib_poc_min_nif, 1, tad=TRUE)
}
\keyword{internal}
