% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_spaghetti_plot}
\alias{nif_spaghetti_plot}
\title{Spaghetti plot over a selected covariate}
\usage{
nif_spaghetti_plot(
  obj,
  points = FALSE,
  lines = TRUE,
  group = "ANALYTE",
  nominal_time = FALSE,
  tad = FALSE,
  dose = NULL,
  analyte = NULL,
  min_x = NULL,
  max_x = NULL,
  log = FALSE,
  point_size = 2
)
}
\arguments{
\item{obj}{The NIF data set.}

\item{points}{Boolean to indicate whether points should be drawn.}

\item{lines}{Boolean to indicate whether lines should be drawn.}

\item{group}{The grouping covariate, defaults to 'ANALYTE'.}

\item{nominal_time}{Boolean to indicate whether the x-axis should be NTIME
rather than TIME.}

\item{tad}{Boolean to select whether time after dose (TAD) rather than TIME
should be plotted.}

\item{dose}{The dose(s) to filter for as numeric.}

\item{analyte}{The analyte as character. If NULL (default), all analytes will
be plotted.}

\item{min_x}{The minimal value for the x scale as numeric.}

\item{max_x}{The maximal value for the x scale as numeric.}

\item{log}{Boolean to define whether y axis is plotted on the log scale.}

\item{point_size}{the point size as numeric.}
}
\value{
A ggplot2 plot object.
}
\description{
Spaghetti plot over a selected covariate
}
\examples{
nif_spaghetti_plot(examplinib_fe_nif)
nif_spaghetti_plot(examplinib_sad_nif)
nif_spaghetti_plot(examplinib_poc_nif, analyte="RS2023")
nif_spaghetti_plot(examplinib_poc_nif, analyte="RS2023", tad = TRUE,
  dose = 500, log = TRUE, points = TRUE, lines = FALSE)
}
