% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{nif_spaghetti_plot}
\alias{nif_spaghetti_plot}
\title{Spaghetti plot over a selected covariate}
\usage{
nif_spaghetti_plot(
  x,
  points = FALSE,
  lines = TRUE,
  group = "ANALYTE",
  nominal_time = FALSE,
  tad = FALSE,
  analyte = NULL
)
}
\arguments{
\item{x}{The NIF data set.}

\item{points}{Boolean to indicate whether points should be drawn.}

\item{lines}{Boolean to indicate whether lines should be drawn.}

\item{group}{The grouping covariate, defaults to 'ANALYTE'.}

\item{nominal_time}{Boolean to indicate whether the x-axis should be NTIME
rather than TIME.}

\item{tad}{Boolean to select whether time after dose (TAD) rather than TIME
should be plotted.}

\item{analyte}{The analyte as character. If NULL (default), all analytes will
be plotted.}
}
\value{
A ggplot2 plot object.
}
\description{
Spaghetti plot over a selected covariate
}
\examples{
nif_spaghetti_plot(examplinib_fe_nif)
nif_spaghetti_plot(examplinib_sad_nif)
}
