% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{pin_list_nif}
\alias{pin_list_nif}
\title{List nif objects in pinboard}
\usage{
pin_list_nif(board = NULL)
}
\arguments{
\item{board}{A pinboard path, as character.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
