% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{pin_list_object}
\alias{pin_list_object}
\title{List contents of pinboard}
\usage{
pin_list_object(board = NULL, object_type)
}
\arguments{
\item{board}{The board path as character.}

\item{object_type}{The object to filter for (sdtm or nif)}
}
\value{
The board folder, defaults to the respective nif_option setting.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
