% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pinboard.R
\name{pin_read_sdtm}
\alias{pin_read_sdtm}
\title{Read sdtm object from pinboard}
\usage{
pin_read_sdtm(name, board = NULL)
}
\arguments{
\item{name}{The pin name, as character.}

\item{board}{The path to the board object. Defaults to the 'pinboard' value
of nif_option, or the value of the 'NIF_PINBOARD' key in the .Renviron
setting if the 'nif_option' is not set.}
}
\value{
A sdtm object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated, please use \code{\link[=pb_read_sdtm]{pb_read_sdtm()}} instead.
}
