% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_load.R
\name{pin_write.nif}
\alias{pin_write.nif}
\title{Write to pin board}
\usage{
\method{pin_write}{nif}(obj, board = NULL, name = NULL, title = NULL, silent = NULL)
}
\arguments{
\item{obj}{nif object.}

\item{board}{Path to pin board, as character, defaults to respective
nif_option setting if NULL.}

\item{name}{Name for the nif pin object, as character. Defaults to
'xx_nif' with xx the studies included.}

\item{title}{Title for the nif pin object, as character. Defaults to
the studies included.}

\item{silent}{Suppress messages. Defaults to nif_option setting if NULL.}
}
\value{
Nothing.
}
\description{
Write to pin board
}
