% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pinboard.R
\name{pin_write}
\alias{pin_write}
\title{Generic pin_write function}
\usage{
pin_write(
  obj,
  name = NULL,
  board = NULL,
  title = NULL,
  dco = NULL,
  silent = NULL
)
}
\arguments{
\item{obj}{The object to pin.}

\item{name}{Name for the pin object, as character. Defaults to
'xx_sdtm' or 'xx_nif' with xx the study name.}

\item{board}{The path to the board object. Defaults to the 'pinboard' value
of nif_option, or the value of the 'NIF_PINBOARD' key in the .Renviron
setting if the 'nif_option' is not set.}

\item{title}{Title for the pin object, as character. Defaults to
the study name(s).}

\item{dco}{Data cut off as character.}

\item{silent}{Suppress messages. Defaults to nif_option setting if NULL.}
}
\value{
Nothing.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
