% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_pivot.R
\name{pivot_analytes}
\alias{pivot_analytes}
\title{Convert nif object to wide data frame by NTIME}
\usage{
pivot_analytes(
  obj,
  analyte = NULL,
  keep = NULL,
  duplicates = "stop",
  duplicate_function = mean,
  silent = NULL
)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{Analytes to include, defaults to all, if NULL.}

\item{keep}{Fields to keep in pivoted output. Defaults to standard subject-
level covariates, if NULL.}

\item{duplicates}{Selection how to deal with duplicate observations with
respect to the ID, ANALYTE, NTIME and TRTDY fields:
\itemize{
\item 'stop': Stop execution and produce error message
\item 'ignore': Include duplicates in the data set
\item 'identify': Return a list of duplicate entries
\item 'resolve': Resolve duplicates, applying the \code{duplicate_function} to the
duplicate entries.
}}

\item{duplicate_function}{Function to resolve duplicate values, defaults to
\code{mean}.}

\item{silent}{Suppress messages.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
