% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{pk_sim}
\alias{pk_sim}
\title{Simulate PK based on the examplinib PopPK model}
\usage{
pk_sim(event_table)
}
\arguments{
\item{event_table}{The event table as required by RxODE.}
}
\value{
PK simulation as data frame
}
\description{
Simulate PK based on the examplinib PopPK model
}
\keyword{internal}
