% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{plot.domain}
\alias{plot.domain}
\title{Plot domain object}
\usage{
\method{plot}{domain}(x, testcd = NULL, points = TRUE, lines = FALSE, legend = TRUE, ...)
}
\arguments{
\item{x}{A domain object.}

\item{testcd}{Testcd field to filter for, defaults to all if NULL.}

\item{points}{Plot points, as logical.}

\item{lines}{Plot lines, as logical.}

\item{legend}{Plot legend, as logical.}

\item{...}{Further parameters.}
}
\value{
A ggplot2 object.
}
\description{
Plot domain object
}
\examples{
plot(domain(examplinib_sad, "lb"))
}
