% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot NIF object}
\usage{
\method{plot}{nif}(x, mean = FALSE, ...)
}
\arguments{
\item{x}{The NIF object to be plotted.}

\item{mean}{Plot the mean instead individual subjects.}

\item{...}{Further parameters}
}
\value{
A ggplot2 object.
}
\description{
This function plots a NIF object, grouped by the variable \code{group}. If no
grouping variable is provided, \code{DOSE} will be used.
}
\examples{
plot(examplinib_fe_nif)
plot(examplinib_fe_nif, analyte = "RS2023", points = TRUE)
plot(examplinib_fe_nif, analyte = "RS2023", group = "FASTED")
plot(examplinib_fe_nif, max_x = 24, points = TRUE)
plot(examplinib_fe_nif,
  analyte = "RS2023", group = "FASTED",
  nominal_time = TRUE, points = TRUE
)
plot(examplinib_fe_nif, analyte = "RS2023", group = "FASTED", mean = TRUE)
plot(examplinib_fe_nif, analyte = "RS2023", group = "FASTED", mean = TRUE,
  max_x = 24)
plot(examplinib_poc_nif, dose = 500, analyte = "RS2023", points = TRUE,
  lines = FALSE, tad = TRUE, log =TRUE)
}
