% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot nif data set}
\usage{
\method{plot}{nif}(
  obj,
  y_scale = "lin",
  max_x = NULL,
  analyte = NULL,
  mean = FALSE,
  doses = NULL,
  points = F,
  id = NULL,
  usubjid = NULL
)
}
\description{
Plot nif data set
}
