% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot nif data set}
\usage{
\method{plot}{nif}(obj, y_scale = "lin")
}
\arguments{
\item{obj}{The nif object}

\item{y_scale}{Type of y-axis scale. Can be 'log' or 'lin' (default).}
}
\value{
The plot object
}
\description{
Plot nif data set
}
