% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot nif object.}
\usage{
\method{plot}{nif}(
  x,
  analyte = NULL,
  dose = NULL,
  group = NULL,
  time_field = "TIME",
  min_time = 0,
  max_time = NULL,
  min_x = NULL,
  max_x = NULL,
  limit_time_to_observations = TRUE,
  points = FALSE,
  lines = TRUE,
  point_size = 2,
  alpha = 1,
  tad = FALSE,
  nominal_time = FALSE,
  cfb = FALSE,
  log = FALSE,
  admin = FALSE,
  summary_function = median,
  title = "",
  mean = F,
  ...
)
}
\arguments{
\item{x}{A nif object.}

\item{analyte}{The analyte.}

\item{dose}{The dose(s) to select as numeric.}

\item{group}{The grouping variable(s).}

\item{time_field}{The field to use as the time metric, as character.}

\item{min_time}{The minimum time as numeric.}

\item{max_time}{The maximum time as numeric.}

\item{min_x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use min_time instead.}

\item{max_x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use max_time instead. values
are present.}

\item{limit_time_to_observations}{Truncate time axis to last observation, as
logical.}

\item{points}{Plot points as logical.}

\item{lines}{Plot lines as logical.}

\item{point_size}{The point size as numeric.}

\item{alpha}{The alpha value for point plotting.}

\item{tad}{Time after dose as logical.}

\item{nominal_time}{Plot NTIME on x axis, as logical.}

\item{cfb}{Change from baseline as logical.}

\item{log}{Logarithmic y scale as logical.}

\item{admin}{Plot administrations as logical.}

\item{summary_function}{The summary function to apply if multiple baseline}

\item{title}{The plot title as character.}

\item{mean}{Mean plot as logical.}

\item{...}{Further graphical parameters.}
}
\description{
Plot nif object.
}
