% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot nif object.}
\usage{
\method{plot}{nif}(
  nif,
  analyte = NULL,
  dose = NULL,
  log = FALSE,
  time = "TIME",
  group = NULL,
  min_time = 0,
  max_time = NULL,
  points = FALSE,
  lines = TRUE,
  admin = FALSE,
  cfb = FALSE,
  summary_function = median,
  silent = TRUE,
  mean = FALSE,
  title = "",
  caption = "",
  integrate_predose = TRUE,
  legend = TRUE,
  show_n = FALSE,
  ...
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte as character.}

\item{dose}{The dose as numeric.}

\item{log}{Logarithmic y scale}

\item{time}{The time field to use, as character. Can be 'TIME', 'NTIME',
'TAFD' or 'TAD'.}

\item{group}{The grouping variable as character.}

\item{min_time}{The minimal time as numeric.}

\item{max_time}{The maximal time as numeric.}

\item{points}{Show data points as logical.}

\item{lines}{Show lines as logical.}

\item{admin}{Show vertical lines for administrations, as logical.}

\item{cfb}{Show change from baseline, as logical.}

\item{summary_function}{The summarizing function to apply to multiple
baseline values.}

\item{silent}{Suppress messages as logical.}

\item{mean}{Show a mean plot, as logical.}

\item{title}{The plot title as character.}

\item{caption}{The plot caption line as character.}

\item{integrate_predose}{Complete 'DOSE' field for predose values.}

\item{legend}{Show legend, as logical.}

\item{show_n}{Show sample size in mean plot, as logical.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Plot nif object.
}
\examples{
plot(examplinib_fe_nif, points = TRUE)
plot(examplinib_fe_nif, nominal_time = TRUE, group = "FASTED")
plot(examplinib_sad_nif)
plot(examplinib_sad_nif, cfb = TRUE)
plot(examplinib_poc_nif, analyte="RS2023", admin = TRUE)
plot(examplinib_poc_nif)
plot(examplinib_poc_nif, analyte="RS2023", time = "NTIME",
  group = "SEX", points = TRUE, lines = FALSE)
plot(examplinib_poc_nif, analyte="RS2023", time = "TAD",
  dose = 500, log = FALSE, points = TRUE, lines = FALSE)
plot(examplinib_sad_min_nif)
plot(examplinib_poc_min_nif, dose = 500, cmt = 2)
plot(examplinib_poc_min_nif, dose = 500, cmt = 2, time = "TAD",
  points = TRUE, lines = FALSE)
}
