% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot NIF object}
\usage{
\method{plot}{nif}(
  x,
  y_scale = "lin",
  log = FALSE,
  min_x = 0,
  max_x = NA,
  analyte = NULL,
  mean = FALSE,
  doses = NULL,
  points = FALSE,
  id = NULL,
  usubjid = NULL,
  group = NULL,
  administrations = FALSE,
  nominal_time = FALSE,
  tad = FALSE,
  lines = TRUE,
  alpha = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{The NIF object to be plotted.}

\item{y_scale}{Type of y-axis scale. Can be 'log' or 'lin'. Default is "lin".}

\item{log}{Boolean to select whether the y-scale is logarithmic. Alternative
\code{y_scale="log"}.}

\item{min_x}{Minimum x (time) scale value.}

\item{max_x}{Maximal x (time) scale value.}

\item{analyte}{The analyte to be plotted If this argument is not supplied
(or set to NULL), all analytes are shown.}

\item{mean}{Boolean value to indicate whether the mean value by dose and
analyte is to be plotted. In that case, the nominal time (NTIME) is shown
on the x axis. 'mean=T' will cast an error if multiple analytes are in the
data set of have been defined with 'analyte'.}

\item{doses}{The doses to be plotted. Can be a scalar value or a numeric
vector or NULL to plot all doses.}

\item{points}{Boolean value to define whether points should be plotted.}

\item{id}{Numerical scalar or vector of IDs to be plotted.}

\item{usubjid}{Character scalar or vector of USUBJIDs to be plotted.}

\item{group}{Character scalar to define a grouping variable. If specified,
this will cast an error if multiple analytes are in the data set or have
been defined by 'analyte'.}

\item{administrations}{Boolean value to indicate whether vertical lines
marking the administration times should be plotted. Has no function if
mean=F.}

\item{nominal_time}{Boolean to indicate whether NTIME rather than TIME should
be plotted on the x-axis.}

\item{tad}{Boolean to select wheter time after dose (TAD) rather than TIME
should be plotted.}

\item{lines}{Boolean to define whether lines are to be drawn.}

\item{alpha}{Numerical value between 0 and 1, passed on as alpha to ggplot.}

\item{title}{The plot title as string.}

\item{...}{Further arguments.}
}
\value{
The plot object
}
\description{
This function plots a NIF object, grouped by the variable \code{group}. If no
grouping variable is provided, \code{DOSE} will be used.
}
\examples{
plot(examplinib_fe_nif)
plot(examplinib_fe_nif, analyte = "RS2023", points = TRUE)
plot(examplinib_fe_nif, analyte = "RS2023", group = "FASTED")
plot(examplinib_fe_nif, max_x = 24, point = TRUE)
plot(examplinib_fe_nif,
  analyte = "RS2023", group = "FASTED",
  nominal_time = TRUE
)
plot(examplinib_fe_nif, analyte = "RS2023", group = "FASTED", mean = TRUE)
plot(examplinib_fe_nif, analyte = "RS2023", group = "FASTED", mean = TRUE,
  max_x = 24)
plot(examplinib_poc_nif, dose = 500, analyte = "RS2023", points = TRUE,
  lines = FALSE, tad = TRUE, log =TRUE)
}
\seealso{
\code{\link[=nif_viewer]{nif_viewer()}}
}
