% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot NIF data set}
\usage{
\method{plot}{nif}(
  x,
  y_scale = "lin",
  max_x = NULL,
  analyte = NULL,
  mean = FALSE,
  doses = NULL,
  points = F,
  id = NULL,
  usubjid = NULL,
  group = NULL,
  administrations = F,
  nominal_time = F,
  ...
)
}
\arguments{
\item{x}{The NIF object to be plotted.}

\item{y_scale}{Type of y-axis scale. Can be 'log' or 'lin'. Default is "lin".}

\item{max_x}{Maximal x (time) scale value.}

\item{analyte}{The analyte to be plotted If this argument is not supplied
(or set to NULL), all analytes are shown.}

\item{mean}{Boolean value to indicate whether the mean value by dose and
analyte is to be plotted. In that case, the nominal time (NTIME) is shown
on the x axis. 'mean=T' will cast an error if multiple analytes are in the
data set of have been defined with 'analyte'.}

\item{doses}{The doses to be plotted. Can be a scalar value or a numeric
vector or NULL to plot all doses.}

\item{points}{Boolean value to define whether points should be plotted.}

\item{id}{Numerical scalar or vector of IDs to be plotted.}

\item{usubjid}{Character scalar or vector of USUBJIDs to be plotted.}

\item{group}{Character scalar to define a grouping variable. If specified,
this will cast an error if multiple analytes are in the data set or have
been defined by 'analyte'.}

\item{administrations}{Boolean value to indicate whether vertical lines
marking the administration times should be plotted. Has no function if
mean=F.}

\item{nominal_time}{Boolean to indicate whether NTIME rather than TIME should
be plotted on the x-axis.}

\item{...}{Further arguments.}
}
\value{
The plot object
}
\description{
This function plots a NIF data set, grouped by the variable \code{group}. If no
grouping variable is provided, \code{DOSE} will be used.
}
\seealso{
\code{\link[=nif_viewer]{nif_viewer()}}
}
