% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot nif object.}
\usage{
\method{plot}{nif}(
  x,
  analyte = NULL,
  dose = NULL,
  log = FALSE,
  time = "TAFD",
  group = NULL,
  min_time = NULL,
  max_time = NULL,
  points = FALSE,
  lines = TRUE,
  admin = NULL,
  cfb = FALSE,
  summary_function = median,
  mean = FALSE,
  title = "",
  caption = "",
  integrate_predose = TRUE,
  legend = TRUE,
  show_n = FALSE,
  shading = TRUE,
  watermark = "",
  silent = deprecated(),
  ...
)
}
\arguments{
\item{x}{A nif object.}

\item{analyte}{The analyte as character.}

\item{dose}{The dose as numeric.}

\item{log}{Logarithmic y scale}

\item{time}{The time field to use, as character. Can be 'TIME', 'NTIME',
'TAFD' or 'TAD'.}

\item{group}{The grouping variable as character.}

\item{min_time}{The minimal time as numeric.}

\item{max_time}{The maximal time as numeric.}

\item{points}{Show data points as logical.}

\item{lines}{Show lines as logical.}

\item{admin}{Show vertical lines for administrations, as logical.}

\item{cfb}{Show change from baseline, as logical.}

\item{summary_function}{The summarizing function to apply to multiple
baseline values.}

\item{mean}{Show a mean plot, as logical.}

\item{title}{The plot title as character.}

\item{caption}{The plot caption line as character.}

\item{integrate_predose}{Complete 'DOSE' field for predose values.}

\item{legend}{Show legend, as logical.}

\item{show_n}{Show sample size in mean plot, as logical. Does currently not
implement grouping! \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}

\item{shading}{Show ribbons when mean = TRUE, as logical.}

\item{watermark}{Show watermark as character or logical. If watermark = TRUE,
a standard text will be shown.}

\item{silent}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Dummy option for
compatibility, set the global option \code{\link[=nif_option]{nif_option()}} with \code{silent = TRUE} to
suppress messages.}

\item{...}{Further graphical parameters for \code{geom_point()}.}
}
\value{
A ggplot object.
}
\description{
Plot nif object.
}
\examples{
plot(examplinib_fe_nif, points = TRUE)
plot(examplinib_fe_nif, nominal_time = TRUE, group = "FASTED")
plot(examplinib_sad_nif, mean = TRUE, max_time = 24, show_n = TRUE)
plot(examplinib_poc_nif, analyte="RS2023", admin = TRUE)
plot(examplinib_poc_nif, analyte="RS2023", time = "NTIME",
  group = "SEX", points = TRUE, lines = FALSE)
plot(examplinib_poc_nif, analyte="RS2023", time = "TAD",
  dose = 500, log = FALSE, points = TRUE, lines = FALSE)
plot(examplinib_poc_min_nif, dose = 500, cmt = 2)
plot(examplinib_fe_nif, points = TRUE, watermark = TRUE)
plot(examplinib_fe_nif, mean = TRUE, group = "FASTED", max_time = 24)
plot(examplinib_fe_nif, points = TRUE, watermark = "Examplinib food effect")
}
