% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_plot.R
\name{plot.nif}
\alias{plot.nif}
\title{Plot NIF object.}
\usage{
\method{plot}{nif}(
  x,
  analyte = NULL,
  dose = NULL,
  time = "TAFD",
  color = NULL,
  facet = "DOSE",
  min_time = NULL,
  max_time = NULL,
  cfb = FALSE,
  dose_norm = FALSE,
  admin = NULL,
  points = FALSE,
  lines = TRUE,
  log = FALSE,
  mean = FALSE,
  title = NULL,
  legend = TRUE,
  size = 1.5,
  scales = "fixed",
  alpha = 1,
  caption = NULL,
  ribbon = TRUE,
  group = deprecated(),
  na_value = NA,
  ...
)
}
\arguments{
\item{x}{The NIF object to be plotted.}

\item{analyte}{The analyte(s) to be plotted, as character.}

\item{dose}{The dose(s) to be filtered for.}

\item{time}{The time field to be plotted on the x-axis. Can be any of 'TIME',
'NTIME', 'TAFD' or 'TAD'.}

\item{color}{The column(s) to be used for coloring.}

\item{facet}{The column(s) to be used for faceting.}

\item{min_time}{The minimal time in units of the selected time field, as
numeric.}

\item{max_time}{The minimal time in units of the selected time field, as
numeric.}

\item{cfb}{Plot change from baseline, as logical.}

\item{dose_norm}{Dose-normalized values, as logical.}

\item{admin}{The analyte to be plotted as administration markers, as
character.}

\item{points}{Plot points, as logical.}

\item{lines}{Plot lines, as logical.}

\item{log}{Logarithmic y axis, as logical.}

\item{mean}{Plot means, as logical.}

\item{title}{The plot title, as character.}

\item{legend}{Show the plot legend, as logical.}

\item{size}{The \code{size} parameter to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} as numeric.}

\item{scales}{The \code{scales} parameter to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, can be
"fixed" (default), "free", "free_x" or "free_y".}

\item{alpha}{The \code{alpha} parameter to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, as numeric.}

\item{caption}{The caption line as per \code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{ribbon}{Plot ribbon in mean plot, as logical.}

\item{group}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Grouping variable. Use the
\code{color} or \code{facet} parameters instead.}

\item{na_value}{Value to replace NA values with, as numeric.}

\item{...}{Further parameters.}
}
\value{
A ggplot2 object.
}
\description{
Plot NIF object.
}
\examples{
plot(examplinib_fe_nif)
plot(examplinib_fe_nif, facet = "FASTED", time = "TAD")
plot(examplinib_fe_nif, color = "FASTED", time = "TAD", facet = "SEX")
plot(examplinib_sad_nif,
  mean = FALSE, points = TRUE, dose_norm = FALSE,
  facet = "RACE", log = TRUE, max_time = 72
)
plot(examplinib_sad_nif,
  dose_norm = TRUE, facet = NULL, log = TRUE,
  max_time = 48, points = TRUE
)
plot(examplinib_fe_nif,
  points = TRUE, color = c("FASTED"), mean = TRUE,
  max_time = 12, admin = "RS2023"
)
plot(examplinib_poc_nif,
  points = TRUE, dose_norm = TRUE, facet = NULL,
  color = "SEX", max_time = 25, time = "TAD", lines = FALSE, size = 3,
  alpha = 0.5, title = "POC study: all analytes and doses"
)
plot(examplinib_fe_nif,
  points = TRUE, color = c("FASTED"), mean = TRUE,
  max_time = 12, admin = "RS2023"
)
}
