% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_exploration.R
\name{plot.sdtm}
\alias{plot.sdtm}
\title{Plot SDTM object}
\usage{
\method{plot}{sdtm}(
  x,
  domain = "dm",
  usubjid = NULL,
  lines = TRUE,
  points = FALSE,
  analyte = NULL,
  log = FALSE,
  ...
)
}
\arguments{
\item{x}{The sdtm object.}

\item{domain}{The domain to be plotted, defaults to 'dm'.}

\item{usubjid}{The USUBJID to filter for. All subjects if \code{NULL} (default).}

\item{lines}{Boolean whether to plot lines.}

\item{points}{Boolean whether to plot points.}

\item{analyte}{The analyte to be plotted as character.}

\item{log}{Boolean whether to use a logarithmic y axis.}

\item{...}{Further printing parameters.}
}
\value{
Nothing.
}
\description{
Plot SDTM object
}
\examples{
plot(examplinib_poc)
plot(examplinib_poc, "dm")
plot(examplinib_poc, domain="ex")
plot(examplinib_poc, domain="pc")
plot(examplinib_poc, domain="vs", lines = FALSE, points = TRUE)
}
