% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_analysis.R
\name{plot.summary_nif}
\alias{plot.summary_nif}
\title{Plot NIF summary}
\usage{
\method{plot}{summary_nif}(x, ...)
}
\arguments{
\item{x}{A NIF data set.}

\item{...}{Further arguments.}
}
\value{
Nothing.
}
\description{
Plot NIF summary
}
