% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{print.nif}
\alias{print.nif}
\title{print() implementation for nif objects}
\usage{
\method{print}{nif}(x, ...)
}
\arguments{
\item{x}{A nif object.}

\item{...}{Additional parameters}
}
\description{
TO DO:
It cannot be assumed that a nif file has all the fields used in this
function. Update this function using defensive programming.
}
