% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{print.sdtm}
\alias{print.sdtm}
\title{print() implementation for SDTM}
\usage{
\method{print}{sdtm}(x, ...)
}
\arguments{
\item{x}{A SDTM object.}

\item{...}{Further parameters}
}
\description{
print() implementation for SDTM
}
