% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_qtime.R
\name{qtime_plot}
\alias{qtime_plot}
\title{Plot analyte over time by discretized time after first dose (QTIME)}
\usage{
qtime_plot(obj, analyte, breaks, points = FALSE, caption = TRUE, alpha = 1)
}
\arguments{
\item{obj}{A nif object.}

\item{analyte}{The analyte as character.}

\item{breaks}{The breaks for TAFD, as numeric.}

\item{points}{Plot original data points as logical.}

\item{caption}{Show caption as logical.}

\item{alpha}{The alpha parameter for the data points.}
}
\value{
A ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
