% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{race_coding}
\alias{race_coding}
\title{Race coding table}
\format{
A data frame with 8 rows and 3 columns:
\describe{
\item{RACEN}{Numeric code}
\item{RACE}{RACE CDISC submission value as per NCI code C74457}
\item{LABEL}{Abbreviation for labeling purpose}
}
}
\usage{
race_coding
}
\description{
Standard race coding table with numeric codes and labels
}
\keyword{datasets}
