% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{race_coding}
\alias{race_coding}
\title{Race coding table}
\format{
A data frame with 8 rows and 3 columns:
\describe{
\item{RACEN}{Numeric race code}
\item{RACE}{Race abbreviation}
\item{LABEL}{Full race description}
}
}
\usage{
race_coding
}
\description{
Standard race coding table with numeric codes and labels
}
\keyword{datasets}
