% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_load.R
\name{read_adam}
\alias{read_adam}
\title{Load ADaM data sets from disk}
\usage{
read_adam(data_path, dataset = NULL, format = "sas", delim = ",", ...)
}
\arguments{
\item{data_path}{The path to read from.}

\item{dataset}{The data set names to load, defaults to all, if NULL.}

\item{format}{The data format, can be one of "sas", "xpt" or "csv."}

\item{delim}{Delimiter in csv files.}

\item{...}{Further arguments to \code{haven::read_sas()}, \code{haven::read_xpt()} or
\code{readr::read_delim()}.}
}
\value{
An adam object.
}
\description{
Load ADaM data sets from disk
}
