% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{read_sdtm}
\alias{read_sdtm}
\title{Read SDTM data}
\usage{
read_sdtm(data_path, domain = c("dm", "vs", "ex", "pc"), format = "sas")
}
\arguments{
\item{data_path}{The file system path to the source folder as character.}

\item{domain}{The domain name(s) as character, defaults to
\code{c("dm", "vs", "ex", "pc")}.}

\item{format}{The format of the source files as character, either 'sas'
(default) or 'xpt'.}
}
\value{
A \code{sdtm} object.
}
\description{
This function reads SDTM-formatted data as SAS or XPT files from a folder
location.
}
