% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_load.R
\name{read_sdtm_sas}
\alias{read_sdtm_sas}
\title{Load SDTM data from sas7bdat files}
\usage{
read_sdtm_sas(data_path, ...)
}
\arguments{
\item{data_path}{Path to the required SDTM files in .sas7bdat format.}

\item{...}{Further optional
parameters may specify the individual SDTM domains to be loaded (lowercase,
no file extensions). If no further parameters are given, the standard set of
dm, ex, pc and vs are loaded.}
}
\value{
A named list of the relevant SDTM domains as data.frames
}
\description{
This function loads relevant SDTM domain files required for the generation
of NIF files from a folder. It assumes that DM, EX, PC and VS are available
at this location and will throw an error if not.
}
