% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recode_race}
\alias{recode_race}
\title{Recode RACE columns in nif object}
\usage{
recode_race(obj, coding_table = NULL, silent = NULL)
}
\arguments{
\item{obj}{A nif object with RACE as character field.}

\item{coding_table}{A data frame with the columns RACE and RACEN. Uses
default coding, if NULL.}

\item{silent}{Suppress messages, defaults to nif_option setting, if NULL.}
}
\value{
A nif object with the original RACE replaced by the numerical race
code.
}
\description{
For some purposes, e.g., NONMEM-based modeling, numerical values are expected
in the RACE field. This function recodes RACE based on the following
associations:
}
\examples{
nif::race_coding
recode_race(examplinib_sad_nif)
}
