% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{recode_sex}
\alias{recode_sex}
\title{Recode SEX field in a data frame}
\usage{
recode_sex(obj)
}
\arguments{
\item{obj}{The data.frame containing a SEX field}
}
\value{
The output data frame
}
\description{
This function recodes the SEX field in a data frame. All numerical values are
kept while "M" is recoded to 0 and "F" to 1.
If your downstream analysis requires different coding, please manually
re-code.
}
\keyword{internal}
