% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{sdtm}
\alias{sdtm}
\title{sdtm object constructor}
\usage{
sdtm(sdtm_data, source = "")
}
\arguments{
\item{sdtm_data}{The SDTM domains as a named list of data frames.}

\item{source}{Source information as character.}
}
\value{
A sdtm object.
}
\description{
Create an \code{sdtm} object from a named list of domain data frames. Domain names
are stored in lowercase. The object is that named list (class \code{c("sdtm", "list")}); extract a domain with \code{obj$dm} or \code{\link[=domain]{domain()}}.
}
