% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{sdtm}
\alias{sdtm}
\title{SDTM class constructor, creating a sdtm object from a set of SDTM domains}
\usage{
sdtm(sdtm.data)
}
\arguments{
\item{sdtm.data, }{a list of SDTM domains as data.frames}
}
\value{
A sdtm object
}
\description{
SDTM class constructor, creating a sdtm object from a set of SDTM domains
}
