% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{sdtm}
\alias{sdtm}
\title{SDTM class constructor}
\usage{
sdtm(sdtm_data)
}
\arguments{
\item{sdtm_data}{The SDTM domains as list of data frames.}
}
\value{
A sdtm object.
}
\description{
SDTM class constructor
}
