% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_domain_model}
\alias{sdtm_domain_model}
\title{SDTM Domain Model variables}
\format{
A data frame with 1969 rows and 6 columns.

\describe{
\item{DOMAIN}{Domain name}
\item{VARNAM}{Variable name}
\item{VARLABEL}{Long-form variable name}
\item{DEFTYPE}{Variable type}
\item{LENGTH}{Length of variable}
\item{CORE}{Level of requiredness}
}
}
\usage{
sdtm_domain_model
}
\description{
An overview on the variables included in the SDTM specification.
}
\details{
The first few lines of the data set for reference:

\preformatted{
DOMAIN   VARNAM                            VARLABEL DEFTYPE LENGTH CORE
    TA  STUDYID                    Study Identifier    text     10  Req
    TA   DOMAIN                 Domain Abbreviation    text      2  Req
    TA    ARMCD                    Planned Arm Code    text     20  Req
    TA      ARM          Description of Planned Arm    text    100  Req
    TA  TAETORD Planned Order of Element within Arm integer      8  Req
    TA     ETCD                        Element Code    text      8  Req
    TA  ELEMENT              Description of Element    text    200 Perm
    TA TABRANCH                              Branch    text    200  Exp
    TA  TATRANS                     Transition Rule    text    200  Exp
    TA    EPOCH                               Epoch    text    100  Req
    ...
}
}
\keyword{datasets}
