% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\docType{data}
\name{standard_nif_fields}
\alias{standard_nif_fields}
\title{Standard nif fields}
\format{
An object of class \code{character} of length 34.
}
\usage{
standard_nif_fields
}
\description{
Standard nif fields
}
\keyword{datasets}
