% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nif.R
\name{standardize_date_format}
\alias{standardize_date_format}
\title{Convert date fileds to POSIX format}
\usage{
standardize_date_format(obj, fields = NULL)
}
\arguments{
\item{obj}{A data frame.}

\item{fields}{Date variable names as character.}
}
\value{
A data frame
}
\description{
This function converts date time code (DTC) variables from the format used
in SDTM (i.e., something like "2001-01-02T09:59" where date and time are
separated by "T") to standard POSIX format. The names of teh variables to be
converted need to be provided by \code{fields}.
}
\examples{
standardize_date_format(examplinib_fe$domains[["ex"]], c("RFSTDTC", "EXSTDTC", "EXENDTC"))
}
