% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{studies}
\alias{studies}
\title{Studies within a nif object}
\usage{
studies(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A character vector of all STUDYIDs in the data set.
}
\description{
Studies within a nif object
}
\examples{
studies(examplinib_poc_nif)

}
