% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subject_info}
\alias{subject_info}
\title{Subject information generic method}
\usage{
subject_info(obj, id)
}
\arguments{
\item{obj}{The object, either an SDTM or NIF object.}

\item{id}{The ID or USUBJID as numeric or character.}
}
\description{
Subject information generic method
}
