% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subject_info}
\alias{subject_info}
\title{Details on selected subjects}
\usage{
subject_info(obj, id)
}
\arguments{
\item{obj}{The object, either an SDTM or NIF object.}

\item{id}{The ID or USUBJID as numeric or character.}
}
\description{
Details on selected subjects
}
\examples{
subject_info(examplinib_fe, subjects(examplinib_fe)[1, "USUBJID"])
subject_info(examplinib_poc_nif, 1)
unclass(subject_info(examplinib_poc_nif, 1))
subject_info(examplinib_poc_nif, 1)$administrations
}
