% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subject_info}
\alias{subject_info}
\title{Subject demographic information}
\usage{
subject_info(obj, id = "")
}
\arguments{
\item{obj}{A SDTM object.}

\item{id}{The USUBJID}
}
\description{
Subject demographic information
}
