% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subject_info}
\alias{subject_info}
\title{subject demographic information}
\usage{
subject_info(obj, id = "")
}
\arguments{
\item{obj, }{a sdtm class object}

\item{id}{the USUBJID}
}
\description{
subject demographic information
}
