% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subject_info.sdtm}
\alias{subject_info.sdtm}
\title{Baseline details for specific subjects}
\usage{
\method{subject_info}{sdtm}(obj, id)
}
\arguments{
\item{obj}{The object, either an SDTM or NIF object.}

\item{id}{The ID or USUBJID as numeric or character.}
}
\description{
Baseline details for specific subjects
}
\examples{
subject_info(examplinib_fe, subjects(examplinib_fe)[1, "USUBJID"])
subject_info(examplinib_fe, subjects(examplinib_fe)[1:3, "USUBJID"])
}
\keyword{internal}
