% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subject_info.sdtm}
\alias{subject_info.sdtm}
\title{Subject demographic information}
\usage{
\method{subject_info}{sdtm}(obj, id)
}
\arguments{
\item{obj}{A SDTM object.}

\item{id}{The USUBJID.}
}
\description{
Subject demographic information
}
