% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{subjects.nif}
\alias{subjects.nif}
\title{Subjects within a nif object}
\usage{
\method{subjects}{nif}(obj)
}
\arguments{
\item{obj}{A nif object}
}
\value{
A character vector of all USUBJIDs in the data set.
}
\description{
Subjects within a nif object
}
