% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{subjects}
\alias{subjects}
\title{Subjects within a NIF object}
\usage{
subjects(obj)
}
\arguments{
\item{obj}{A NIF object}
}
\value{
A data frame of all ID - USUBJID pairs in the data set.
}
\description{
Subjects within a NIF object
}
