% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R, R/sdtm_class.R
\name{subjects}
\alias{subjects}
\title{Studies within a nif object}
\usage{
subjects(obj)

subjects(obj)
}
\arguments{
\item{obj}{The sdtm object.}
}
\value{
A character vector of all STUDYIDs in the data set.
}
\description{
Studies within a nif object

Get the USUBJID from an sdtm object
}
