% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{subjects.sdtm}
\alias{subjects.sdtm}
\title{Unique subjects within a sdtm object}
\usage{
\method{subjects}{sdtm}(obj)
}
\arguments{
\item{obj}{The sdtm object.}
}
\value{
A data frame.
}
\description{
Unique subjects within a sdtm object
}
\examples{
subjects(examplinib_poc)
}
