% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{suggest}
\alias{suggest}
\title{Suggest data programming steps to generate a nif object from an sdtm object}
\usage{
suggest(obj, show_all = FALSE)
}
\arguments{
\item{obj}{A sdtm object}

\item{show_all}{Show all lines in longer data frames, as logical.}
}
\value{
Nothing.
}
\description{
Suggest data programming steps to generate a nif object from an sdtm object
}
\examples{
suggest(examplinib_poc)
}
