% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{suggest}
\alias{suggest}
\title{Suggest manual data programming steps for a sdtm data set}
\usage{
suggest(obj)
}
\arguments{
\item{obj}{A sdtm object}
}
\description{
Suggest manual data programming steps for a sdtm data set
}
