% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_domain.R
\name{summary.domain}
\alias{summary.domain}
\title{Summary for domain object}
\usage{
\method{summary}{domain}(object, ...)
}
\arguments{
\item{object}{Domain object.}

\item{...}{Further parameters.}
}
\value{
A summary object
}
\description{
Summary for domain object
}
