% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{summary.nif}
\alias{summary.nif}
\title{NIF data set overview}
\usage{
\method{summary}{nif}(object, egfr_function = egfr_cg, ...)
}
\arguments{
\item{object}{The NIF data set.}

\item{egfr_function}{The function to be used for estimation of the renal
function classes, see \code{\link[=add_bl_crcl]{add_bl_crcl()}} for reference.}

\item{...}{Further arguments.}
}
\value{
A summary_nif object.
}
\description{
NIF data set overview
}
\examples{
summary(examplinib_poc_nif)

}
