% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summary.nif}
\alias{summary.nif}
\title{NIF data set overview}
\usage{
\method{summary}{nif}(object, ...)
}
\arguments{
\item{object}{The NIF data set.}

\item{...}{Further arguments.}
}
\value{
A summary_nif object.
}
\description{
NIF data set overview
}
