% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{summary.sdtm}
\alias{summary.sdtm}
\title{SDTM summary}
\usage{
\method{summary}{sdtm}(object, ...)
}
\arguments{
\item{object}{A SDTM object.}

\item{...}{Further parameters.}
}
\value{
A sdtm_summary object.
}
\description{
This function returns a named list of properties of the SDTM object:
}
\details{
\itemize{
\item \code{study} The study identifier as character.
\item \code{subjects} The USUBJIDs as character.
\item \code{domains} The included domains by their SDTM code and the number of
respective unique USUBJID, as data frame,
\item \code{treatments} The unique \code{EXTRT} as character.
\item \code{arms} The unique \code{ACTAMCD} and \code{ACTARM} as data frame.
\item \code{doses} The unique \code{EXTRT} and \code{EXDOSE} as data frame.
\item \code{specimens} The unique \code{PCSPEC} as character.
\item \code{analytes} The unique \code{PCTEST} and \code{PCTESTCD} as data frame.
\item \code{pc_timepoints} The unique \code{PCTPT} and \code{PCTPTNUM} as data frame.
\item \code{analyte_mapping} The analyte mapping as data frame.
\item \code{metabolite_mapping} The metabolite mapping as data frame.
\item \code{time_mapping} The time mapping as data frame.
}
}
\examples{
summary(examplinib_poc)
}
\keyword{internal}
