% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{summary.sdtm}
\alias{summary.sdtm}
\title{SDTM summary}
\usage{
\method{summary}{sdtm}(object, ...)
}
\arguments{
\item{object}{A SDTM object.}

\item{...}{Further parameters.}
}
\value{
A sdtm_summary object.
}
\description{
SDTM summary
}
