% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{synthesize_sdtm_food_effect_study}
\alias{synthesize_sdtm_food_effect_study}
\title{Synthesize SDTM data for a fictional food effect study}
\usage{
synthesize_sdtm_food_effect_study()
}
\value{
The SDTM data as sdtm object.
}
\description{
Synthesize SDTM data for a fictional food effect study
}
\keyword{internal}
