% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_synthesis.R
\name{synthesize_sdtm_poc_study}
\alias{synthesize_sdtm_poc_study}
\title{MD study}
\usage{
synthesize_sdtm_poc_study(
  studyid = "2023000022",
  dose = 500,
  nrich = 12,
  nsubs = 80,
  nsites = 8,
  duration = 100,
  red_prob = 0.3
)
}
\arguments{
\item{studyid}{The study identifyer.}

\item{dose}{The dose to be administered. Defauts to 500.}

\item{nrich}{The number of subjects with rich PK sampling.}

\item{nsubs}{The number of subjects.}

\item{nsites}{The number of clinical sites.}

\item{duration}{The study duration in days.}

\item{red_prob}{The dose reduction probability.}
}
\value{
A stdm object.
}
\description{
MD study
}
\keyword{internal}
