% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_check.R
\name{time_plot}
\alias{time_plot}
\title{Plot time metrics against each other for observations}
\usage{
time_plot(
  obj,
  xtime = "TIME",
  ytime = "TAD",
  analyte = NULL,
  max_time = NULL,
  color = "CHECK",
  ...
)
}
\arguments{
\item{obj}{A nif object.}

\item{xtime}{The x axis time metric as character.}

\item{ytime}{The y axis time metric as character.}

\item{analyte}{The analyte as character. Defaults to all, if NULL.}

\item{max_time}{The maximum x axis time as numeric.}

\item{color}{The field to color by, as character. Defaults to CHECK.}

\item{...}{Further graphical parameters.}
}
\value{
A ggplot object.
}
\description{
Plot time metrics against each other for observations
}
