% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_class.R
\name{trial_title}
\alias{trial_title}
\title{Trial title}
\usage{
trial_title(obj)
}
\arguments{
\item{obj}{A sdtm object.}
}
\value{
The title as ASCII character, or NULL.
}
\description{
Extract trial title from the TS domain, if available. Non-ASCII characters
are removed from the output to ensure compatibility.
}
\examples{
trial_title(examplinib_sad)
}
