% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{usubjid}
\alias{usubjid}
\title{Get the USUBJID of subject}
\usage{
usubjid(obj, id, silent = NULL)
}
\arguments{
\item{obj}{A NIF object.}

\item{id}{A subject ID in numeric form.}

\item{silent}{Suppress messages, defaults to nif_option setting.}
}
\value{
The USUBJID as character.
}
\description{
Get the USUBJID of subject
}
\examples{
usubjid(examplinib_fe_nif, 1, silent = FALSE)
}
