% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_sdtm}
\alias{validate_sdtm}
\title{Validate sdtm object}
\usage{
validate_sdtm(obj, expected_domains = NULL)
}
\arguments{
\item{obj}{A stdm object.}

\item{expected_domains}{Expected domains as character.}
}
\value{
Nothing or stop.
}
\description{
Validate sdtm object
}
