% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_class.R
\name{write_nif}
\alias{write_nif}
\title{Write as space-delimited, fixed-width file as required by NONMEM}
\usage{
write_nif(obj, filename, fields = NULL)
}
\arguments{
\item{obj}{The NIF data set.}

\item{filename}{The filename as string.}

\item{fields}{The fields to export.}
}
\description{
Write as space-delimited, fixed-width file as required by NONMEM
}
