% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_write.R
\name{write_nif}
\alias{write_nif}
\title{Write as space-delimited, fixed-width file as required by NONMEM or a
character-separated file}
\usage{
write_nif(obj, filename = NULL, fields = NULL, sep = NULL)
}
\arguments{
\item{obj}{The NIF object.}

\item{filename}{The filename as string. If not filename is specified, the
file is printed only.}

\item{fields}{The fields to export. If NULL (default), all fields will be
exported.}

\item{sep}{The separating character, e.g. ',' or ';'. If NULL (default), the
output has a space-separated, fixed-width format.}
}
\value{
Nothing.
}
\description{
All numeric fields are reduced to 4 significant places. For this, IEC 60559
is applied, i.e., rounding for a last digit of 5 is to the next even number.
All fields are converted to character, and NA-values are converted to '.'.
}
\examples{
head(write_nif(examplinib_fe_nif), 5)
}
