% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_exploration.R
\name{wt_by_ht}
\alias{wt_by_ht}
\title{Weight by height scatterplot}
\usage{
wt_by_ht(obj)
}
\arguments{
\item{obj}{The NIF object.}
}
\value{
A plot object.
}
\description{
Weight by height scatterplot
}
